% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Prediction accuracy}
\usage{
accuracy(table)
}
\arguments{
\item{table}{contingency table produced from \code{terra::crosstab}, or
\code{base::table}}
}
\value{
list with names \code{augm, userproducer, overall, kappa}
}
\description{
Calculates user, producer, overall, and kappa from an error matrix.
NB! Reference (truth) must be along the columns of the matrix
(e.g. second argument in \code{raster::crosstab}, or \code{base::table} )
}
\examples{
a<- terra::rast(system.file("ex/logo.tif", package="terra"))
# make everything above 50 to be 50 and everything below NA
res<- r_reclassify_threshold(a, 50, c(1, 50))
res_table<- terra::crosstab(c(res[[1]], res[[2]]))
res_table
accuracy(res_table)
}
