% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{acf_plus}
\alias{acf_plus}
\title{Calculate the autocorrelation function for a time series of returns, and plot
it.}
\usage{
acf_plus(x_ts, lagg = 10, plo_t = TRUE, xlab = "Lag", ylab = "",
  main = "", ...)
}
\arguments{
\item{\code{x_ts}}{A vector, matrix, or time series of returns.}

\item{\code{lagg}}{The maximum lag at which to calculate the ACF (default is
\code{10}).}

\item{\code{plo_t}}{A \emph{Boolean} argument: should a plot be made?
(default is \code{TRUE})}

\item{\code{xlab}}{A string with the x-axis label.}

\item{\code{ylab}}{A string with the y-axis label.}

\item{\code{main}}{A string with the plot title.}

\item{\code{...}}{Additional arguments to the function \code{stats::acf()}.}
}
\value{
Returns the ACF data invisibly and creates a plot.
}
\description{
Calculate the autocorrelation function for a time series of returns, and plot
it.
}
\details{
The function \code{acf_plus()} calls the function
  \code{stats::acf()}, removes the first element of the ACF data (which is
  spurious), creates a plot, and returns the ACF data invisibly.
}
\examples{
# Plot the ACF of random returns
rutils::acf_plus(rnorm(1e4), lag=10, main="ACF of Random Returns")
# Plot the ACF of VTI returns
rutils::acf_plus(na.omit(rutils::etf_env$re_turns$VTI), lag=10, main="ACF of VTI Returns")
}
