% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{add_month}
\alias{add_month}
\title{Adds a certain number of months to the given date in %y-%m-%d format.}
\usage{
add_month(monthstr, num = 1)
}
\arguments{
\item{monthstr}{(charachter): Given date %y-%m-%d format as character}

\item{num}{(numeric or integer): Number of months to be added to the given date. It can be a negative number.}
}
\value{
(character): Refers to the first day of the date with added months
}
\description{
Adds a certain number of months to the given date in %y-%m-%d format.
}
\examples{
add_month('2012-06-01', 3)
[1] "2012-09-01"
add_month('2010-03-12', -6)
[1] "2009-09-01"

}
