% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{add_year}
\alias{add_year}
\title{Add or subtract years to a given date character.}
\usage{
add_year(input_date, num_years = 1)
}
\arguments{
\item{num_years}{(numeric or integer): Number of years to be added to the given date. It can be a negative number.}
}
\value{
(character): date pointing to the same day and month of the input date with years added
}
\description{

}
\examples{
add_year('2012-06-11', 3)
[1] "2015-06-11"
add_month('2010-03-12', -6)
[1] "2004-03-12"

}
