% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address_to_coords.R
\name{address_to_coords}
\alias{address_to_coords}
\title{Get coordinates of any location in the World}
\usage{
address_to_coords(address)
}
\arguments{
\item{address}{a \code{character} of length 1. The location for which coordinates
will be found.}
}
\value{
A \code{data.frame} with two columns: \code{lon}, the longitude and \code{lat},
the latitude of the location.
}
\description{
Retrieves geographic coordinates (longitude and latitude) of any location in
the World using the OpenStreetMap Data Search Engine Nominatim
\url{http://nominatim.openstreetmap.org}.
}
\examples{
address_to_coords("Houston, TX, USA")
address_to_coords("5 rue de l'école de Médecine, Montpellier, France")
}
