% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{adjust_ohlc}
\alias{adjust_ohlc}
\title{Adjust the first four columns of \emph{OHLC} data using the "adjusted" price
column.}
\usage{
adjust_ohlc(oh_lc)
}
\arguments{
\item{\code{oh_lc}}{An \emph{OHLC} time series of prices in \emph{xts} format.}
}
\value{
An \emph{OHLC} time series with the same dimensions as the input
  series.
}
\description{
Adjust the first four columns of \emph{OHLC} data using the "adjusted" price
column.
}
\details{
Adjusts the first four \emph{OHLC} price columns by multiplying them
  by the ratio of the "adjusted" (sixth) price column, divided by the
  \emph{Close} (fourth) price column.
}
\examples{
# Adjust VTI prices
VTI <- rutils::adjust_ohlc(rutils::etf_env$VTI)
}
