% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{all_na}
\alias{all_na}
\title{Are all values in a column NA?}
\usage{
all_na(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Logical
}
\description{
Returns a logical value describing if all the values in the input are NA
}
\examples{
all_na(c(1,2,NA,3,4))

# Use with purrr to test columns individually

nadf <- data.frame(x = rep(NA, 3), y = c(1, 2, NA))
purrr::map_lgl(nadf, ~ all_na(.))

}
