% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{appender_rolling}
\alias{appender_rolling}
\title{Logger appender with rolling file capabilities}
\usage{
appender_rolling(
  filename,
  console = FALSE,
  inherit = TRUE,
  max_size = 10 * 1024 * 1024,
  max_files = 5,
  lock_file = NULL
)
}
\arguments{
\item{filename}{path of the logging file}

\item{console}{output to console}

\item{inherit}{inherit levels from upper level loggers}

\item{max_size}{size in bytes before rolling (defaults 10Mb)}

\item{max_files}{number of files to keep (defaults 5)}

\item{lock_file}{lock file to prevent access to lock during rolling
operations. If `NULL` no locking is performed (at your own risk)}
}
\description{
This extends the futile.logger::appender.file2 adding
rolling file capabilities
}
