% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdataframeExtent.R
\name{as.data.frame.Extent}
\alias{as.data.frame.Extent}
\title{as.data.frame method for extent}
\usage{
\method{as.data.frame}{Extent}(x, row.names, optional, ...)
}
\arguments{
\item{x}{extent}

\item{row.names}{optional}

\item{...}{}
}
\description{
Method for as.data.frame for extent objects
}
\examples{


res<- terra::ext(30, 50, 30, 50)
as.data.frame.Extent(res)

}
