% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{as_numeric_factor}
\alias{as_numeric_factor}
\title{Convert factor to numeric}
\usage{
as_numeric_factor(x)
}
\arguments{
\item{x}{factor vector}
}
\value{
numeric vector
}
\description{
Convert factor to numeric
}
