% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Build intuitive assertion calls}
\usage{
assert(message, ...)
}
\arguments{
\item{message}{(character) The error message to throw if the assertion fails}

\item{...}{(boolean) Statements that evaluate to TRUE or FALSE}
}
\value{
None
}
\description{
Build intuitive assertion calls
}

