% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{assign_files}
\alias{assign_files}
\title{Assign Files to the Global Environment}
\usage{
assign_files(.lst, .prefix = "", .suffix = "")
}
\arguments{
\item{.lst}{A list with unique names and file paths}

\item{.prefix}{Prefix for the names}

\item{.suffix}{Suffix for the names}
}
\value{
Object(s) in the Global Environment
}
\description{
Assign Files to the Global Environment
}
\examples{

library(rUtils)

dir <- system.file("extdata/files", package = "rUtils")

files <- lfc(dir)
assign_files(files)

rm(dir, files, fst, rds)
}
