% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{atomic_saveRDS}
\alias{atomic_saveRDS}
\title{Atomically save an R object to an RDS file}
\usage{
atomic_saveRDS(object, file, compression = "gzip")
}
\arguments{
\item{object}{The R object to save.}

\item{file}{The target file path.}

\item{compression}{Compression method (default "gzip").}
}
\value{
None.
}
\description{
Atomically save an R object to an RDS file
}
