% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{atomic_write_json}
\alias{atomic_write_json}
\title{Atomically write JSON content to a file}
\usage{
atomic_write_json(data, path, pretty = TRUE, auto_unbox = TRUE)
}
\arguments{
\item{data}{The R object to write as JSON.}

\item{path}{The file path where the JSON should be saved.}

\item{pretty}{Logical; if TRUE, the JSON output will be pretty-printed. Default is TRUE.}

\item{auto_unbox}{Logical; if TRUE, single-element vectors will be unboxed. Default is TRUE.}
}
\description{
Writes an R object as JSON to a temporary file and then renames it to the target path,
ensuring that the write operation is atomic.
}
