% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{backup_script}
\alias{backup_script}
\title{Backup Script}
\usage{
backup_script(.dir_here, .dir_script = NULL, .name_script)
}
\arguments{
\item{.dir_here}{Main Directory.}

\item{.dir_script}{Script Directory (optional).}

\item{.name_script}{Script Name.}
}
\value{
No return value; creates backup copies of specified files.
}
\description{
This function creates a backup of the specified Rmd and function files by copying them to the "_backup"
subdirectory within the "1_code" directory.
}
