% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandnames.R
\name{bandnames}
\alias{bandnames}
\title{Band names}
\usage{
bandnames(x = "s2_fullname", smallcaps = TRUE)
}
\arguments{
\item{x}{currently one of \code{c("s2", "rgb", "rgbnir")}}

\item{smallcaps}{}
}
\description{
Most common band names as character vector
}
\examples{
bandnames(smallcaps = TRUE)
}
