% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{bic_to_biclust}
\alias{bic_to_biclust}
\title{Converts two vectors of gene and cell clusters to biclustlib results.}
\usage{
bic_to_biclust(cell_clusters, gene_clusters, params = list())
}
\arguments{
\item{cell_clusters}{Named vector of cell clusters.}

\item{gene_clusters}{Named vector of gene clusters.}
}
\value{
An object of type "Biclust".
}
\description{
Converts two vectors of gene and cell clusters to biclustlib results.
}
