% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{Bioinformatics}
\alias{Bioinformatics}
\title{Bioinformatics}
\description{
Some scripts are present at \href{https://github.com/llrs/aligning/}{llrs/aligning}.

See the functions from \code{llrs_cnag} among others.
}
