% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_docs.R
\name{browse_docs}
\alias{browse_docs}
\alias{browse_wiki}
\title{Browse Package Website}
\usage{
browse_docs(package)

browse_wiki()
}
\arguments{
\item{package}{the package name. Opens \code{gdfR} if nothing provided}
}
\description{
Opens package website.
}
\details{
\code{browse_wiki()} is a shorthand for \code{browse_docs("wiki")}

check for valid package names using \code{fs::path_file(fs::dir_ls(getOption("path")$docs))}
}
\examples{
# browse_documentation("gdfR")
}
