% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{bu_name}
\alias{bu_name}
\title{Get Backup Name}
\usage{
bu_name(.name = "", .prec = 6, .sep = "_")
}
\arguments{
\item{.name}{A String}

\item{.prec}{1 = "\%Y"
2 = "\%Y-\%m"
3 = "\%Y-\%m-\%d"
4 = "\%Y-\%m-\%d-\%H"
5 = "\%Y-\%m-\%d-\%H-\%M"
6 = "\%Y-\%m-\%d-\%H-\%M-\%S"}

\item{.sep}{S separatot between time and name}
}
\value{
A string
}
\description{
Get Backup Name
}
\examples{
bu_name("file", 6)
bu_name("file", 5)
bu_name("file", 4)
bu_name("file", 3)
bu_name("file", 2)
bu_name("file", 1)
}
