% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{bucket}
\alias{bucket}
\title{Percentile-based bucketing function}
\usage{
bucket(v, ncut = 10, round.clever = F)
}
\arguments{
\item{v}{input numeric vector}

\item{ncut}{number of buckets to create}

\item{round.clever}{should values of v be rounded}
}
\value{
a numeric vector where each value represents a class/bucket
}
\description{
Associates each value of a numeric input vector to a percentile-based bucket.
}
\examples{
bucket(rnorm(100, 100, 25), 10, TRUE)
}
\keyword{bucket}
\keyword{ditribution}
