% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{bucket2}
\alias{bucket2}
\title{Threshold-based bucketing function}
\usage{
bucket2(
  v,
  splits = quantile(v, probs = seq(0, 1, by = 1/10)[2:(10 + 1)], na.rm = TRUE)
)
}
\arguments{
\item{v}{input numeric vector}

\item{splits}{sorted numeric thresholds delimiting the buckets (defaults to the deciles of `v`)}
}
\value{
an integer vector where each value represents a class/bucket
}
\description{
Associates each value of a numeric input vector to a threshold-based bucket.
}
\examples{
bucket2(rnorm(100, 100, 25), c(50, 75, 100, 125, 150))
}
\keyword{bucket}
\keyword{ditribution}
