% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_obj}
\alias{cache_obj}
\title{Cache an Object to Disk}
\usage{
cache_obj(
  .expr = {
 },
  .dir,
  .name,
  .rerun = FALSE
)
}
\arguments{
\item{.expr}{An expression that generates the object to be cached.
This is evaluated only if the object does not exist in the cache
or if \code{.rerun} is \code{TRUE}.}

\item{.dir}{A character string specifying the directory in which to store the
cached object. The directory will be created if it does not exist.}

\item{.name}{A character string specifying the name of the file to save the
cached object. This name is used as the base for the filename,
with appropriate file extensions added for the storage format.}

\item{.rerun}{A logical value indicating whether to re-run the expression
and re-cache the object, even if it already exists in the cache.
Defaults to \code{FALSE}.}
}
\value{
The function does not explicitly return a value. It saves or loads the
cached object as a side effect.
}
\description{
Caches an R object to disk using either the 'fst' format for data frames or
'qs' format for other objects. It creates a specified directory if it does
not exist and saves the object in that directory.
}
