% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch-memoised-functions.R
\name{cached_result}
\alias{cached_result}
\title{Evaluate the memoised version of a function}
\usage{
cached_result()
}
\value{
The result of calling a memoised function
}
\description{
This function is intended to be called within other
functions. When called within a function \code{foo} it searches
for another function called \code{foo_mem} and uses that instead.
\code{foo} should have an argument \code{use_cached_result}, which
controls whether or not this function gets called.
}
\examples{
runif2 <- function(n, use_cached_result = FALSE) {

if (use_cached_result) { return(cached_result()) }

runif(n)

}

runif2_mem <- memoise::memoise(runif2)

# These give different results
runif2(2)
runif2(2)

# These give the same results
runif2(2, TRUE)
runif2(2, TRUE)s
}
