% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeHome.R
\name{cacheFile}
\alias{cacheFile}
\title{provide wrapper to path to reference population data files
creates cache directory if not exist}
\usage{
cacheFile(fileName, extDisk = FALSE)
}
\arguments{
\item{fileName}{of expected file in path}

\item{extDisk}{for large files to be stored on external drives}
}
\description{
provide wrapper to path to reference population data files
creates cache directory if not exist
}
