% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CacheListBatch}
\alias{CacheListBatch}
\title{CacheListBatch class}
\description{
CacheListBatch class

CacheListBatch class
}
\details{
A class for caching results in list-based batches, storing them as JSON files.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cache_name}}{Name of the cache.}

\item{\code{cache_dir}}{Directory for cache files.}

\item{\code{cache_filename}}{File path for the cached JSON.}

\item{\code{list_of_keys_to_cache}}{List of keys to cache.}

\item{\code{get_batch_func}}{Function to fetch a batch of items for given keys.}

\item{\code{batch_size}}{Number of items per batch.}

\item{\code{sleep_between_calls}}{Delay in seconds between calls.}

\item{\code{cached_list_results}}{Cached list of results.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CacheListBatch-new}{\code{CacheListBatch$new()}}
\item \href{#method-CacheListBatch-init_cache}{\code{CacheListBatch$init_cache()}}
\item \href{#method-CacheListBatch-get_cached_list}{\code{CacheListBatch$get_cached_list()}}
\item \href{#method-CacheListBatch-is_key_in_cached_list}{\code{CacheListBatch$is_key_in_cached_list()}}
\item \href{#method-CacheListBatch-get_cached_item}{\code{CacheListBatch$get_cached_item()}}
\item \href{#method-CacheListBatch-clone}{\code{CacheListBatch$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheListBatch-new"></a>}}
\if{latex}{\out{\hypertarget{method-CacheListBatch-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new CacheListBatch object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheListBatch$new(
  cache_name,
  cache_dir,
  list_of_keys_to_cache,
  get_batch_func,
  batch_size,
  sleep_between_calls
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cache_name}}{Name of the cache.}

\item{\code{cache_dir}}{Directory to store cache files.}

\item{\code{list_of_keys_to_cache}}{List of keys to cache.}

\item{\code{get_batch_func}}{Function to fetch a batch of items for given keys.}

\item{\code{batch_size}}{Number of items per batch.}

\item{\code{sleep_between_calls}}{Delay in seconds between calls.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new CacheListBatch object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheListBatch-init_cache"></a>}}
\if{latex}{\out{\hypertarget{method-CacheListBatch-init_cache}{}}}
\subsection{Method \code{init_cache()}}{
Initialize or refresh the cache by fetching and storing data for specified keys.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheListBatch$init_cache()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None. Populates the cache and writes JSON file to disk.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheListBatch-get_cached_list"></a>}}
\if{latex}{\out{\hypertarget{method-CacheListBatch-get_cached_list}{}}}
\subsection{Method \code{get_cached_list()}}{
Retrieve the full cached results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheListBatch$get_cached_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of cached results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheListBatch-is_key_in_cached_list"></a>}}
\if{latex}{\out{\hypertarget{method-CacheListBatch-is_key_in_cached_list}{}}}
\subsection{Method \code{is_key_in_cached_list()}}{
Check if a key exists in the cached list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheListBatch$is_key_in_cached_list(item_key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item_key}}{The key to check.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if the key is in cache; FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheListBatch-get_cached_item"></a>}}
\if{latex}{\out{\hypertarget{method-CacheListBatch-get_cached_item}{}}}
\subsection{Method \code{get_cached_item()}}{
Retrieve a specific cached item by key.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheListBatch$get_cached_item(item_key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item_key}}{The key of the item to retrieve.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The cached item corresponding to the key.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CacheListBatch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CacheListBatch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CacheListBatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
