% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{calc_sharpe}
\alias{calc_sharpe}
\title{Calculate the Sharpe and Sortino ratios of a time series of returns.}
\usage{
calc_sharpe(retp, riskf = 0, nperiods = 252)
}
\arguments{
\item{\code{retp}}{A time series of returns, with multiple columns.}

\item{\code{riskf}}{The annual risk-free interest rate (the default is
\code{0}).}

\item{\code{nperiods}}{The number of time periods in a year (the default is
\code{252} days).}
}
\value{
A matrix of the Sharpe and Sortino ratios.
}
\description{
Calculate the Sharpe and Sortino ratios of a time series of returns.
}
\details{
The function \code{calc_sharpe()} calculates the Sharpe and Sortino
  ratios of a time series of returns.

  The function \code{calc_sharpe()} performs an sapply() loop over the
  columns of the \code{retp} argument. It calculates the Sharpe and Sortino
  ratios for each column. It subtracts from \code{retp} the annual risk-free
  interest rate \code{riskf} divided by \code{nperiods}. It multiplies the
  Sharpe and Sortino ratios by the square root of \code{nperiods}, in order
  to obtain the annual ratios.

  The Sharpe ratio \eqn{S_r} is defined as:
  \deqn{
    S_r = \sqrt{n} \frac{\bar{r}}{\sigma_r}
  }
  Where \eqn{r} are the daily excess returns (the returns minus the risk-free
  rate), \eqn{\bar{r}} are the average excess returns, and \eqn{\sigma_r} is
  their daily standard deviation.

  The Sortino ratio \eqn{{So}_r} is defined as:
  \deqn{
    {So}_r = \sqrt{n} \frac{\bar{r}}{\sigma_d}
  }
  Where \eqn{r} are the daily excess returns (the returns minus the risk-free
  rate), \eqn{\bar{r}} are the average excess returns, and \eqn{\sigma_d} is
  their daily \emph{downside deviation}.
  The \emph{downside deviation} \eqn{\sigma_d} is equal to the standard
  deviation of the downside returns \eqn{r_d}, the returns that are less than
  the risk-free rate.
}
\examples{
# Calculate the Sharpe and Sortino ratios of VTI and IEF returns
rutils::calc_sharpe(rutils::etfenv$returns[, c("VTI", "IEF")])
}
