% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{calc_skew}
\alias{calc_skew}
\title{Calculate the skewness or kurtosis of a time series of returns.}
\usage{
calc_skew(retp, expn = 3)
}
\arguments{
\item{\code{retp}}{A time series of returns, with multiple columns.}

\item{\code{expn}}{The power (exponent) to raise the returns (the default is
\code{3} for skewness).}
}
\value{
A matrix of the skewness or kurtosis values.
}
\description{
Calculate the skewness or kurtosis of a time series of returns.
}
\details{
The function \code{calc_skew()} calculates the skewness or kurtosis
  of a time series of returns.

  The function \code{calc_skew()} performs an sapply() loop over the columns
  of the \code{retp} argument. It raise the returns to the power \code{expn}.
  If \code{expn = 3} it calculates the skewness.
  If \code{expn = 4} it calculates the kurtosis.

  The skewness \eqn{\varsigma} is defined as:
  \deqn{
    \varsigma = \frac{1}{n-1} \sum_{i=1}^n {(\frac{r - \bar{r}}{\sigma})^3}
  }
  Where \eqn{r} are the daily returns, \eqn{\bar{r}} are the average returns,
  and \eqn{\sigma} is their standard deviation.

  The kurtosis \eqn{\kappa} is defined as:
  \deqn{
    \kappa = \frac{1}{n-1} \sum_{i=1}^n {(\frac{r - \bar{r}}{\sigma})^4}
  }
}
\examples{
# Calculate the kurtosis of VTI and IEF returns
rutils::calc_skew(rutils::etfenv$returns[, c("VTI", "IEF")], expn=4)
}
