% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Character Capitalization}
\usage{
capitalize(character.vec)
}
\arguments{
\item{character.vec}{a character vector to be capitalized.}
}
\description{
Capitalize all elements of a character vector.
}
\examples{
vec <- c("ab", "ac", "ad")
capitalize(vec)
## prints
## "Ab"    "Ac"    "Ad"
}

