% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/character.contains.R
\name{character.contains}
\alias{character.contains}
\title{Character Matching}
\usage{
character.contains(sub.character, character)
}
\arguments{
\item{sub.character}{the character to seek.}

\item{character}{the character to be tested if contains
\code{sub.character}}
}
\description{
Helper function wrapping over native \code{\link[base]{grepl}}
function to provide a more readable way of testing whether or
not a character contains other.
}
\examples{
character.contains("ab", "abc") ## returns TRUE
character.contains("abc", "ab") ## returns FALSE.
}

