% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{chart_dygraph}
\alias{chart_dygraph}
\title{Plot an interactive \emph{dygraphs} candlestick plot with background shading
for an \emph{OHLC} time series in \emph{xts} format.}
\usage{
chart_dygraph(oh_lc, in_dic = NULL, ...)
}
\arguments{
\item{\code{oh_lc}}{An \emph{OHLC} time series in \emph{xts} format.}

\item{\code{in_dic}}{A \emph{Boolean} time series in \emph{xts} format for
specifying the shading areas, with TRUE indicating "lightgreen" shading,
and FALSE indicating "antiquewhite" shading (default is \code{NULL}).}

\item{\code{...}}{Additional arguments to function \code{dygraphs::dygraph()}.}
}
\value{
A \code{dygraphs} plot object, and a \code{dygraphs} plot produced as
  a side effect.
}
\description{
Plot an interactive \emph{dygraphs} candlestick plot with background shading
for an \emph{OHLC} time series in \emph{xts} format.
}
\details{
The function \code{chart_dygraph()} creates an interactive dygraphs
  candlestick plot with background shading for an \emph{OHLC} time series.
  The function \code{chart_dygraph()} uses plotting functions from the
  package
  \href{https://cran.r-project.org/web/packages/dygraphs/index.html}{dygraphs}.
}
\examples{
# Plot an interactive dygraphs candlestick plot with background shading
oh_lc <- rutils::etf_env$VTI
v_wap <- TTR::VWAP(price=quantmod::Ad(oh_lc), volume=quantmod::Vo(oh_lc), n=20)
oh_lc <- cbind(oh_lc[, c(1:3, 6)], v_wap)["2009-02/2009-04"]
rutils::chart_dygraph(oh_lc, in_dic=(oh_lc[, 4] > v_wap))
}
