% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{check_datasets}
\alias{check_datasets}
\title{Check dataset}
\usage{
check_datasets(x, timeline, participant_col, timediff, time)
}
\arguments{
\item{x}{A data.frame with information about the cohort.}

\item{timeline}{A data.frame with three columns, the visit name, the time since the first visit and a category.}

\item{participant_col}{The name of the column holding the identification of the participant.}

\item{timediff}{The name of the column with the time difference between visits in \code{timeline}.}

\item{time}{The name of the column with the visit time in \code{x}.}
}
\description{
Not tested!
Checks if the visits of a cohort are consistently.
That samples have enough/all the visits.
}
\details{
Assumes that x and timelines have the same column names for sample id and visit timepoints.
In addition, there should be a third column in timeline about the
}
