% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc_meta.R
\docType{data}
\name{clc_meta}
\alias{clc_meta}
\title{Corine land cover nomenclature}
\format{
A data frame with 44 rows and 5 variables:
\describe{
  \item{value}{three digit CLC code}
  \item{color}{hexadecimal color code}
  \item{level1-3}{Class name for each abstraction level }
  ...
}
}
\source{
\url{https://developers.google.com/earth-engine/datasets/catalog/COPERNICUS_CORINE_V20_100m}
}
\usage{
clc_meta
}
\description{
A dataset containing the codes and names for the Corine Land Cover Classification
}
\keyword{datasets}
