% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_date.R
\name{closest_date}
\alias{closest_date}
\title{Find a closest date}
\usage{
closest_date(among, to)
}
\arguments{
\item{among}{vector of class Date to pick from}

\item{to}{the targeted date}
}
\description{
Find the closest date among a vector of dates. Any NA values are removed
}
\examples{
dates <- as.Date(c("2010-01-01", "2010-05-01"))
date  <- as.Date("2010-01-05")

closest_date(among = dates , to = date)
}
