% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\docType{class}
\name{Cluster}
\alias{Cluster}
\alias{Cluster-class}
\title{Cluster wrapper}
\description{
A Cluster proxy to internal cluster functions
}
\details{
This proxy has been made to be less error prone and a little easier
to emply clusters in computations

no value added, just a small API
}
\section{Fields}{

\describe{
\item{\code{ncores}}{stores the number of cores used for this cluster}

\item{\code{cluster}}{the internal cluster}

\item{\code{working}}{a logical status to fix some erratic behaviour of the internal cluster}

\item{\code{active}}{a logical status (TRUE if Cluster is active)}
}}

\section{Methods}{

\describe{
\item{\code{export(Class)}}{Returns the result of coercing the object to
Class.  No effect on the object itself.}

\item{\code{initialize(ncores = detectCores())}}{Initializes the Cluster by the number of cores argument (ncores)}

\item{\code{isWorking()}}{Returns `TRUE` is the cluster is in working state}

\item{\code{submit(X, fun, ...)}}{Submit a `fun` to the cluster to operate on the `X` data}
}}

