% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnag.R
\name{cnag_file}
\alias{cnag_file}
\alias{llrs_cnag_project_file}
\alias{llrs_cnag_stats_file}
\alias{llrs_cnag_id}
\title{File names of the CNAG projects.}
\usage{
llrs_cnag_project_file(project, n = 1L, path = "")

llrs_cnag_stats_file(project, n = 1L, path = "")

llrs_cnag_id(project, n = 1L)
}
\arguments{
\item{project}{Name of the project.}

\item{n}{Integer value of the project (increasing from 1).}

\item{path}{Path to file.}
}
\value{
A path for the stats or the file describing the samples of a project.
}
\description{
Create the name of the files of the CNAG projects.
}
\section{Functions}{
\itemize{
\item \code{llrs_cnag_project_file()}: Create a project file path.

\item \code{llrs_cnag_stats_file()}: Create a project stats file path.

\item \code{llrs_cnag_id()}: Create a project name from the id.

}}
\examples{
llrs_cnag_id("project", 1:10)
llrs_cnag_project_file("project")
llrs_cnag_stats_file("project")

}
