% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeHome.R
\name{codeFile}
\alias{codeFile}
\title{provide wrapper to path code files}
\usage{
codeFile(fileName)
}
\arguments{
\item{fileName}{of expected file in path}
}
\description{
provide wrapper to path code files
}
