% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohens_d.R
\name{cohens_d}
\alias{cohens_d}
\title{Compute Cohen's d statistic}
\usage{
cohens_d(x, y, t = NULL, abs = TRUE)
}
\arguments{
\item{x}{A \code{\link[base:numeric]{numeric}} vector.}

\item{y}{A \code{\link[base:numeric]{numeric}} vector.}

\item{t}{(optional) A \code{\link[base:numeric]{numeric}} value for the t statistic
(Default: \code{NULL}).}

\item{abs}{(optional) A \code{\link[base:logical]{logical}} flag indicating whether
to return the absolute value of the Cohen's d statistic (Default: \code{TRUE}).}
}
\value{
A \code{\link[base:numeric]{numeric}} value representing the Cohen's d
statistic.
}
\description{
\code{cohens_d()} computes Cohen's d statistic for two independent samples.

This function is based on Cohen (1988) and Frey (2022) calculations
for samples with equal and unequal sizes.
}
\examples{
cohens_d(c(1, 2, 3, 4, 5), c(6, 7, 8, 9, 10))
#> [1] 3.162278 # Expected
}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences}
(2nd ed.). Lawrence Erlbaum Associates.

Frey, B. B. (Ed.). (2022). _The SAGE encyclopedia of research design
(2. ed.). SAGE Publications. \doi{10.4135/9781071812082}
}
\seealso{
Other statistical functions.: 
\code{\link{mode}()}
}
\concept{statistical functions.}
