% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2hex.R
\name{col2hex}
\alias{col2hex}
\title{Convert a color name to hexadecimal format}
\usage{
col2hex(x)
}
\arguments{
\item{x}{A character vector of color names present in
\code{\link[grDevices:colors]{grDevices::colors()}}. Hex colors are returned
as is.}
}
\value{
A \code{\link[base:character]{character}} vector of hexadecimal color codes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{col2hex()} converts a color name present in
\code{\link[grDevices:colors]{grDevices::colors()}} to hexadecimal format.
}
\details{
Adapted from Mikael Jagan's contribution in a
\href{https://stackoverflow.com/a/70121688/8258804}{Stack Overflow discussion}.
}
\examples{
col2hex("red")
#> [1] "#FF0000"

col2hex(c("red", "green", "blue"))
#> [1] "#FF0000" "#00FF00" "#0000FF"

col2hex(c("red", "#000"))
#> [1] "#FF0000" "#000000"
}
\concept{color functions.}
