% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{colnames_grepl}
\alias{colnames_grepl}
\title{Get the column names from an object that match the given regex pattern}
\usage{
colnames_grepl(obj, pattern)
}
\description{
Get the column names from an object that match the given regex pattern
}
