% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{column.feed.forward}
\alias{column.feed.forward}
\title{Feeds missing values of each row with the values of its previous row. Order columns appropriately brfore calling this function.
If id_col is specified, then the table is grouped by that column before 
Input tbl is a table: data.frame, tibble or matrix}
\usage{
column.feed.forward(tbl, col = 1, id_col = NULL)
}
\description{
Feeds missing values of each row with the values of its previous row. Order columns appropriately brfore calling this function.
If id_col is specified, then the table is grouped by that column before 
Input tbl is a table: data.frame, tibble or matrix
}
