% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utils.r
\name{combine2}
\alias{combine2}
\title{Esegue il prodotto cartesiano degli array di stringhe passati come argomento}
\usage{
combine2(..., prefix, sep)
}
\arguments{
\item{...}{stringhe su cui eseguire il prodotto cartesiano}

\item{prefix}{eventuale prefisso da applicare}

\item{sep}{separatore da utilizzare}
}
\description{
Esegue il prodotto cartesiano degli array di stringhe passati come argomento
}
\examples{
a = c("A", "B")
          b = c("1", "2", "3")
          combine2(a,b)
          # ritorna "A1" "B1" "A2" "B2" "A3" "B3"
}

