% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_json_lists}
\alias{compare_json_lists}
\title{Compare two JSON files containing lists}
\usage{
compare_json_lists(file1, file2)
}
\arguments{
\item{file1}{Path to the first JSON file.}

\item{file2}{Path to the second JSON file.}
}
\value{
A list with elements 'common', 'unique_to_file1', and 'unique_to_file2'.
}
\description{
Compare two JSON files containing lists
}
\examples{
result <- compare_json_lists(
  "data/accessions_of_interest_from_code_api.json",
  "data/accessions_of_interest_from_web_search.json"
)
}
