% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compress_rds_files}
\alias{compress_rds_files}
\title{Compress all RDS files in a directory}
\usage{
compress_rds_files(dir_path)
}
\arguments{
\item{dir_path}{The directory containing RDS files.}
}
\value{
A list of results for each file.
}
\description{
Compress all RDS files in a directory
}
