% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{.containsString}
\alias{.containsString}
\title{controlla se una stringa e' nell'altra}
\usage{
.containsString(stringa, substring)
}
\arguments{
\item{stringa}{stringa da controllare}

\item{substring}{sottoscringa da cercare in `stringa`}
}
\value{
-1 se non trova `substring` in `stringa` altrimenti la posizione
}
\description{
controlla se una stringa e' nell'altra
}
