% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_myrds_to_json}
\alias{convert_myrds_to_json}
\title{Convert a cached RDS file to JSON format}
\usage{
convert_myrds_to_json(interpro_filename)
}
\arguments{
\item{interpro_filename}{The base filename (without extension) for the RDS file in the cache and the JSON file in the temp directory.}
}
\value{
None. Calls convert_rds_to_json with constructed file paths.
}
\description{
Convert a cached RDS file to JSON format
}
