% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_rds_to_json}
\alias{convert_rds_to_json}
\title{Convert an RDS file to a JSON file}
\usage{
convert_rds_to_json(rds_file, json_file)
}
\arguments{
\item{rds_file}{The path to the input RDS file.}

\item{json_file}{The path where the output JSON file will be written.}
}
\value{
None. Writes a JSON file to disk.
}
\description{
Convert an RDS file to a JSON file
}
