% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_address.R
\name{coords_to_address}
\alias{coords_to_address}
\title{Retrieve OSM address from coordinates}
\usage{
coords_to_address(coords)
}
\arguments{
\item{coords}{a \code{numeric} of length 2. The first element is the \code{latitude}
and the second the \code{longitude}.}
}
\value{
A \code{character} of length 1 with the OSM name of the location.
}
\description{
Retrieves any location name in the World from geographic coordinates
(longitude and latitude) using the OpenStreetMap Data Search Engine Nominatim
\url{http://nominatim.openstreetmap.org}.
}
\examples{
coords_to_address(c(43.61277, 3.873486))
coords_to_address(c(29.75894, -95.3677))
}
