% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_na.R
\name{count_na}
\alias{count_na}
\title{Count the amount of \code{NA} values in a vector}
\usage{
count_na(x)
}
\arguments{
\item{x}{An \code{\link[checkmate:checkAtomic]{atomic}} object of any type.}
}
\value{
An \code{\link[base:integer]{integer}} number with the amount of \code{NA} values
in \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{count_na()} counts the amount of \code{NA} values in a vector.
}
\examples{
count_na(c(1, 2, NA, 4, 5))
#> [1] 1 # Expected

count_na(c(1, 2, 3, 4, 5))
#> [1] 0 # Expected

count_na(c(NA, NA, NA, NA, NA))
#> [1] 5 # Expected
}
\seealso{
Other vector functions.: 
\code{\link{drop_na}()},
\code{\link{inbetween_integers}()}
}
\concept{vector functions.}
