% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_interval_mean.R
\name{cut_interval_mean}
\alias{cut_interval_mean}
\title{Compute the interval means of a cut factor}
\usage{
cut_interval_mean(x, round = TRUE)
}
\arguments{
\item{x}{A \code{\link[base:factor]{factor}} or \code{\link[base:character]{character}} vector
with \code{\link[base:cut]{cut}} intervals.}

\item{round}{A \code{\link[base:logical]{logical}} flag indicating whether to round
the means to \verb{0`` decimal places (Default: }TRUE`).}
}
\value{
A \code{\link[base:numeric]{numeric}} vector with the interval means.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{cut_interval_mean()} computes the interval means of a \code{\link[base:cut]{cut}}
factor, regardless of the their open or closed nature.
}
\examples{
cut(1:5, breaks = 3)
#> [1] (0.996,2.33] (0.996,2.33] (2.33,3.67]  (3.67,5] # Expected
#> [5] (3.67,5]
#> Levels: (0.996,2.33] (2.33,3.67] (3.67,5]

cut(1:5, breaks = 3) |> cut_interval_mean()
#> [1] 2 2 3 4 4 # Expected

cut(1:5, breaks = 3) |> cut_interval_mean(round = FALSE)
#> [1] 1.663 1.663 3.000 4.335 4.335 # Expected
}
\seealso{
Other parsing/conversion functions: 
\code{\link{na_as}()}
}
\concept{parsing/conversion functions}
