% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inc.utils.r
\name{ddmergev}
\alias{ddmergev}
\title{merges 2 data.frame using a formula
example of formula is var1 + var2 | key1 + key2}
\usage{
ddmergev(data1, data2, form, verbose = FALSE)
}
\description{
merges 2 data.frame using a formula
example of formula is var1 + var2 | key1 + key2
}

