% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete.data.table}
\alias{delete.data.table}
\title{This function allows deleting a data.table by removing all references to it in a given environment.
In memory-strapped environments, we'd like to help the garbage collector reclaim memory that's not logically referenced anymore.
Note that any errors will be logged without a traceback for uninterrupted execution.
TODO: Note that this function is NOT thread-safe.}
\usage{
delete.data.table(target, environ_objs, environ = .GlobalEnv)
}
\arguments{
\item{target}{data.table object to be deleted}

\item{environ}{environment to search for this object (default: .GlobalEnv)}

\item{environObjs}{list of R objects in the given environment (see: ls())}
}
\description{
This function allows deleting a data.table by removing all references to it in a given environment.
In memory-strapped environments, we'd like to help the garbage collector reclaim memory that's not logically referenced anymore.
Note that any errors will be logged without a traceback for uninterrupted execution.
TODO: Note that this function is NOT thread-safe.
}
