% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_packages.R
\name{desc_pkgs}
\alias{desc_pkgs}
\title{Update DESCRIPTION file with package dependencies}
\usage{
desc_pkgs(
  dir = ".",
  file = "DESCRIPTION",
  suggests = c("^inst/"),
  ignore = c("^data-raw/"),
  must_keep = character(0),
  quiet = TRUE
)
}
\arguments{
\item{dir}{Root package directory}

\item{file}{DESCRIPTION file name}

\item{suggests}{Path(s) to search for dependencies to put in SUGGESTS (supports regular expressions)}

\item{ignore}{Path(s) to ignore in getting dependencies (supports regular expressions)}

\item{must_keep}{List of packages that must be retained even if not a current dependency}

\item{quiet}{Whether to print messages when \code{renv::dependencies} is called}
}
\description{
\code{desc_pkgs} will search code in a package for dependencies,
using both \code{renv::dependencies} to search R and RMD files,
and by detecting dependencies declared with \code{roxygen2} decorators.
It will write these
package dependencies to the DESCRIPTION file.
}
