% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_packages.R
\name{desc_rewrite}
\alias{desc_rewrite}
\title{Rewrite DESCRIPTION file}
\usage{
desc_rewrite(file = "DESCRIPTION", dir = ".")
}
\arguments{
\item{file}{Name of the description file}

\item{dir}{Root package directory}
}
\description{
Wrapper to normalize the DESCRIPTION file in a package by invoking
the functions in \link[desc:description]{desc::description}
}
