% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_document.R
\name{dev_document}
\alias{dev_document}
\title{Document outside of package}
\usage{
dev_document(x, output_directory = "man")
}
\arguments{
\item{x}{path to a \code{function.R} file. if not supplied it will look for the \code{R} folder in the current working directory.}

\item{output_directory}{defaults to \code{man} in the current working directory. will be created if does not exist.}
}
\description{
This function creates an html documentation outside of package.
}
