% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_set_project.R
\name{dev_set_project}
\alias{dev_set_project}
\title{Set Project}
\usage{
dev_set_project(
  path,
  type = c("package", "project"),
  dirs,
  open = FALSE,
  check_name = FALSE,
  ...
)
}
\arguments{
\item{type}{one of  \code{package}, or  \code{project}}

\item{dirs}{vector of dirs to create. If missing, creates \code{"data"},  \code{"analysis"}, and   \code{"R"} folders}

\item{open}{Whether to activate the new project and open a new RStudio sesion.}

\item{check_name}{Whether to check if the name is valid for CRAN and throw an error}

\item{...}{further options to \code{\link[usethis]{create_package}}}
}
\description{
Convenient wrapper around \code{\link[usethis]{create_package}} to set up a project. Only works from within the project.
}
\details{
Workflow:

\enumerate{
    \item Create a GitLab repository and initiate a new project with version control from within RStudio.
    \item run \code{dev_set_project()}
}

Parameters passed to \code{...}

\enumerate{
  \item \code{fields} A named list of fields to add to \code{DESCRIPTION},
                     potentially overriding default values. See \code{\link[usethis]{use_description}}
                     for how you can set personalized defaults using package options.
}
}
