% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_split.R
\name{df_split}
\alias{df_split}
\title{Split a data frame to chunks.}
\usage{
df_split(df, n)
}
\arguments{
\item{df}{a data.frame or object from which nrow() can be taken}

\item{n}{number of chunks}
}
\description{
Calculates the size of the chunk internally
The last chunk being the smallest.
}
\details{
Split a data frame like object to smaller data frames
}
\examples{
df_split(mtcars, 3)
}
