% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{diff.xts}
\alias{diff.xts}
\title{Calculate the time differences of an \code{xts} time series.}
\usage{
\method{diff}{xts}(x_ts, lag = 1)
}
\arguments{
\item{x_ts}{an \code{xts} time series.}

\item{lag}{integer equal to the number of time periods of lag.}
}
\value{
\code{xts} time series with the same dimensions and the same time
  index as the input series.
}
\description{
Calculate the time differences of an \code{xts} time series.
}
\details{
Calculates the time differences of an \code{xts} time series and
  pads with \code{zeros} instead of \code{NAs}.  Positive \code{lag} means
  differences are calculated with values from \code{lag} periods in the past
  (vice versa negative \code{lag}).  The function \code{diff()} is just a
  wrapper for \code{diff.xts()} from package
  \href{https://cran.r-project.org/web/packages/xts/index.html}{xts}, but it
  pads with \code{zeros} instead of \code{NAs}.
}
\examples{
# calculate time differences over lag by 10 periods
rutils::diff.xts(env_etf$VTI, lag=10)
}

