% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{diffit}
\alias{diffit}
\title{Calculate the row differences of a \emph{numeric} or \emph{Boolean} vector,
matrix, or \emph{xts} time series.}
\usage{
diffit(inputv, lagg = 1, ...)
}
\arguments{
\item{\code{inputv}}{A \emph{numeric} or \emph{Boolean} vector or matrix, or
\emph{xts} time series.}

\item{\code{lagg}}{An integer equal to the number of time periods of lag
(the default is \code{1}).}
}
\value{
A vector, matrix, or \emph{xts} time series. with the same dimensions
  as the input object.
}
\description{
Calculate the row differences of a \emph{numeric} or \emph{Boolean} vector,
matrix, or \emph{xts} time series.
}
\details{
The function \code{diffit()} calculates the row differences between
  rows that are \code{lagg} rows apart. Positive \code{lagg} means that the
  difference is calculated as the current row minus the row that is
  \code{lagg} rows above. (vice versa for a negative \code{lagg}).  This also
  applies to vectors, since they can be viewed as single-column matrices. The
  leading or trailing stub periods are padded with \emph{zeros}.

  When applied to \emph{xts} time series, the function \code{diffit()} calls
  the function \code{diff.xts()} from package
  \href{https://cran.r-project.org/web/packages/xts/index.html}{xts}, but it
  pads the output with zeros instead of with \code{NA}s.
}
\examples{
# Diff vector by 2 periods
rutils::diffit(1:10, lagg=2)
# Diff matrix by negative 2 periods
rutils::diffit(matrix(1:10, ncol=2), lagg=-2)
# Diff an xts time series
rutils::diffit(rutils::etfenv$VTI, lagg=10)
}
