% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{diffohlc}
\alias{diffohlc}
\title{Calculate the reduced form of an \emph{OHLC} time series, or calculate the
standard form from the reduced form of an \emph{OHLC} time series.}
\usage{
diffohlc(ohlc, reducit = TRUE, ...)
}
\arguments{
\item{\code{ohlc}}{An \emph{OHLC} time series of prices in \emph{xts} format.}

\item{\code{reducit}}{A \emph{Boolean} argument: should the reduced form be
calculated or the standard form? (default is \code{TRUE})}

\item{\code{...}}{Additional arguments to function \code{xts::diff.xts()}.}
}
\value{
An \emph{OHLC} time series with five columns for the \emph{Open},
  \emph{High}, \emph{Low}, \emph{Close} prices, and the \emph{Volume}, and
  with the same time index as the input series.
}
\description{
Calculate the reduced form of an \emph{OHLC} time series, or calculate the
standard form from the reduced form of an \emph{OHLC} time series.
}
\details{
The reduced form of an \emph{OHLC} time series is obtained by
  calculating the time differences of its \emph{Close} prices, and by
  calculating the differences between its \emph{Open}, \emph{High}, and
  \emph{Low} prices minus the \emph{Close} prices. The standard form is the
  original \emph{OHLC} time series, and can be calculated from its reduced
  form by reversing those operations.
}
\examples{
# Calculate reduced form of an OHLC time series
diff_VTI <- rutils::diffohlc(rutils::etfenv$VTI)
# Calculate standard form of an OHLC time series
VTI <- rutils::diffohlc(diff_VTI, reducit=FALSE)
identical(VTI, rutils::etfenv$VTI[, 1:5])
}
