% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{diffxts}
\alias{diffxts}
\title{Calculate the time differences of an \emph{xts} time series.}
\usage{
diffxts(xtsv, lagg = 1, ...)
}
\arguments{
\item{\code{xtsv}}{An \emph{xts} time series.}

\item{\code{lagg}}{An integer equal to the number of time periods of lag
(the default is \code{1}).}

\item{\code{...}}{Additional arguments to function \code{xts::diff.xts()}.}
}
\value{
An \emph{xts} time series with the same dimensions and the same time
  index as the input series.
}
\description{
Calculate the time differences of an \emph{xts} time series.
}
\details{
The function \code{diffxts()} calculates the time differences of an
  \emph{xts} time series and pads with \emph{zeros} instead of \code{NA}s.
  Positive \code{lagg} means differences are calculated with values from
  \code{lagg} periods in the past (vice versa for a negative \code{lagg}).
  The function \code{diff()} is just a wrapper for \code{diff.xts()} from
  package \href{https://cran.r-project.org/web/packages/xts/index.html}{xts},
  but it pads with \emph{zeros} instead of \code{NA}s.

  The function \code{diffit()} has incorporated the functionality of
  \code{diffxts()}, so that \code{diffxts()} will be retired in future
  package versions.
}
\examples{
# Calculate time differences over lag by 10 periods
rutils::diffxts(rutils::etfenv$VTI, lag=10)
}
