% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{display_schema}
\alias{display_schema}
\title{Display JSON schema}
\usage{
display_schema(jsonschema, cds_package = "rvispack")
}
\arguments{
\item{jsonschema}{a JSON file stored in ext/data}

\item{cds_package}{a containerized data science package that has the `jsonchema` parameter}
}
\description{
Displays detailed information about the parameters required in a JSON object.
}
