\name{dist}
\alias{dist}
\title{computes normalized distance between 2 vectors
error will recover the context. The next time, the error is back to normal
and the context is not recovered}
\usage{
  dist(M1, M2, type = 0)
}
\description{
  computes normalized distance between 2 vectors error will
  recover the context. The next time, the error is back to
  normal and the context is not recovered
}

