% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{dloc}
\alias{dloc}
\title{Default Location}
\usage{
dloc(..., .init = FALSE)
}
\arguments{
\item{...}{Character vectors representing file paths. Long vectors are not supported.}

\item{.init}{Logical, if TRUE, initializes the default directory. Default is FALSE.}
}
\value{
A path or a directory. If \code{.init} is set to TRUE, the function returns NULL.
}
\description{
This function sets the default directory and returns the path of the files and directories
within that directory. The function also initializes the default directory if \code{.init} is set to TRUE.
}
