% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_shell.R
\name{docker_run_rserver}
\alias{docker_run_rserver}
\title{Run an RStudio instance on the bioconductor docker image}
\usage{
docker_run_rserver(
  image = "bioconductor/bioconductor_docker:devel",
  port = 8888,
  user = "rstudio",
  password = "bioc",
  name = "dz_bioc",
  rm = FALSE,
  volumes_ro = NULL,
  volumes = NULL,
  verbose = FALSE
)
}
\arguments{
\item{image}{\code{character(1)} name of the image to use}

\item{port}{\code{integer(1)} port of the host which will be used to access
RStudio server via \code{localhost::port}}

\item{user}{\code{character(1)} a UN for the docker image - by default this will
be part of the group ID 1024. The group is what permits \code{volumes} to be
write-able. Volumes will inherit permissions from the host and therefore,
must have group write permissions on the host (based on
https://dille.name/blog/2018/07/16/handling-file-permissions-when-writing-to-volumes-from-docker-containers/)}

\item{password}{\code{character(1)} password to used for RStudio server}

\item{name}{\code{character(1)} name for the container.}

\item{rm}{\code{logical(1)} whether to use the \code{-rm} flag when running the
container. Should the container be removed when stopped?}

\item{volumes_ro}{\code{character(1)} paths for the hosts files that you want
accessible in the container. These will be mounted using the identical
paths on the host and with \code{read-only} permissions.}

\item{volumes}{\code{character(1)} paths for the hosts files that you want
accessible in the container. These will NOT be mounted with \code{read-only}
permissions. Use this option with care, recommended only for directories
that you want to output results into.}

\item{verbose}{\code{logical(1)} whether to display the command to be run (for
debugging purposes).}
}
\description{
A wrapper for the docker function used to run a RStudio container on a
specified docker image. This will be available at the \code{port} specified on the
host.
}
\examples{

\dontrun{
docker_run_rserver()
}
}
\references{
\code{docker_run_rserver} is based on
\verb{https://github.com/LieberInstitute/megadepth/blob/master/R/megadepth.R}.
}
