% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{.debug}
\alias{.debug}
\title{Logs an DEBUG message}
\usage{
.debug(...)
}
\arguments{
\item{...}{params passed to the logger}
}
\description{
Logs an DEBUG message
}
\note{
watchout: don't call this "debug": Baaad things will happen
}
