% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Download files}
\usage{
download_file(
  url,
  connection_timeout = 10,
  max_tries = 3,
  retry_on_failure = TRUE,
  backoff = function(attempt) 5^attempt,
  dir = tempdir()
)
}
\arguments{
\item{url}{A \code{\link[base:character]{character}} vector of URLs pointing to
remote files.}

\item{connection_timeout}{(optional) A \code{\link[base:numeric]{numeric}} value
specifying the connection timeout in seconds for HTTP requests
(default: \code{10}).}

\item{max_tries}{(optional) A \code{\link[base:numeric]{numeric}} value specifying
the maximum number of retry attempts (default: \code{3}).}

\item{retry_on_failure}{(optional) A \code{\link[base:logical]{logical}} value
indicating whether to retry on failure (default: \code{TRUE}).}

\item{backoff}{(optional) A \code{\link[base:function]{function}} that takes the
current attempt number as input and returns the number of seconds to wait
before the next attempt (default: \verb{\\(attempt) 5^attempt}).}

\item{dir}{(optional) A string specifying the directory where the files
should be downloaded (default: \code{tempdir()}).}
}
\value{
A invisible \code{\link[base:character]{character}} vector of file paths
where the files were downloaded.
}
\description{
\code{download_file()} is a wrapper around the \code{httr2} package that provides a
user-friendly interface for downloading files, with built-in error handling
and progress reporting.
}
\examples{
library(httr2)

if (is_online()) {
  urls <- file.path(
    "ftp://ftp.datasus.gov.br/dissemin/publicos/IBGE/POPSVS",
     c("POPSBR00.zip", "POPSBR01.zip")
  )

  download_file(urls)
}
}
\seealso{
Other file functions: 
\code{\link{get_file_size}()}
}
\concept{file functions}
