% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na.R
\name{drop_na}
\alias{drop_na}
\title{Remove \code{NA} values from an atomic object}
\usage{
drop_na(x)
}
\arguments{
\item{x}{An \code{\link[checkmate:checkAtomic]{atomic}} object of any type.}
}
\value{
An object of the same type as \code{x} with all \code{NA} values removed.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{drop_na()} remove all \code{NA} values from an atomic object.
}
\examples{
drop_na(c(1, NA, 2, 3, NA, 4))
#> [1] 1 2 3 4 # Expected

drop_na(c("a", NA, "b", "c"))
#> [1] "a" "b" "c" # Expected
}
\concept{vector functions.}
