% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{end_points}
\alias{end_points}
\title{Create an index (integer vector) of equally spaced end points for a time
series.}
\usage{
end_points(x_ts, inter_val = 10, off_set = 0)
}
\arguments{
\item{x_ts}{vector or time series.}

\item{inter_val}{the number of data points per interval.}

\item{off_set}{the number of data points in the first interval (stub
interval).}
}
\value{
\code{integer} vector of equally spaced end points.
}
\description{
Create an index (integer vector) of equally spaced end points for a time
series.
}
\details{
The end points divide the time series into equally spaced intervals.
  The off_set argument shifts the end points forward and creates an initial
  stub interval.
}
\examples{
# create end points with initial stub interval
end_points(env_etf$VTI, inter_val=7, off_set=4)
}

