% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/eval.expression.R
\name{eval.expression}
\alias{eval.expression}
\title{Metaprogramming}
\usage{
eval.expression(expression.as.character)
}
\arguments{
\item{expression.as.character}{a vector of length 1 containing the
code to be executed as a character element.}
}
\description{
Helper function to parse a vector with a single character element as
code.
}
\examples{
eval.expression("n <- 1 + 1")
cat(n)
}

