% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_expr.R
\name{eval_expr}
\alias{eval_expr}
\title{Metaprogramming}
\usage{
eval_expr(expr)
}
\arguments{
\item{expr}{a vector of length 1 containing the
code to be executed as a character element.}
}
\description{
Helper function to parse a vector with a single character element as
code.
}
\examples{

eval_expr('n <- 1 + 1')
cat(n)
## 2

}

