% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evalExpression.R
\name{evalExpression}
\alias{evalExpression}
\title{Metaprogramming}
\usage{
evalExpression(expressionAsCharacter)
}
\arguments{
\item{expressionAsCharacter}{a vector of length 1 containing the
code to be executed as a character element.}
}
\description{
Helper function to parse a vector with a single character element as
code.
}
\examples{
evalExpression("n <- 1 + 1")
cat(n) ## 2
}

