% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{ex_tract}
\alias{ex_tract}
\title{Extract columns of prices from an \code{OHLC} time series.}
\usage{
ex_tract(oh_lc, col_name = "Close")
}
\arguments{
\item{oh_lc}{an \code{OHLC} time series.}

\item{col_name}{string with the field name of the column to be be extracted.
(default is "Close")}
}
\value{
A single column \code{OHLC} time series in \code{xts} format.
}
\description{
Extract columns of prices from an \code{OHLC} time series.
}
\details{
Extracts columns of prices from an \code{OHLC} time series by
  \code{grepping} column names for the \code{col_name} string. The
  \code{OHLC} column names are assumed to be in the format
  \code{"symbol.field_name"}, for example \code{"VTI.Close"}. Performs a
  similar operation to the extractor functions \code{Op()}, \code{Hi()},
  \code{Lo()}, \code{Cl()}, and \code{Vo()}, from package
  \href{https://cran.r-project.org/web/packages/quantmod/index.html}{quantmod}.
   But \code{ex_tract()} is able to handle symbols like \code{"LOW"}, which
  the function \code{Lo()} can't handle. The col_name argument is partially
  matched, for example "vol" is matched to "Volume".
}
\examples{
# get close prices for VTI
ex_tract(env_etf$VTI)
# get volumes for VTI
ex_tract(env_etf$VTI, col_name="vol")
}

