% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/execute.with.parallel.backend.R
\name{execute.with.cpu.parallel.backend}
\alias{execute.with.cpu.parallel.backend}
\title{CPU Parallel Processing}
\usage{
execute.with.cpu.parallel.backend(task.to.be.executed, n.cores = NULL, ...)
}
\arguments{
\item{task.to.be.executed}{a function representing a task to be executed
with parallel processing.}

\item{n.cores}{the numbers of CPU cores to be used in the parallel
processing. By default, it uses the numbers of cores returned by the
\code{\link[parallel]{detectCores}} function. It must be an integer.}

\item{...}{additional paramaters to be delegated to \code{task.to.be.executed}
function.}
}
\description{
Helper function which automatically configures a CPU parallel
backend either on Linux or Windows (MAC OS is not supported).
It is specially useful when there is a requirement to write R
code that should be portable between windows and linux.
}
\details{
This function saves in global environment either a boolean variable
called \code{linux.parallel.backend} or \code{windows.parallel.backend},
indicating if the parallel backend is available for use and the type of
the operating system in which the code is being run.
}
\examples{
library(foreach)

# a function simulating an useful parallel task
summation <- function(){
   # saves the result in the parent enviroment
   result <<- sum(foreach(i = 1:10000, .combine = "c") \%dopar\% {sqrt(i)})
}

# this code is portable between windows and linux, if the right packages
# are installed in both of them.
execute.with.cpu.parallel.backend(summation)
print(result) ## 666716.5
}

