% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/execute.with.linux.cpu.parallel.backend.R
\name{execute.with.linux.cpu.parallel.backend}
\alias{execute.with.linux.cpu.parallel.backend}
\title{CPU Parallel Processing in Linux}
\usage{
execute.with.linux.cpu.parallel.backend(task.to.be.executed, n.cores = NULL,
  ...)
}
\arguments{
\item{task.to.be.executed}{a function representing a task to be executed
with parallel processing.}

\item{n.cores}{the numbers of CPU cores to be used in the parallel
processing. By default, it uses the numbers of cores returned by the
\code{\link[parallel]{detectCores}} function. It must be an integer.}

\item{...}{additional paramaters to be delegated to \code{task.to.be.executed}
function.}
}
\description{
Utility function which executes a task with automatic configured linux parallel
backend. Under the hoods, it uses the \code{\link[doMC]{registerDoMC}}
function from the \code{doMC} package.
}
\details{
This function also saves in the global environment a boolean variable called
\code{linux.parallel.backend} indicating if the parallel backend is
available to use.
}
\examples{
library(foreach)

# a function simulating an useful parallel task
summation <- function(){
   # saves the result in the parent enviroment
   result <<- sum(foreach(i = 1:10000, .combine = "c") \%dopar\% {sqrt(i)})
}

execute.with.linux.cpu.parallel.backend(summation)
print(result) ## 666716.5
}

