% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expm1.R
\name{Expm1}
\alias{Expm1}
\title{Expm1}
\usage{
Expm1(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric, same size as x
}
\description{
Compute exp(x)-1 in a numerically-stable fashion
}
\examples{
Expm1(1e-20)
# compare to naive approach:
exp(1e-20)-1
}
