% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_initials.R
\name{extract_initials}
\alias{extract_initials}
\title{Extract initials from a character vector}
\usage{
extract_initials(x, sep = ". ")
}
\arguments{
\item{x}{A \code{\link[base:character]{character}} vector with names.}

\item{sep}{A \code{\link[base:character]{character}} string to separate initials
(Default: ". ").}
}
\value{
A \code{\link[base:character]{character}} vector with the initials of
names in \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{extract_initials()} extracts the initials of names in a character
vector.
}
\examples{
extract_initials(c("John Doe", "Jane Doe", "John Smith"))
#> [1] "J. D." "J. D." "J. S." # Expected

extract_initials(c("John Doe", "Jane Doe", "John Smith"), sep = ".")
#> [1] "J.D." "J.D." "J.S." # Expected
}
\concept{string functions.}
