% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_path.R
\name{file_path}
\alias{file_path}
\title{Create a file path and/or check if it exists}
\usage{
file_path(..., check_exists = TRUE, force_exists = FALSE)
}
\arguments{
\item{...}{passed to \code{base::file.path}.}

\item{check_exists}{\code{logical(1)} specifying whether to print the path and if
it exists locally.}

\item{force_exists}{\code{logical(1)} specifying whether to return an error if the
path does not exist.}
}
\value{
file path
}
\description{
Creates a file path and also can check/force it to exists in one step. Used
mostly in conjunction with \code{here::here}, for checking whether a file exists
(you've typed the path correctly).
}
\examples{

file_path("/")
}
