% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{filter_unprocessed_files}
\alias{filter_unprocessed_files}
\title{Filter Unprocessed Files}
\usage{
filter_unprocessed_files(.tab, .dir, .verbose = TRUE)
}
\arguments{
\item{.tab}{A data frame that should at least contain the column 'doc_id' (usually created by lft()).}

\item{.dir}{Directory where the files are located.}

\item{.verbose}{Logical. If TRUE (default), prints out the number of files to be processed.}
}
\value{
A filtered data frame containing only the unprocessed files.
}
\description{
This function filters out files that have already been processed.
}
