% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_orphan_files.R
\name{find_orphan_files}
\alias{find_orphan_files}
\title{Find orphan files in a Zotero library}
\usage{
find_orphan_files(
  lib_file = tcltk::tk_choose.files(caption = "Select the Zotero library CSV file", multi
    = FALSE),
  file_folder = tcltk::tk_choose.dir(caption = "Select the folder containing the files")
)
}
\arguments{
\item{lib_file}{(optional) A string specifying the path to the Zotero library
exported as a \strong{CSV} file.
(default: \code{\link[tcltk:tk_choose.files]{tk_choose.files()}}).}

\item{file_folder}{(optional) A string specifying the path to the folder
containing the files linked to references in the Zotero library.
(default: \code{\link[tcltk:tk_choose.dir]{tk_choose.dir()}}).}
}
\value{
A \code{\link[base:character]{character}} vector with the names of the
orphan files.
}
\description{
\code{find_orphan_files()} compares the files present in a specified folder
with those linked to references in a Zotero library, returning the names
of files that are not linked (i.e., orphan files).
}
\details{
To export your library from Zotero, go to the menu \verb{File > Export Library...}
and choose the \strong{CSV} format.
}
\examples{
\dontrun{
  find_orphan_files()
}
}
\seealso{
Other Zotero functions: 
\code{\link{list_linked_files}()}
}
\concept{Zotero functions}
