% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.R
\name{first}
\alias{first}
\title{Element Retrieval}
\usage{
first(x, is_list = FALSE)
}
\arguments{
\item{x}{a vector or list}

\item{is_list}{is x a list?}
}
\description{
Retrieve the first element of a vector or list.
}
\examples{

first(1:100)
## 1

}

