% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/flypwd.r
\name{flypwd}
\alias{flypwd}
\title{Returns the password stored by flypwd}
\usage{
flypwd(service = NULL, clean = FALSE)
}
\value{
the current user's password
}
\description{
Returns the password stored by flypwd
}
\author{
Giuseppe Acito
}

