% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy.R
\name{fuzzy}
\alias{fuzzy}
\title{Fuzzy functions}
\usage{
fuzzy(
  shape = c("trapezoid", "triangular", "normal", "linear"),
  breaks = c(1, 2, 3, 4)
)
}
\arguments{
\item{shape}{of the relationship one of \code{c("trapezoid", "triangular", "normal")}}

\item{breaks}{cornerpoints for the inflection points. 4 - trapezoid, 3 - triangular}
}
\description{
Returns a fuzzy function
}
\details{
TODO implement those in https://search.r-project.org/CRAN/refmans/sets/html/fuzzyfuns.html
}
\examples{

x <- 1:100
trap <- fuzzy("trapezoid", breaks = c(20, 40, 60, 80))
plot(x, trap(x))
trian <- fuzzy("triangular", breaks = c(25, 50, 75))
plot(x, trian(x))
}
