% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_unique_file.r
\name{gen_unique_filename}
\alias{gen_unique_filename}
\title{Generates a unique file name in a directory \code{prefix}}
\usage{
gen_unique_filename(n = 7, prefix = tempdir(), ext = NA, values = letters)
}
\arguments{
\item{n}{number of character per file}

\item{prefix}{directory where the file must be created}

\item{ext}{filename extension}

\item{values}{available values for random filenamem creation
defaults to \code{letters}}
}
\value{
the path where it's safe to create a unique filename
}
\description{
it's safe to call this in a multithreaded environment
}
