% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMessage.R
\name{GenerateMessage}
\alias{GenerateMessage}
\title{GenerateMessage}
\usage{
GenerateMessage(message.text)
}
\arguments{
\item{\code{message.text}}{The text to print}
}
\description{
Print a message to stdout with the date and \code{message.text}
}
