% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeometricMean.R
\name{GeometricMean}
\alias{GeometricMean}
\title{GeometricMean}
\usage{
GeometricMean(x, na.rm = TRUE, zero.propagate = FALSE)
}
\arguments{
\item{x}{Numerical vector}

\item{na.rm}{Remove NA values from x}

\item{zero.propagate}{Propagate zeroes from x}
}
\value{
Numeric: geometrical mean of x
}
\description{
Take the geometric mean of a numerical vector. Stolen from
\code{http://stackoverflow.com/a/25555105/3017012}.
}
