% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_and_aggregate_sidra_by_year.R
\name{get_and_aggregate_sidra_by_year}
\alias{get_and_aggregate_sidra_by_year}
\title{Aggregate data by year from SIDRA API}
\usage{
get_and_aggregate_sidra_by_year(years, api_start, api_end, year_col = "Ano")
}
\arguments{
\item{years}{A \code{\link[checkmate:checkIntegerish]{integerish}} vector with the
years to download.}

\item{api_start}{A string with the start of the API URL.}

\item{api_end}{A string with the end of the API URL. It's important to note
that the year will be pasted after \code{api_start} and before \code{api_end}.}

\item{year_col}{A string with the name of the column that contains the year.}
}
\value{
A \link[tibble:tibble]{tibble} with data from the SIDRA API.
}
\description{
This function downloads data from the Brazilian Institute of Geography and
Statistics (\href{https://www.ibge.gov.br/}{IBGE}) Automatic Retrieval System
(\href{https://sidra.ibge.gov.br/}{SIDRA}) API for a given range of years.
}
\details{
To get the API call, go to the SIDRA, look for your data in one of the
tables, then click on the share button and copy the \emph{Parâmetros para a API}
value. The API call starts with "/t" (see the examples).
}
\examples{
\dontrun{
  sidrar_aggregator(
    years = 2010:2011,
    api_start = "/t/1612/n6/all/v/109/p/",
    api_end = "/c81/2692"
  )
}
}
\author{
Daniel Vartanian
}
