% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-package-functions.R
\name{get_args}
\alias{get_args}
\title{Get the inputs to a function}
\usage{
get_args(f, ...)
}
\arguments{
\item{f}{A function}

\item{...}{Arguments to pass to \code{f}}
}
\value{
A named list
}
\description{
This gets all inputs to a function \code{f}, including
default arguments, when the arguments in \code{...} are
passed. All returned arguments are named, except those
which would be in a \code{...} argument, if \code{f} has one.
}
\examples{
get_args(paste, "hi", "there!")
}
