% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_functions.R
\name{get_auc}
\alias{get_auc}
\title{Get AUC}
\usage{
get_auc(results)
}
\arguments{
\item{results}{data frame with predicted probabilities under column named 'prob' and results
under column named 'actual'}
}
\value{
AUC of the ROC curve
}
\description{
\code{get_auc} returns the area under the ROC curve for a list of predictions and actuals
}
