% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_city_admin_info.R
\name{get_city_admin_info}
\alias{get_city_admin_info}
\title{Retrieve French city INSEE code}
\usage{
get_city_admin_info(city, county = NULL, state = NULL, limit = 1L)
}
\arguments{
\item{city}{a \code{character} of length 1.}

\item{county}{(optional) a \code{character} of length 1. The county (departement)
of the city used to refine the query.}

\item{state}{(optional) a \code{character} of length 1. The state (region)
of the city used to refine the query.}

\item{limit}{an \code{integer} of length 1. The number of results to return.
Default is \code{1}.}
}
\value{
A \code{data.frame} with the following variables:
\itemize{
\item \code{input}: the value of the argument \code{city}
\item \code{village}: the name of the \code{city} in Open Street Map database
\item \code{city}: the name of the \code{city} in Open Street Map database
\item \code{town}: the name of the \code{city} in Open Street Map database
\item \code{municipality}: the name of the \code{city} in Open Street Map database
\item \code{county}: the name of the French departement
\item \code{state}: the name of the French region
\item \code{region}: the name of the region of France
\item \code{country}: the name of the region of France
\item \code{postcode}: the \emph{Code postal}
\item \code{code_insee}: the \emph{Code officiel geographique}
\item \code{longitude}: the longitude of the administrative center of the city
\item \code{latitude}: the latitude of the administrative center of the city
}
}
\description{
Uses the Nominatim API \url{https://nominatim.org/release-docs/develop/api/Search/}
to retrieve the \emph{Code officiel geographique (COG)} and the \emph{Code postal} of a
French town.
}
\examples{
get_city_admin_info(city = "Florac")
get_city_admin_info(city = "La Salle-Prunet")
get_city_admin_info(city = "Florac-Trois-Rivieres")
}
