% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{get_col}
\alias{get_col}
\title{Extract columns of data from an \emph{OHLC} time series using column field
names.}
\usage{
get_col(oh_lc, field_name = "Close")
}
\arguments{
\item{oh_lc}{an \emph{OHLC} time series.}

\item{field_name}{string with the field name of the column to be be extracted.
(default is \emph{Close})}
}
\value{
The specified columns of the \emph{OHLC} time series in a single
  \emph{xts} series, with the same number of rows as the input time series.
}
\description{
Extract columns of data from an \emph{OHLC} time series using column field
names.
}
\details{
Extracts columns from an \emph{OHLC} time series by partially
  matching field names with column names. The \emph{OHLC} column names are
  assumed to be in the format \code{"symbol.field_name"}, for example
  \code{"VTI.Close"}. Performs a similar operation to the extractor functions
  \code{Op()}, \code{Hi()}, \code{Lo()}, \code{Cl()}, and \code{Vo()}, from
  package
  \href{https://cran.r-project.org/web/packages/quantmod/index.html}{quantmod}.
   But \code{get_col()} is able to handle symbols like \emph{LOW}, which the
  function \code{Lo()} can't handle. The field_name argument is partially
  matched, for example "Vol" is matched to \code{Volume} (but it's case
  sensitive).
}
\examples{
# get close prices for VTI
rutils::get_col(rutils::env_etf$VTI)
# get volumes for VTI
rutils::get_col(rutils::env_etf$VTI, "Vol")
# get close prices and volumes for VTI
rutils::get_col(rutils::env_etf$VTI, c("Cl", "Vol"))
}

