% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{get_data}
\alias{get_data}
\title{Load \emph{OHLC} time series data into an environment, either from an
external source (download from \emph{YAHOO}), or from \emph{CSV} files in a
local drive.}
\usage{
get_data(
  symbolv,
  datadir = NULL,
  datenv,
  startd = "2007-01-01",
  endd = Sys.Date(),
  func = match.fun("as.Date"),
  formatv = "\%Y-\%m-\%d",
  header = TRUE,
  echo = TRUE,
  scrub = TRUE
)
}
\arguments{
\item{\code{symbolv}}{A vector of strings representing instrument symbols
(tickers).}

\item{\code{datadir}}{The directory containing \emph{CSV} files (the default is
\code{NULL}).}

\item{\code{datenv}}{The environment for loading the data into.}

\item{\code{startd}}{The start date of time series data (the default is
"2007-01-01").}

\item{\code{endd}}{The end date of time series data (the default is
\code{Sys.Date()}).}

\item{\code{func}}{The name of the function for formatting the date
fields in the \emph{CSV} files (the default is \code{as.Date()}).}

\item{\code{formatv}}{The format of the date fields in the \emph{CSV} files
(the default is \code{\%Y-\%m-\%d}).}

\item{\code{header}}{A \emph{Boolean} argument: if \code{TRUE} then read the
header in the \emph{CSV} files (the default is \code{TRUE}).}

\item{\code{echo}}{A \emph{Boolean} argument: if \code{TRUE} then print to
console information on the progress of \emph{CSV} file loading (the default is
\code{TRUE}).}

\item{\code{scrub}}{A \emph{Boolean} argument: if \code{TRUE} then remove
\code{NA} values using function \code{rutils::nalocf()} (the default is
\code{TRUE}).}
}
\value{
A vector of \code{symbolv} returned invisibly.
}
\description{
Load \emph{OHLC} time series data into an environment, either from an
external source (download from \emph{YAHOO}), or from \emph{CSV} files in a
local drive.
}
\details{
The function \code{get_data()} loads \emph{OHLC} time series data
  into an environment (as a side-effect), and returns invisibly the vector of
  \code{symbolv}.

  If the argument \code{datadir} is specified, then \code{get_data()} loads
  from \emph{CSV} files in that directory, and overwrites \code{NA} values if
  \code{scrub=TRUE}.
  If the argument \code{datadir} is \emph{not} specified, then
  \code{get_data()} downloads adjusted \emph{OHLC} prices from \emph{YAHOO}.

  The function \code{get_data()} calls the function \code{getSymbols.yahoo()}
  for downloading data from \emph{YAHOO}, and performs a similar operation to
  the function \code{getSymbols()} from package
  \href{https://cran.r-project.org/web/packages/quantmod/index.html}{quantmod}.
  But \code{get_data()} is faster because it performs less overhead
  operations, and it's able to handle symbols like \emph{LOW}, which
  \code{getSymbols()} can't handle because the function \code{Lo()} can't
  handle them. The \code{startd} and \code{endd} must be either of
  class \emph{Date}, or a string in the format \emph{"YYYY-mm-dd"}.
}
\examples{
\dontrun{
newenv <- new.env()
# Load prices from local csv files
rutils::get_data(symbolv=c("SPY", "EEM"),
            datadir="C:/Develop/data/bbg_records",
            datenv=newenv)
# Download prices from YAHOO
rutils::get_data(symbolv=c("MSFT", "XOM"),
            datenv=newenv,
            startd="2012-12-01",
            endd="2015-12-01")
}
}
