% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{get_dups}
\alias{get_dups}
\title{Get Duplicates in a Data Frame}
\usage{
get_dups(.tab, ...)
}
\arguments{
\item{.tab}{A data frame or tibble to check for duplicates.}

\item{...}{Columns to consider for identifying duplicates. These should be unquoted column names.}
}
\value{
A tibble containing the rows from the original data frame that have duplicates, along with additional columns:
\describe{
\item{id_dup}{A unique identifier for each group of duplicate rows.}
\item{n_dup}{The number of occurrences of each duplicate group.}
}
}
\description{
This function identifies duplicate rows in a data frame based on specified columns.
}
\details{
The Function groups the data by the specified columns, assigns a unique identifier to each group of duplicates,
counts the number of duplicates, and then filters out non-duplicate rows.
The resulting tibble is sorted by the duplicate identifier and includes all original columns
with the addition of \code{id_dup} and \code{n_dup}.
}
\examples{
df <- tibble::tibble(
  id = 1:10,
  group = c(rep("A", 3), rep("B", 4), rep("C", 3)),
  value = c(1, 2, 2, 3, 3, 3, 4, 5, 5, 6)
)
get_dups(df, group, value)

}
