% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_ext.R
\name{get_file_ext}
\alias{get_file_ext}
\title{Get file extension}
\usage{
get_file_ext(file)
}
\arguments{
\item{file}{A \code{\link[base:character]{character}} string with the file paths.}
}
\value{
A \code{\link[base:character]{character}} string containing the file
extension of the files.
}
\description{
\code{get_file_ext()} is similar to \code{\link[tools:fileutils]{file_ext()}}, but it
returns the file extension with the dot separator (e.g., \code{.csv}).
}
\examples{
get_file_ext("example.txt")
#> [1] ".txt" # Expected

get_file_ext("/path/to/file.tar.gz")
#> [1] ".gz"  # Expected

get_file_ext(c("data.csv", "data.rds"))
#> [1] ".csv" ".rds" # Expected
}
\concept{string functions}
