% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_name_without_ext.R
\name{get_file_name_without_ext}
\alias{get_file_name_without_ext}
\title{Get file names without the extension}
\usage{
get_file_name_without_ext(file)
}
\arguments{
\item{file}{A \code{\link[base:character]{character}} vector with file paths.}
}
\value{
A \code{\link[base:character]{character}} vector containing file names
without the extension part.
}
\description{
\code{get_file_name_without_ext()} returns file names without the extension
part.
}
\examples{
get_file_name_without_ext("example.txt")
#> [1] "example" # Expected

get_file_name_without_ext("/path/to/file.tar.gz")
#> [1] "file.tar"  # Expected

get_file_name_without_ext("no-extension")
#> [1] "no-extension" # Expected

get_file_name_without_ext(c("data.csv", "data.rds"))
#> [1] "data" "data" # Expected
}
\seealso{
Other string functions: 
\code{\link{get_file_ext}()}
}
\concept{string functions}
