% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_size.R
\name{get_file_size}
\alias{get_file_size}
\title{Get the sizes of local files or files from URLs}
\usage{
get_file_size(
  file,
  connection_timeout = 10,
  max_tries = 3,
  retry_on_failure = TRUE
)
}
\arguments{
\item{file}{A \code{\link[base:character]{character}} vector of file paths.
The function also works with URLs.}

\item{connection_timeout}{(optional) A \code{\link[base:numeric]{numeric}} value
specifying the connection timeout in seconds for HTTP requests
(default: \code{10}).}

\item{max_tries}{(optional) A \code{\link[base:numeric]{numeric}} value specifying
the maximum number of retry attempts (default: \code{3}).}

\item{retry_on_failure}{(optional) A \code{\link[base:logical]{logical}} value
indicating whether to retry on failure (default: \code{TRUE}).}
}
\value{
A \code{\link[fs:fs_bytes]{fs_bytes}} vector of file sizes.
}
\description{
\code{get_file_size()} returns the sizes of files in bytes. It works with local
and remote files.
}
\examples{
library(fs)
library(readr)

files <- c("file1.txt", "file2.txt", "file3.txt")

dir <- tempfile("dir")
dir.create(dir)

for (i in files) {
  letters |>
    rep(sample(1000:10000, 1)) |>
    write_lines(file.path(dir, i))
}

urls <- c(
  paste0(
    "https://geodata.ucdavis.edu/climate/worldclim/2_1/base/",
    "wc2.1_2.5m_tavg.zip"
  ),
  paste0(
    "https://geodata.ucdavis.edu/climate/worldclim/2_1/base/",
    "wc2.1_10m_prec.zip"
  )
)

c(urls, path(dir, files)) |> get_file_size()
}
\seealso{
Other file functions: 
\code{\link{download_file}()}
}
\concept{file functions}
