% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locale.R
\name{get_locale}
\alias{get_locale}
\title{Get the current locale settings}
\usage{
get_locale()
}
\value{
An invisible \code{NULL}. This function is called for its side effects.
}
\description{
\code{get_locale()} prints the current locale settings in system and local
environments.
}
\examples{
get_locale()
}
\seealso{
Other system functions.: 
\code{\link{set_en_us_locale}()}
}
\concept{system functions.}
