% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{get_name}
\alias{get_name}
\title{Extract a symbol name (ticker) from a \emph{character} string.}
\usage{
get_name(str_ing, separator = "_", field = 1)
}
\arguments{
\item{str_ing}{a vector of \emph{character} strings containing symbol names.}

\item{separator}{the name separator, i.e. the single \emph{character} that
separates the symbol name from the rest of the string. (default is "_")}

\item{field}{the position of the name in the string, i.e. the integer index
of the field to be extracted.  (default is \code{1}, i.e. the name is at
the beginning of the string,)}
}
\value{
A vector of \emph{strings} with the names of objects.
}
\description{
Extract a symbol name (ticker) from a \emph{character} string.
}
\details{
Extracts the symbol name from a \emph{character} string. For
  example, it extracts the name "XLU" from the string "XLU_2017_09_05.csv".
  The input string is assumed to be in the format "\emph{name}_date.csv",
  with the name at the beginning of the string, but \code{get_name()} can
  also parse the name from other string formats as well.  If the input is a
  vector of strings, then \code{get_name()} returns a vector of names.
}
\examples{
# extract the ticker symbol from file names
rutils::get_name("XLU_2017_09_05.csv")
rutils::get_name("XLU 2017 09 05.csv", sep=" ")
rutils::get_name("XLU.csv", sep="[.]")
}

