% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_os}
\alias{get_os}
\title{Get OS}
\usage{
get_os()
}
\value{
3-lower-cased-characters string
}
\description{
Identify the OS you're currently running on.
}
\examples{
get_os()
}
\author{
Hadley https://github.com/r-lib/rappdirs/blob/master/R/utils.r#L1
}
\keyword{OS}
