% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phylopic_image.R
\name{get_phylopic_image}
\alias{get_phylopic_image}
\title{Download a PhyloPic silhouette}
\usage{
get_phylopic_image(uuid, size = 512)
}
\arguments{
\item{uuid}{a \code{character} of length 1. The UUID (unique taxa identifier in
the Phylopic database).}

\item{size}{an \code{integer} of length 1. The size of the PNG to download. An
error will be raised in the size is not available.}
}
\value{
An array of dimensions red x green x blue x alpha.
}
\description{
This function downloads a silhouette from the
\href{https://www.phylopic.org/}{PhyloPic} database. With the new API, some
functions of the package
\href{https://github.com/palaeoverse-community/rphylopic}{\code{rphylopic}} don't
work anymore. This function aims to temporary replace the function
\code{rphylopic::image_data()} as this package seems to be looking for a new
maintainer.
}
\examples{
\dontrun{
mammals_pic <- get_phylopic_image("8cad2b22-30d3-4cbd-86a3-a6d2d004b201")

## Plot the silhouette ----
ggplot2::ggplot(x = Sepal.Length, y = Sepal.Width, data = iris, 
                geom = "point") +
rphylopic::add_phylopic(mammals_pic, alpha = 1)
}
}
