% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_functions.R
\name{get_pr}
\alias{get_pr}
\title{Get PR}
\usage{
get_pr(results)
}
\arguments{
\item{results}{data frame with predicted probabilities under column named 'prob' and numeric
results under column named 'actual_num'}
}
\value{
AUC of the PR curve
}
\description{
\code{get_auc} returns the area under the PR curve for a list of predictions and actuals
}
