% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{get_templates}
\alias{get_templates}
\title{Get Templates}
\usage{
get_templates(.template = NULL, .name, .open = TRUE)
}
\arguments{
\item{.template}{Not relevant (NULL by default).}

\item{.name}{File Name.}

\item{.open}{Logical, if TRUE, opens the script in a web browser. Default is TRUE.}
}
\value{
Copied and modified template file.
}
\description{
This function retrieves a template file from the "rUtils" package, replaces a placeholder
with the given file name, and saves the modified file to the "1_code" directory.
Optionally, the function can open the file in a web browser.
}
