% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_world_basemap.R
\name{get_world_basemap}
\alias{get_world_basemap}
\title{Download IPBES World countries boundaries}
\usage{
get_world_basemap(path = ".", force = FALSE)
}
\arguments{
\item{path}{a \code{character} of length 1. The folder inside which the \code{zip}
will be downloaded and extracted.
Default is \code{getwd()}.}

\item{force}{a \code{logical}. If \code{TRUE} previous files will be erased.
Default is \code{FALSE}.}
}
\value{
No return value.
}
\description{
This function downloads a spatial vector of world countries boundaries as
defined by the IPBES (available at \url{https://zenodo.org/record/3928281}).
A \code{zip} file is downloaded in the folder \code{path} (working directory by
default) and files are extracted in the same folder. If this folder doesn't
exist it will be created.

\strong{Note:} if you delete the zip\verb{ file (after files extraction) and re-run  this function, the zip} file will be downloaded again. It this is your wish
do not forget to use \code{force = TRUE} to erase previous files.

Original Projection System: WGS84
}
\examples{
\dontrun{
## Download and extract zip file ----
get_world_basemap()

## Check ----
list.files()

## Re-run function ----
get_world_basemap()              # No download (zip file already present)
get_world_basemap(force = TRUE)  # Erase previous files
}
}
