% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{get_zeisel_brain}
\alias{get_zeisel_brain}
\title{Download and preprocess Zeisel brain data.}
\usage{
get_zeisel_brain(subset_by_MHVG = FALSE)
}
\description{
Download and preprocess Zeisel brain data.
}
