% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{getpoly}
\alias{getpoly}
\title{Download an \emph{OHLC} time series of prices from Polygon.}
\usage{
getpoly(
  symbol = "SPY",
  startd = as.Date("1997-01-01"),
  endd = Sys.Date(),
  tspan = "day",
  apikey
)
}
\arguments{
\item{\code{symbol}}{The stock symbol (ticker).}

\item{\code{startd}}{The start date (default is "1997-01-01").}

\item{\code{endd}}{The end date (default is \code{Sys.Date()}).}

\item{\code{tspan}}{The data frequency, i.e. the time span for data
aggregations (default is "day" for daily data).}

\item{\code{apikey}}{The API key issued by Polygon.}
}
\value{
An \emph{OHLC} time series of class \emph{xts}.
}
\description{
Download an \emph{OHLC} time series of prices from Polygon.
}
\details{
The function \code{getpoly()} downloads historical prices from
  \href{https://polygon.io}{Polygon}, and returns an \emph{OHLC} time series
  of class \emph{xts}.

  \href{https://polygon.io}{Polygon} is a provider of live and historical
  prices for stocks, options, foreign currencies, and cryptocurrencies.

  The function \code{getpoly()} sends a request for data to the
  \href{https://polygon.io}{Polygon} rest API, using the function
  \code{read_json()} from package
  \href{https://cran.r-project.org/web/packages/jsonlite/index.html}{jsonlite}.
  The query requires an API key issued by \href{https://polygon.io}{Polygon}.
  The API key must be passed to the argument \code{apikey}.

  \href{https://polygon.io}{Polygon} returns data in \emph{JSON} format,
  which is then formatted into an \emph{OHLC} time series of class
  \emph{xts}.

  The argument \code{tspan} determines the data frequency, i.e. it's the time
  span for data aggregations.  The default is "day" for daily data.  Other
  possible values of \code{tspan} are "minute", "hour", "week", and "month".
}
\examples{
\dontrun{
# Polygon API key - user must obtain their own key
apikey <- "0Q2f7j8CwAbdY5M8VYt_8pwdP0V4TunxbvRVC_"
# Download SPY prices from Polygon
ohlc <- rutils::getpoly(symbol="SPY", apikey=apikey)
# Plot candlesticks of SPY OHLC prices
library(highcharter)
highcharter::highchart(type="stock") \%>\% hc_add_series(ohlc, type="candlestick")
}
}
