% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{getSCEWithSymbols}
\alias{getSCEWithSymbols}
\title{change gene names to gene symbol}
\usage{
getSCEWithSymbols(sce, keytype = "ENTREZID", org.db = org.Mm.eg.db)
}
\description{
change gene names to gene symbol
}
