% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Ggplot default colors}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{number of colors}
}
\description{
Emulate ggplot's default colors.
}
\details{
https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette
}
