% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitSessionInfo.R
\name{GitSessionInfo}
\alias{GitSessionInfo}
\title{GitSessionInfo}
\usage{
GitSessionInfo()
}
\value{
Character vector for printing to a logfile, e.g. with
  \code{writeLines}
}
\description{
Get session info with git branch and has and system date.
}

