% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_plot.R
\name{go_plot}
\alias{go_plot}
\title{Visualise reduced pathway enrichments}
\usage{
go_plot(
  reduced_pathway_df,
  col_gene_list = "gene_list",
  col_fdr = "fdr",
  fct_gene_list = NULL
)
}
\arguments{
\item{reduced_pathway_df}{the \link[tibble:tbl_df-class]{tibble} derived from
\code{\link{go_reduce}}.}

\item{col_gene_list}{a \code{character()} vector, with the name of column
containing the name(s) of the inputted gene list(s). Defaults to
\code{"gene_list"}.}

\item{col_fdr}{a \code{character()} vector, with the name of column containing
enrichment p-values/fdr-values. Defaults to \code{"fdr"}.}

\item{fct_gene_list}{a \code{character()} vector, with the order in which input
gene lists should be factored. Defaults to NULL.}
}
\value{
\code{ggplot} displaying the pathway enrichments using reduced perent
terms. \itemize{
\item x-axis displays the inputted gene list(s).
\item
y-axis displays the reduced parent term, with the number of associated
child GO terms in brackets. Parent terms are ordered first by the factoring
of the input gene list (as provided in the argument, \code{fct_gene_list}) and
then by the number of associated child terms.
\item size of the dot indicates the proportion of child terms associated
with an inpute gene list that are annotated to the parent term.
\item fill of dot indicates the minimum FDR
assigned to a child term that is associated with the parent term (within an
input gene list).
}
}
\description{
\code{go_plot} plots GO terms that have been reduced using \code{\link{go_reduce}}.
}
\examples{
file_path <- 
   system.file(
     "testdata", 
     "go_test_data.txt", 
     package = "rutils", 
     mustWork = TRUE
   )
   
pathway_df <- 
   readr::read_delim(file_path,
                     delim = "\t") \%>\% 
   dplyr::mutate(gene_list = c(rep("a", 5), rep("b", 5)),
                 fdr = runif(10, min=0, max=0.05))
 
reduced_pathway_df <-  
  go_reduce(
   pathway_df = pathway_df,
   threshold = 0.9,
   scores = NULL,
   measure = "Wang")

go_plot(
  reduced_pathway_df,
  col_gene_list = "gene_list",
  col_fdr = "fdr",
  fct_gene_list = c("b", "a"))
}
\seealso{
\code{\link{go_reduce}} for reducing GO terms

Other GO-related functions: 
\code{\link{go_reduce}()}
}
\concept{GO-related functions}
