% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_reduce.R
\name{go_reduce}
\alias{go_reduce}
\title{Reduce redundancy of human GO terms}
\usage{
go_reduce(pathway_df, threshold = 0.7, scores = NULL, measure = "Wang")
}
\arguments{
\item{pathway_df}{a \code{data.frame} or \link[tibble:tbl_df-class]{tibble} object,
with the following columns:
\itemize{
\item \code{go_type}: the sub-ontology the
GO term relates to. Should be one of \code{c("BP", "CC", "MF")}.
\item \code{go_id}: the gene ontology identifier (e.g. GO:0016209)
}}

\item{threshold}{\code{numeric()} vector. Similarity threshold (0-1) for
\code{rrvgo::\link[rrvgo:reduceSimMatrix]{reduceSimMatrix()}}. Default
option is 0.7. Some guidance:
\itemize{
\item For large term groupings, use \code{threshold = 0.9}
\item For medium term groupings, use \code{threshold = 0.7}
\item For small term groupings, use \code{threshold = 0.5}
\item For tiny term groupings, use \code{threshold = 0.4}
}}

\item{scores}{\strong{named} vector, with scores (weights) assigned to each
term. Higher is better. Can be NULL (default, means no scores. In this case,
a default score based on set size is assigned, thus favoring larger sets).
Note: if you have p-values as scores, consider log-transforming them
(\code{-log10(p)}).}

\item{measure}{\code{character()} vector, indicating method to be used to
calculate semantic similarity measure. Must be one of the methods supported
by GOSemSim: c("Resnik", "Lin", "Rel", "Jiang", "Wang"). Default is "Wang".}
}
\value{
a \link[tibble:tbl_df-class]{tibble} object of pathway results, a
"reduced" parent term to which pathways have been assigned. New columns:
\itemize{
\item \code{parent_id}: the GO ID of the parent term
\item \code{parent_term}: a description of the GO ID
\item \code{parent_sim_score}: the similarity score between the child GO term and
its parent term
}
}
\description{
This function will reduce GO redundancy first by creating a
semantic similarity matrix (using
\code{GOSemSim::\link[GOSemSim:mgoSim]{mgoSim}}), which is then passed
through \code{rrvgo::\link[rrvgo:reduceSimMatrix]{reduceSimMatrix()}},
which will reduce a set of GO terms based on their semantic similarity and
scores (in this case, a default score based on set size is assigned.)
}
\details{
Semantic similarity is calculated using the "Wang" method, a
graph-based strategy to compute semantic similarity using the topology of
the GO graph structure. \code{GOSemSim::\link[GOSemSim:mgoSim]{mgoSim}}
does permit use of other measures (primarily information-content measures),
but "Wang" is used as the default in GOSemSim (and was, thus, used as the
default here). If you wish to use a different measure, please refer to the
\href{https://yulab-smu.top/biomedical-knowledge-mining-book/semantic-similarity-overview.html}{GOSemSim documentation}.

\code{rrvgo::\link[rrvgo:reduceSimMatrix]{reduceSimMatrix()}} creates a
distance matrix, defined as (1-simMatrix). The terms are then hierarchically
clustered using complete linkage (an agglomerative, or "bottom-up"
clustering approach), and the tree is cut at the desired threshold. The term
with the highest "score" is used to represent each group.
}
\examples{
file_path <-
    system.file(
        "testdata",
        "go_test_data.txt",
        package = "rutils",
        mustWork = TRUE
    )

pathway_df <-
    readr::read_delim(file_path,
        delim = "\t"
    )

go_reduce(
    pathway_df = pathway_df,
    threshold = 0.9,
    scores = NULL,
    measure = "Wang"
)
}
\references{
\itemize{
\item Yu et al. (2010) GOSemSim: an R package for measuring semantic
similarity among GO terms and gene products \emph{Bioinformatics} (Oxford,
England), 26:7 976--978, April 2010.
\url{http://bioinformatics.oxfordjournals.org/cgi/content/abstract/26/7/976}
PMID: 20179076
\item Yu (2021) Biomedical Knowledge Mining using GOSemSim and
clusterProfiler
\url{https://yulab-smu.top/biomedical-knowledge-mining-book/index.html}
\item Sayols S (2020). rrvgo: a Bioconductor package to
reduce and visualize Gene Ontology terms. \url{https://ssayols.github.io/rrvgo}
}
}
\seealso{
\code{\link{go_plot}} for plotting the output of \code{go_reduce},
\code{GOSemSim::\link[GOSemSim:mgoSim]{mgoSim}} for calculation of semantic
similarity and
\code{rrvgo::\link[rrvgo:reduceSimMatrix]{reduceSimMatrix()}} for reduction
of similarity matrix

Other GO-related functions: 
\code{\link{go_plot}()}
}
\concept{GO-related functions}
