% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_fun_par.R
\name{grab_fun_par}
\alias{grab_fun_par}
\title{Grab all the parameters in a function}
\usage{
grab_fun_par()
}
\value{
A \code{\link[base:list]{list}} containing the parameters of the function
from which it was called.
}
\description{
\code{grab_fun_par()} grabs all parameters defined in a function. It's
particularly useful when you need to pass all parameters from one function
to another (e.g., using \code{\link[base:do.call]{do.call()}}).

Credits: Function adapted from B. Christian Kamgang's contribution in a
\href{https://bit.ly/3QD48Wy}{Stack Overflow discussion}.
}
\examples{
foo <- function(a = 1) grab_fun_par()

foo()
#> $a # Expected
#> [1] 1
}
\concept{scoping functions.}
