% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicals.R
\name{\%ni\%}
\alias{\%ni\%}
\title{NOT IN function}
\usage{
a \%ni\% b
}
\arguments{
\item{a}{contained vector}

\item{b}{containing vector}
}
\value{
logical vector for each value in a, is there a match in b
}
\description{
This function is a shortcut for `!(a %in% b)`
}
\examples{
"a" \%ni\% letters[1:10]
}
\keyword{in}
