% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{group_proportions}
\alias{group_proportions}
\title{Calculate proportions within groups}
\usage{
group_proportions(df, ...)
}
\arguments{
\item{df}{dataframe containing the grouping variables}

\item{...}{any number of grouping variables passed as names}
}
\value{
dataframe with counts and proportions of groups within groups
}
\description{
Given a dataframe and one or more grouping variables this function will calculate the
proportion of group observations within groups
}
\examples{
group_proportions(mtcars, hp, cyl)
}
