% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutils-package.R
\docType{package}
\name{gutils-package}
\alias{gutils}
\alias{gutils-package}
\title{gutils: Utility Functions Created by the GIPERBIO Developer Team}
\description{
An R package made to store and organize utility functions created by the GIPERBIO developer team. We don’t intend to publish this package to CRAN.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://giperbio.github.io/gutils/}
  \item \url{https://github.com/giperbio/gutils/}
  \item Report bugs at \url{https://github.com/giperbio/gutils/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
