% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/has.only.NaNs.R
\name{has.only.NaNs}
\alias{has.only.NaNs}
\title{NaN Check}
\usage{
has.only.NaNs(vec)
}
\arguments{
\item{vec}{a vector}
}
\description{
Test if a vector has just NaN elements.
}
\examples{
has.only.NaNs(c(NaN, NaN, NaN)) ## TRUE
has.only.NaNs(c(1, NaN, NaN))   ## FALSE
}

