% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiking_time.R
\name{hiking_time}
\alias{hiking_time}
\title{Hiking time calculation}
\usage{
hiking_time(distance, elevation_gain, elevation_loss)
}
\arguments{
\item{distance}{a \code{numeric} of length 1. The total distance of the hike
(in kilometers). Must be strictly positive.}

\item{elevation_gain}{a \code{numeric} of length 1. The total elevation gain
(cumulative positive elevation). Must be positive or zero.}

\item{elevation_loss}{a \code{numeric} of length 1. The total elevation loss
(cumulative negative elevation). Must be positive or zero.}
}
\value{
A \code{character} of length 1 returning the hiking time as \code{"23h59"}.
}
\description{
Estimates the hiking time for a planned route by considering horizontal
distance, elevation gain and elevation loss.

This function implements the DIN 33466 of the German Alpine Club. Different
speeds are assumed according to this standard:
\itemize{
\item 4 kilometers per hour in the horizontal direction
\item 300 vertical meters in the ascent per hour
\item 500 vertical meters in the descent per hour
}
}
\examples{
## Lac du Montagnon par le Col d’Iseye 
## https://www.visorando.com/randonnee-lac-du-montagnon-par-le-col-d-iseye/

hiking_time(distance = 20, elevation_gain = 1460, elevation_loss = 1460)
}
