% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{ifelse}
\alias{ifelse}
\title{ifelse come dio comanda}
\usage{
ifelse(test, a, b)
}
\arguments{
\item{test}{un predicato che deve restituire `TRUE/FALSE`}

\item{a}{l'oggetto ritornato se `test` e' `TRUE`}

\item{b}{l'oggetto ritornato se `test` e' `FALSE`}
}
\description{
Questo `ifelse`, a differenze di `base::ifelse` non modifica
lo shape dell'oggetto ritornato. (il che, oggettivamente
l'e' una bischerata)
}
