% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbetween_integers.R
\name{inbetween_integers}
\alias{inbetween_integers}
\title{Get all the integers between two integers}
\usage{
inbetween_integers(x, y)
}
\arguments{
\item{x}{An integer number.}

\item{y}{An integer number.}
}
\value{
An \code{\link[base:integer]{integer}} vector with all the integers between
\code{x} and \code{y}.
}
\description{
\code{inbetween_integers()} returns all the integers between two integers.
This is useful while subsetting a vector.
}
\examples{
inbetween_integers(1, 5)
#> [1] 2 3 4 # Expected

inbetween_integers(5, 1)
#> [1] 2 3 4 # Expected
}
\seealso{
Other vector functions.: 
\code{\link{count_na}()},
\code{\link{drop_na}()}
}
\concept{vector functions.}
