% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ini_parse.r
\name{ini_parse}
\alias{ini_parse}
\title{Parses a file INI}
\usage{
ini_parse(filename)
}
\arguments{
\item{filename}{filename to parse}
}
\value{
a list containing the INI files' informations
}
\description{
Parses a file INI
}
\note{
full credit to \url{http://goo.gl/V0jwlz}
}
