% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{init_logging}
\alias{init_logging}
\title{Init logging based on configuration file}
\usage{
init_logging(config_filename)
}
\arguments{
\item{config_filename}{path to the configuration file}
}
\description{
This tries to mimic the common pattern to use
a single config file to configure loggers and appenders
}
\details{
This is way simpler than any other logging library out there.
}
