% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{init_project}
\alias{init_project}
\title{Initialize Project}
\usage{
init_project()
}
\value{
Directories created in the current working directory.
}
\description{
This function initializes a project by creating three directories named "0_data", "1_code", and "2_output".
It checks if the directories exist, and if not, creates them.
}
