% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeHome.R
\name{inputFile}
\alias{inputFile}
\title{provide wrapper to path input files}
\usage{
inputFile(fileName)
}
\arguments{
\item{fileName}{of expected file in path}
}
\description{
provide wrapper to path input files
}
