% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_old_pkgs_for_new_r.R
\name{install_old_pkgs_for_new_r}
\alias{install_old_pkgs_for_new_r}
\title{Install R packages for new major R version}
\usage{
install_old_pkgs_for_new_r(
  dir_base = "C:/Users/migue/OneDrive/Documents/R/win-library",
  type = "win.binary"
)
}
\arguments{
\item{dir_old}{character. Absolute path to old R version's library (i.e. folder
where packages were installed).}
}
\value{
A character vector of all packages that were not successfully installed.
}
\description{
Install all non-based packages installed for previous version of R
for new version of R.
}
\details{
Installs \code{caret} package with
}
