% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/is.empty.vector.R
\name{is.empty.vector}
\alias{is.empty.vector}
\title{Vector}
\usage{
is.empty.vector(vec)
}
\arguments{
\item{vec}{a vector.}
}
\value{
A boolean vector with just one element indicating if the vector \code{vec} is empty.
}
\description{
This function indicates whether or not a vector is empty.
}
\examples{
is.vector.empty(1:10) ## FALSE
is.vector.empty(numeric()) ## TRUE
}

