% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utils.r
\name{is.jenkins}
\alias{is.jenkins}
\title{Internal Functions}
\usage{
is.jenkins()
}
\value{
`TRUE` if running under Jenkins, `FALSE` otherwise
}
\description{
internal function to check if the runtime environment is jenkins.
}
\details{
more like an hack.
}

