% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.nan.data.frame.R
\name{is.nan.data.frame}
\alias{is.nan.data.frame}
\title{NaN}
\usage{
\method{is.nan}{data.frame}(d)
}
\arguments{
\item{data}{a data.frame}
}
\description{
function to extend data.frame class, adding NaN checking.
}
\examples{

d <- data.frame(c1 = rep(NaN, 10), c2 = rep(NaN, 10))
d[is.nan(d)] <- 0

}

