% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.r
\name{is.Stack}
\alias{is.Stack}
\title{Returns `TRUE` if `x` is a Stack}
\usage{
is.Stack(x)
}
\arguments{
\item{x}{object to check}
}
\value{
`TRUE` if x is a Stack; `FALSE` otherwise
}
\description{
Returns `TRUE` if `x` is a Stack
}
