% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/is.vector.empty.R
\name{is.vector.empty}
\alias{is.vector.empty}
\title{Vector}
\usage{
is.vector.empty(vec)
}
\arguments{
\item{vec}{a vector.}
}
\value{
A boolean vector with just one element indicating if the vector \code{vec} is empty.
}
\description{
This function indicates whether or not a vector is empty.
}
\examples{
is.vector.empty(1:10)
is.vector.empty(numeric())
}

