% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_r_project.r
\name{is_r_project}
\alias{is_r_project}
\title{Checks if \code{path} contains a valid R project structure}
\usage{
is_r_project(path = getwd())
}
\arguments{
\item{path}{path to check for an R project (defaults to \code{getwd()})}
}
\description{
Checks if \code{path} contains:
\itemize{
\item a DESCRIPTION file
\item a NAMESPACE file
\item an R directory
}
}
