% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_test_environment}
\alias{is_test_environment}
\title{Test environment}
\usage{
is_test_environment()
}
\value{
\code{TRUE} if current environment is for testing, \code{FALSE} otherwise.
}
\description{
Checks if current environment is aimed at testing.
Specifically, it tests whether package \code{testthat} was loaded.
}
\examples{
is_test_environment()
\dontrun{
library(testthat)
is_test_environment()
}
}
\keyword{check}
\keyword{test}
