% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{lag.xts}
\alias{lag.xts}
\title{Apply a time lag to an \code{xts} time series.}
\usage{
\method{lag}{xts}(x_ts, k = 1)
}
\arguments{
\item{x_ts}{an \code{xts} time series.}

\item{k}{integer equal to the number of time periods of lag.}
}
\value{
\code{xts} time series with the same dimensions and the same time
  index as the input series.
}
\description{
Apply a time lag to an \code{xts} time series.
}
\details{
Applies a time lag to an \code{xts} time series and pads with
  \code{zeros} instead of \code{NAs}.  Positive lag \code{k} means values
  from \code{k} periods in the past are moved to the present.  Negative lag
  \code{k} moves values from the future to the present.  The function
  \code{lag()} is just a wrapper for function \code{lag.xts()} from package
  \href{https://cran.r-project.org/web/packages/xts/index.html}{xts},
  but it pads with \code{zeros} instead of \code{NAs}.
}
\examples{
# lag by 10 periods
rutils::lag.xts(env_etf$VTI, k=10)
}

