% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{lag_it}
\alias{lag_it}
\title{Apply a lag to a \emph{numeric} vector or matrix.}
\usage{
lag_it(in_put, lagg = 1)
}
\arguments{
\item{in_put}{A \emph{numeric} vector or matrix.}

\item{lagg}{An integer equal to the number of time periods of lag (default is
1).}
}
\value{
A vector or matrix with the same dimensions as the input object.
}
\description{
Apply a lag to a \emph{numeric} vector or matrix.
}
\details{
Applies a lag to a vector or matrix, by shifting its values by a
  certain number of rows, equal to the integer \code{lagg}, and pads the
  leading or trailing stub periods with \emph{zeros}. Positive \code{lagg}
  means that values in the current row are replaced with values from the row
  that are \code{lagg} rows above. (vice versa negative \code{lagg}).  This
  also applies to vectors, since they can be viewed as single-column
  matrices.
}
\examples{
# lag vector by 2 periods
rutils::lag_it(1:10, lag=2)
# lag matrix by negative 2 periods
rutils::lag_it(matrix(1:10, ncol=2), lag=-2)
}

