% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last}
\alias{last}
\title{Element Retrieval}
\usage{
last(x, is_list = FALSE)
}
\arguments{
\item{x}{a vector or list}

\item{is_list}{is x a list?}
}
\description{
Retrieve the last element of a vector or a list.
}
\examples{

last(1:100)
## 100

}

