% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leading_zero.R
\name{leading_zero}
\alias{leading_zero}
\title{Add leading zeros to a vector}
\usage{
leading_zero(x)
}
\arguments{
\item{x}{a \code{character} or a \code{numeric} vector.}
}
\value{
A \code{character} vector of the same length as \code{x}.
}
\description{
Adds n leading zeros to a vector so that each element of the vector has the
same number of characters. See examples below.
}
\examples{
## On numeric vector ----
x <- 1:10
leading_zero(x)

x <- 50:100
leading_zero(x)

## On floating vector ----
x <- c(1, 9.5, 10, 10.2)
leading_zero(x)

## On character vector ----
x <- as.character(1:10)
leading_zero(x)

## Creating sortable identifiers ----
x1 <- as.character(10:1)
x2 <- leading_zero(x)

sort(x1)
sort(x2)

x2 <- paste0("ID", x2)
sort(x2)
}
