% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/len.R
\name{len}
\alias{len}
\title{A shortcut for the function length}
\usage{
len(x)
}
\arguments{
\item{x}{an R object.}
}
\value{
See ?length for a complete description.
}
\description{
A shortcut for the function length
}
\examples{
length(letters)

len(letters)
}
