% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{lft}
\alias{lft}
\title{List Files in Table}
\usage{
lft(.dirs, .reg = NULL, .id = "doc_id", .rec = FALSE)
}
\arguments{
\item{.dirs}{Vector/List or single paths to directory/directories}

\item{.reg}{RegEx to find files (defaults to '*' all files)}

\item{.id}{Column name containing the file name (defaults to 'doc_id')}

\item{.rec}{Should the directories be searched recursively?}
}
\value{
A dataframe with file paths
}
\description{
List Files in Table
}
\examples{
library(rUtils)

dir <- system.file("extdata", package = "rUtils")

lft(dir)

rm(dir)
}
