% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{library_install}
\alias{library_install}
\title{Install and load an R package if not already installed}
\usage{
library_install(library_name)
}
\arguments{
\item{library_name}{The name of the library to install.}
}
\description{
Install and load an R package if not already installed
}
