% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover_coord.R
\name{liftover_coord}
\alias{liftover_coord}
\title{Liftover genomic coordinates from one genome build to another.}
\usage{
liftover_coord(df, path_to_chain)
}
\arguments{
\item{df}{Dataframe with genomic coordinates. Must include columns, \code{CHR} and
\code{BP}, with chromosome and base pair, respectively. If additional
columns included these will be preserved.}

\item{path_to_chain}{Path to UCSC chain file for transformation. This can
either be from hg19 to hg38 or hg38 to hg19.}
}
\value{
Dataframe with coordinates in different genome build.
}
\description{
Liftover genomic coordinates from one genome build to another.
}
