% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inc.utils.r
\name{linear.map}
\alias{linear.map}
\title{linear mapping from [a,b] to [0,c]}
\usage{
linear.map(x, new.up, plotit = FALSE)
}
\arguments{
\item{x}{ordered array. a grid.}

\item{new.up}{new upper bound}

\item{plotit}{boolean true if plot required}
}
\description{
linear mapping from [a,b] to [0,c]
}

