% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{lines2}
\alias{lines2}
\title{Extended lines chart function}
\usage{
lines2(
  y,
  x = seq(ncol(y)),
  ynames = "",
  main = "",
  append = FALSE,
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{y}{data.frame or numeric matrix, each line of which is a line to plot}

\item{x}{x values for each line to plot}

\item{ynames}{how to name each line in legend}

\item{main}{plot title}

\item{append}{TRUE to append those lines to an existing plot,
FALSE to create a new one}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
a plot object that represents a line scatter plot
}
\description{
This function is a shortcut to plot several lines in one call.
}
\examples{
data(iris)
lines2(iris[seq(4), seq(4)], ynames=as.character(seq(4)))
}
\keyword{chart}
\keyword{line}
\keyword{lines}
\keyword{plot}
