% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{list_all_signature}
\alias{list_all_signature}
\title{List all function signatures in a package}
\usage{
list_all_signature(pkg)
}
\arguments{
\item{pkg}{package name in a character}
}
\value{
list
}
\description{
List all function signatures in a package
}
