% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_as_tibble.R
\name{list_as_tibble}
\alias{list_as_tibble}
\title{Convert a \code{list} to a \code{tibble} in long format}
\usage{
list_as_tibble(list)
}
\arguments{
\item{list}{A \code{\link[base:list]{list}} to be converted.}
}
\value{
A \code{\link[tibble:tibble]{tibble}} with two columns: \code{name} and \code{value},
where \code{name} corresponds to the original \code{\link[base:list]{list}} element names
and \code{value} to their respective values as \code{\link[base:character]{character}}`
strings.
}
\description{
\code{list_as_tibble()} converts a \code{\link[base:list]{list}} to a
\code{\link[tibble:tibble]{tibble}}, pivoting it to long format.
}
\examples{
list(a = 1, b = "a", c = TRUE) |>
  list_as_tibble()
}
\seealso{
Other parsing/conversion functions.: 
\code{\link{vector_to_c}()}
}
\concept{parsing/conversion functions.}
