% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_edit.R
\name{list_edit}
\alias{list_edit}
\title{Edit a list}
\usage{
list_edit(.l, ...)
}
\arguments{
\item{.l}{List to edit}

\item{...}{Named elements to add or delete}
}
\description{
Edit the elements of a list by name
}
\details{
\code{list_edit} is similar to \link[purrr:list_modify]{purrr::list_modify} in
\code{purrr}, but unlike \code{list_modify}, does not operate recursively
on a list. It only edits elements at the top level of hierarchy.
Elements can be removed by setting them equal to \code{NULL}.
}
\examples{
list_edit(list(x = 1, y = 2), x = 10, z = 20)
list_edit(list(x = 1, y = 2), x = NULL, y = 5)
list_edit(list(x = 1, y = 2), !!!list(x = NULL, z = 100))
}
