% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_linked_files.R
\name{list_linked_files}
\alias{list_linked_files}
\title{List all files linked to a reference in a Zotero library}
\usage{
list_linked_files(
  lib_file = tcltk::tk_choose.files(caption = "Select Zotero library CSV file", multi =
    FALSE),
  basename = TRUE
)
}
\arguments{
\item{lib_file}{(optional) A string specifying the path to the Zotero library
exported as a \strong{CSV} file.
(default: \link[tcltk:tk_choose.files]{`tk_choose.files()``}).}

\item{basename}{(optional) A \code{\link[base:logical]{logical}} flag indicating
if the function should return the full path to the files or only the file
names (default: \code{TRUE}).}
}
\value{
A \code{\link[base:character]{character}} vector with the names of the
files linked to the references in the Zotero library.
}
\description{
\code{list_linked_files()} reads a CSV file exported from Zotero and extracts the
file paths of all attachments linked to the references in the library.
}
\details{
To export your library from Zotero, go to the menu \verb{File > Export Library...}
and choose the \strong{CSV} format.
}
\examples{
\dontrun{
  list_linked_files()
}
}
\seealso{
Other Zotero functions: 
\code{\link{find_orphan_files}()}
}
\concept{Zotero functions}
