% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance.R
\name{llrs_add_member}
\alias{llrs_add_member}
\title{Add a contributor to the registry}
\usage{
llrs_add_member(
  name = getOption("usethis.full_name"),
  role = "Collaborator",
  path = "."
)
}
\arguments{
\item{name}{Character name of the contributor.}

\item{role}{Either Collaborator or Publisher.}

\item{path}{Path to the project.}
}
\value{
\code{NULL}
}
\description{
Add a contributor to the registry
}
\examples{
\dontrun{
llrs_governance()
llrs_add_member(NULL)
}
}
