% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{llrs_box}
\alias{llrs_box}
\title{Create positions of a box}
\usage{
llrs_box(size)
}
\arguments{
\item{size}{A numeric value with the number of rows or columns.}
}
\value{
A data.frame with the rows and columns.
}
\description{
Create a data.frame with the positions of the sample in a square box
}
\examples{
box10 <- llrs_box(10)
box9 <- llrs_box(9)
}
