% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{llrs_cellranger_aggr}
\alias{llrs_cellranger_aggr}
\title{Aggregate several cellrangers}
\usage{
llrs_cellranger_aggr()
}
\value{
The path to the file saved with the configuration
}
\description{
Aggregate several cellrangers
}
