% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{llrs_cellranger_clean}
\alias{llrs_cellranger_clean}
\title{Clean cellranger multi output}
\usage{
llrs_cellranger_clean(path)
}
\arguments{
\item{path}{Path to the output of cellranger.}
}
\value{
Called by its side effect (It deletes some files and folders).
Returns \code{TRUE} if removed files and directories, \code{NULL} if cancelled, \code{FALSE}
if nothing to remove.
}
\description{
This might squeeze 4Gb of data for each folder...
}
