% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{llrs_cellranger_multi}
\alias{llrs_cellranger_multi}
\title{Create multi config file}
\usage{
llrs_cellranger_multi(config, gex, library, samples, vdj, feature, out)
}
\arguments{
\item{config}{A list with the general config of the samples.}

\item{gex}{A data.frame with the information for the gene expression (GEX).}

\item{library}{A data.frame with the information about the libraries used.}

\item{samples}{A data.frame with the information about the samples used.}

\item{vdj}{A list with the configuration for the vdj assays.}

\item{feature}{A list with the configuration for the featured assays.}

\item{out}{Path to a directory where to save the file.}
}
\value{
The path to a file saved with the configuration as specified.
}
\description{
Prepare the cellranger multi file for some samples.
}
\examples{
#llrs_cellranger_multi(out = "/S2_cellranger.csv")
}
\references{
\url{https://www.10xgenomics.com/support/software/cell-ranger/latest/advanced/cr-multi-config-csv-opts}
}
\seealso{
Set up names of samples with \code{\link[=llrs_cnag_files]{llrs_cnag_files()}}. Prepare folders and files \code{\link[=llrs_cnag_symlinks]{llrs_cnag_symlinks()}}.
}
