% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance.R
\name{llrs_change_role}
\alias{llrs_change_role}
\title{Remove role of a contributor}
\usage{
llrs_change_role(name = NULL, role = "User", path = ".")
}
\arguments{
\item{name}{Character name of the contributor.}

\item{role}{Either User, Collaborator or Publisher.}

\item{path}{Path to the project.}
}
\value{
\code{NULL}
}
\description{
Change a contributor role on the project.
}
\examples{
\dontrun{
llrs_governance()
llrs_add_member(NULL)
llrs_change_member(NULL, "Publisher")
}
}
