% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnag.R
\name{llrs_cnag_files}
\alias{llrs_cnag_files}
\title{Prepare CNAG data for cellranger}
\usage{
llrs_cnag_files(path)
}
\arguments{
\item{path}{Path to a file with information about the delivery file.}
}
\value{
A data.frame with the output of the names of the files ("d1" and "d2")
and its associated information.
}
\description{
CNAG returns the data in a format that is not compatible with cellranger.
This function merges the data of CNAG to be able to prepare for it.
}
\details{
The CNAG project is assumed to be below the file in \code{path} with all the fastq files.
}
\examples{
# out <- llrs_cnag_cellranger("PROJECT_01.xls")
# out[, c("SAMPLE NAME", "d1", "cr1", "d1", "cr2")]
}
