% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnag.R
\name{llrs_cnag_samples}
\alias{llrs_cnag_samples}
\title{Create a sample sheet from files}
\usage{
llrs_cnag_samples(path)
}
\arguments{
\item{path}{A path to a folder containing the files.}
}
\value{
A data.frame with the names of the files.
}
\description{
Create a sample sheet from fastq.gz files for easier creation of sample
sheets for several programs (multiqc, fastqc, ...).
}
\details{
Assumes a paired end sequencing.
}
\examples{
# llrs_cnag_samples("AUSER_01.xls")
}
\references{
\href{https://www.cnag.eu/}{CNAG}
}
