% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnag.R
\name{llrs_cnag_stats}
\alias{llrs_cnag_stats}
\title{Read the CNAG stats}
\usage{
llrs_cnag_stats(path)
}
\arguments{
\item{path}{Path to excel file.}
}
\value{
A data.frame with the data.
}
\description{
Read the CNAG stats
}
\examples{
# llrs_cnag_stats("AUSER_01_Sample_Stats.xls")
}
\references{
\href{https://www.cnag.eu/}{CNAG}
}
