% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnag.R
\name{llrs_cnag_symlinks}
\alias{llrs_cnag_symlinks}
\title{Create symlinks}
\usage{
llrs_cnag_symlinks(x, name, out_dir)
}
\arguments{
\item{x}{A data.frame with the information.
At least \code{c("LANE", "MULTIPLEX INDEX", "FLOWCELL", "d1", "d2")} columns must
be present.}

\item{name}{The name of the sample, it can be a single character with the
name of the column or a vector of characters with the sample name for each file.}

\item{out_dir}{Path to a folder where the files should be placed.
If \code{NULL} it doesn't create the symlinks to the folder.}
}
\value{
Called by its side effect but also returns a data.frame.
}
\description{
Create symlinks for cellranger format (see references).
}
\references{
\url{https://www.10xgenomics.com/support/software/cell-ranger/latest/analysis/inputs/cr-specifying-fastqs#file-naming-convention}
}
