% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{llrs_copy_config}
\alias{llrs_copy_config}
\title{Copy config to config repository}
\usage{
llrs_copy_config(path, dest_dir = "~/Documents/Projects/customizations")
}
\arguments{
\item{path}{Path of file to copy}

\item{dest_dir}{Path to folder were to save it.}
}
\value{
TRUE if successful
}
\description{
Copy file to config repository and add it and push it.
}
\details{
If there are files with the new file and the existing one it will prompt to review them.
}
\examples{
\dontrun{
llrs_copy_config("~/.gitconfig")
}
}
