% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics.R
\name{llrs_diff}
\alias{llrs_diff}
\title{Add column with}
\usage{
llrs_diff(
  x,
  p = "pval",
  fc = "diff",
  fdr = "fdr",
  fdr_threshold = 0.05,
  p_threshold = 0.05
)
}
\arguments{
\item{x}{A data.frame or matrix with differential results (DEG or other bioinformatics data)}

\item{p}{The name of the column with the p-values.}

\item{fc}{The name of the column with the fold change.}

\item{fdr}{The name of the column with the p-values corrected by multiple comparison (usually FDR).}

\item{fdr_threshold}{The value to decide when to classify them.}

\item{p_threshold}{The value to decide when to consider a row}
}
\value{
The same matrix or data.frame with a new column "sign" with 5 possible values:
\code{c("UUP", "UP", "", "DW", "DDW")}.
This new column allows to filter and easily find relevant rows.
}
\description{
Adds a new column with those rows that meet the preference
}
\note{
limma recommends using the B value instead of this \emph{ad-hoc} test (but it is popular among labs).
}
