% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onedrive.R
\name{llrs_download}
\alias{llrs_download}
\title{Download a file}
\usage{
llrs_download(file, dest = ".")
}
\arguments{
\item{file}{Path to file from the One Drive.}

\item{dest}{Path to where the file should be stored.}
}
\value{
The path
}
\description{
Download a file
}
