% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{llrs_figures}
\alias{llrs_figures}
\title{Creates path for a figure}
\usage{
llrs_figures(project, ...)
}
\arguments{
\item{project}{Name of the project to be used in the title.}

\item{...}{Other names/variables. The extensions of the file should be
included on the last element.}
}
\value{
A path to a file in a folder
}
\description{
creates a standardized name for figures and checks if the standardized
folders exist.
}
\examples{
# llrs_figures("rutils", "pca", "sample1.png")
# llrs_figures("rutils", "pca", "sample whatever.png")
}
