% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_hooks.R
\name{llrs_hook}
\alias{llrs_hook}
\title{Insert hook}
\usage{
llrs_hook(file, path = getwd())
}
\arguments{
\item{file}{Name of the file to be used as a hook}

\item{path}{Path to the git repository}
}
\value{
A logical value if the hook can be placed there.
}
\description{
Insert a new hook (It doesn't replace hooks or expand them).
}
\examples{
\dontrun{
hook <- system.file("post-receive", package = "rutils")
llrs_hook(hook, "projects/git_something")
}
}
\references{
<https://git-scm.com/docs/githooks>
}
\seealso{
[llrs_hooks_list()]
}
