% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{llrs_results}
\alias{llrs_results}
\title{Creates path for results}
\usage{
llrs_results(project, ...)
}
\arguments{
\item{project}{Name of the project to be used in the title}

\item{...}{Other names/variables. The extensions of the file should be
included on the last element.}
}
\value{
A path to a file.
}
\description{
It creates a standardized name for results and checks that the standardized
folders exist.
}
\examples{
# llrs_results("rutils", "DEG", "limma", "a vs b.xlsx")
# llrs_results("rutils", "DEG", "pathways.xlsx")
}
