% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mastodon.R
\name{llrs_send_toot}
\alias{llrs_send_toot}
\title{Send toot}
\usage{
llrs_send_toot(x, width = 500, join_text = ". ")
}
\arguments{
\item{x}{A vector of strings if possible they will be joined}

\item{width}{Allowed width on the server}

\item{join_text}{Character used to join the text}
}
\value{
The ids of the toots posted
}
\description{
Send toot
}
\examples{
\dontrun{
if(requireNamespace(“rtoot”)) {
    llrs_send_toot(c("First message", "Second message"))
}
}
}
