% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onedrive.R
\name{llrs_share}
\alias{llrs_share}
\title{Create a link to a shared file}
\usage{
llrs_share(path, type, expiry, password, scope)
}
\arguments{
\item{path}{The path to the file}

\item{type}{One of c("view", "edit", "embed").}

\item{expiry}{Time it is valid: "x days", "x hours".}

\item{password}{Password protected ?}

\item{scope}{I don't know.}
}
\value{
A link to share a resource
}
\description{
Uses the credentials to share a link to a resource.
}
\seealso{
\code{\link[=llrs_download]{llrs_download()}}
}
