% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{llrs_shiny_create}
\alias{llrs_shiny_create}
\title{Create shiny app}
\usage{
llrs_shiny_create(project, path = "~/ShinyApps/", dest = "/srv/shiny-server/")
}
\arguments{
\item{project}{Name of the project. It can be also multiple folders.}

\item{path}{Path to the folder where it should be.}

\item{dest}{Path to the shiny folder used to serve the website.}
}
\value{
Called by its side effects. It will open a new RStudio project.
}
\description{
Create a shiny app in the right place as well as set the local git repository
to synchronize.
}
\seealso{
\code{\link[=llrs_hook]{llrs_hook()}}
}
