% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_hooks.R
\name{llrs_shiny_hook}
\alias{llrs_shiny_hook}
\title{Creates new shiny app and add hooks.}
\usage{
llrs_shiny_hook(project, path = "~/ShinyApps/", dest = "/srv/shiny-server/")
}
\arguments{
\item{project}{Name of the project. It can be also multiple folders.}

\item{path}{Path to the folder where it should be.}

\item{dest}{Path to the shiny folder used to serve the website.}
}
\value{
A logic value if everything went smoothly.
}
\description{
Creates new shiny app and add hooks.
}
\seealso{
Wrapper on \code{\link[=llrs_shiny_create]{llrs_shiny_create()}} and \code{\link[=llrs_hook]{llrs_hook()}}.
}
