% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{targets}
\alias{targets}
\alias{llrs_tar_load}
\alias{llrs_tar_read}
\title{Functions helping with targets}
\usage{
llrs_tar_load(name, ...)

llrs_tar_read(name, ...)
}
\arguments{
\item{name}{Name of the targets to prepare and load or read.}

\item{...}{Other arguments passed around}
}
\value{
Depending of what \code{\link[=tar_load]{tar_load()}} and \code{tar_read()} returns.
}
\description{
Prepare a target and serve it to the user
}
