% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mastodon.R
\name{llrs_toots}
\alias{llrs_toots}
\title{Send toot}
\usage{
llrs_toots(msg, width = 500, join_text = ". ")
}
\arguments{
\item{msg}{A vector of strings if possible they will be joined}

\item{width}{Allowed width on the server}

\item{join_text}{Character used to join the text}
}
\value{
The ids of the toots posted
}
\description{
Split and merge the text as needed to reduce the number of toots to post.
It only handles text and not images as per \code{rtoot::post_thread()}
}
\examples{
\dontrun{
if(requireNamespace(“rtoot”)) {
    llrs_toots(c("First message", "Second message", "Testing rtoo::post_threads"))
}
}
}
