% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onedrive.R
\name{llrs_upload}
\alias{llrs_upload}
\title{Upload a file to Microsoft cloud in One Drive}
\usage{
llrs_upload(file, dest)
}
\arguments{
\item{file}{File you want to upload.}

\item{dest}{Where you want to sve the file.}
}
\value{
The id of the file uploaded
}
\description{
Folders will be created if they don't exists.
}
