% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{lm_pvalue}
\alias{lm_pvalue}
\title{Extract p-value from lm fit object}
\usage{
lm_pvalue(fit)
}
\arguments{
\item{fit}{lm fit object}
}
\value{
numeric scalar
}
\description{
Extract p-value from lm fit object
}
