% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoadFunctions.R
\name{loadData}
\alias{loadData}
\title{Load Data}
\usage{
loadData(name = stop("'name' must be specified"))
}
\arguments{
\item{name}{filename of the object}
}
\value{
the R object
}
\description{
\code{loadData} loads an Robject in datapath specified in user's Rprofile with filename name
}
