% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoadFunctions.R
\name{loadModel}
\alias{loadModel}
\title{Load Model}
\usage{
loadModel(name = stop("'name' must be specified"))
}
\arguments{
\item{name}{filename of the object}
}
\value{
the R object
}
\description{
\code{loadModel} loads an Robject in modelpath specified in user's Rprofile with filename name
}
