% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log1p.R
\name{Log1p}
\alias{Log1p}
\title{log1p}
\usage{
Log1p(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric, same size as x
}
\description{
Compute log(1+x) in a numerically-stable fashion
}
\examples{
log1p(1e-20)
# compare to naive approach:
log(1+1e-20)
}
