% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_error_with_status}
\alias{log_error_with_status}
\title{Log an error message with status information}
\usage{
log_error_with_status(finished_count, total_count, message)
}
\arguments{
\item{finished_count}{The number of completed tasks.}

\item{total_count}{The total number of tasks.}

\item{message}{A glue-formatted message to log.}
}
\description{
Log an error message with status information
}
