% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsumexp.R
\name{logsumexp}
\alias{logsumexp}
\title{LogSumExp}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{numeric, usually a vector}
}
\value{
numeric, single number
}
\description{
Compute log(sum(exp(x))) in a numerically-stable fashion
}
\examples{
logsumexp(c(-1e5,-1e12))
# compare to log(sum(exp(c(-1e5,-1e12))))
log(sum(exp(c(-1e5,-1e12))))
}
