% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lsh}
\alias{lsh}
\title{List Objects}
\usage{
lsh(up = FALSE, split = TRUE)
}
\arguments{
\item{up}{["UsePryr"] \code{TRUE} to use \code{pryr::object_size}, \code{FALSE} to use \code{utils::object.size}}

\item{split}{(default=\code{TRUE}) split the results by class}
}
\description{
Extends \code{ls()} to associate each object (in current environment) with its (human-readable) size.
}
\examples{
\dontrun{lsh()}
}
\seealso{
ls object.size
}
\keyword{ls}
