% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majority.R
\name{majority}
\alias{majority}
\title{Majority}
\usage{
majority(x, numeric = FALSE)
}
\arguments{
\item{x}{a vector}

\item{numeric}{Logical, whether to return a numeric}
}
\value{
By default returns a character
}
\description{
Find the most frequent element in a vector. Returns it as numeric by default!
}
\examples{

x <- c(1,2,2)
majority(x, numeric = TRUE) # otherwise will return "2" (character)
}
