% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_dir_if_not_exist}
\alias{make_dir_if_not_exist}
\title{Create a directory if it does not exist}
\usage{
make_dir_if_not_exist(dir_path)
}
\arguments{
\item{dir_path}{The path of the directory.}
}
\value{
The directory path.
}
\description{
Create a directory if it does not exist
}
