% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{make_dt_from_msec}
\alias{make_dt_from_msec}
\title{Make lubridate duration object from HH:MM:SS string}
\usage{
make_dt_from_msec(msec, tz = "America/New_York")
}
\arguments{
\item{msec}{Elapsed milliseconds from epoch}

\item{tz}{Timezone}
}
\value{
POSIXct with millisecond accuracy
}
\description{
Make lubridate duration object from HH:MM:SS string
}
