% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{make_duration}
\alias{make_duration}
\title{Make lubridate duration object from HH:MM:SS string}
\usage{
make_duration(hhmmss)
}
\arguments{
\item{hhmmss}{character scalar of HH:MM:SS format}
}
\value{
lubridate's duration object
}
\description{
Make lubridate duration object from HH:MM:SS string
}
