% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{make_today_time}
\alias{make_today_time}
\title{Make today's POSIXct of specified time and timezone}
\usage{
make_today_time(hour = 0, min = 0, sec = 0,
  tz = "America/New_York")
}
\arguments{
\item{hour}{Hour}

\item{min}{Minute}

\item{sec}{Second}

\item{tz}{Timezone}
}
\value{
POSIXct
}
\description{
Make today's POSIXct of specified time and timezone
}
