% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/captionChart.R
\name{makeCaption}
\alias{makeCaption}
\title{makeCaption}
\usage{
makeCaption(captionText = "")
}
\arguments{
\item{captionText}{Text to appear in the caption}
}
\description{
standardise making captions for charts
}
