% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_strings_vectors.R
\name{matchingStringVectors}
\alias{matchingStringVectors}
\title{A function to compare strings (exists or not) between two vectors of strings}
\usage{
matchingStringVectors(searched, searching_in)
}
\arguments{
\item{searched}{vector with words that we want to search in "searching_in" parameter}

\item{searching_in}{vector where we want to search words in "searched" parameter}
}
\description{
test if exist string in another vector strings
Use case: compare headers of dataframes to get if variables are the same
}
\examples{
matchingStringVectors()


}
\keyword{match}
\keyword{strings}
