% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch-memoised-functions.R
\name{memoise_package_fns}
\alias{memoise_package_fns}
\title{Replace package functions with their memoised versions}
\usage{
memoise_package_fns(
  groups = c("read_fns", "slow_fns", "other"),
  functions = list(read_fns = c(), slow_fns = c(), other = "runif"),
  unmemoise = FALSE
)
}
\arguments{
\item{groups}{The groups of functions to memoise}

\item{functions}{The names of the functions in the groups}

\item{unmemoise}{Whether to unmemoise instead of memoise}
}
\value{
Quietly returns NULL
}
\description{
You can only memoise functions which are part of
the package!
}
\examples{
# First make sure `runif` is unmemoised
memoise_package_fns(functions = list(other = "runif"), unmemoise = TRUE)

# Gives different values as expected
RUtils::runif(2)
RUtils::runif(2)

# Now memoise `runif`
memoise_package_fns(functions = list(other = "runif"), unmemoise = FALSE)

# Gives the same value due to memoisation
RUtils::runif(2)
RUtils::runif(2)
}
