% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{message_to_console}
\alias{message_to_console}
\title{Writes a message to console}
\usage{
message_to_console(line)
}
\arguments{
\item{line}{string with the message}
}
\description{
Writes a message to console
}
