% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{minmax}
\alias{minmax}
\title{Numeric trim}
\usage{
minmax(v, min = NA, max = NA, na.value = NA, na.post = NA)
}
\arguments{
\item{v}{input numeric vector}

\item{min}{minimum accepted value}

\item{max}{maximum accepted value}

\item{na.value}{value set to NA elements BEFORE minmax is applied}

\item{na.post}{[only if na.value is NA] value set to NA elements AFTER minmax is applied}
}
\value{
a numeric vector where all values are between `min` and `max`
}
\description{
Trims a numeric input within a user-defined range
}
\examples{
minmax(rnorm(100, 100, 25), 75, 125)
summary((minmax(c(rnorm(1000000, 100, 25), rep(NA, 10)), 75, 125, na.value=-10)))
summary((minmax(c(rnorm(1000000, 100, 25), rep(NA, 10)), 75, 125, na.post=-10)))
}
\keyword{max}
\keyword{min}
\keyword{threshold}
\keyword{trim}
