% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Find the most frequent value in a vector}
\usage{
mode(x)
}
\arguments{
\item{x}{An \code{\link[base:is.recursive]{atomic}} vector.}
}
\description{
\code{mode()} returns the most frequent value in a vector.

If there are more than one mode, it returns all of them. If all values are
unique, it returns \code{NA}.

This function is based on
}
\examples{
mode(c(1, 2, 3, 4, 5, 6))
#> [1] NA # Expected

mode(c(1, 2, 3, 4, 5, 5, 6))
#> [1] 5 # Expected

mode(c(1, 2, 3, 4, 5, 5, 6, NA))
#> [1] 5 # Expected

mode(c(1, 2, 3, 4, 5, 5, 6, 6))
#> [1] 5 6 # Expected
}
\seealso{
Other statistical functions.: 
\code{\link{cohens_d}()}
}
\concept{statistical functions.}
