% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{mondf}
\alias{mondf}
\title{Month difference (vectorized)}
\usage{
mondf(id1, id2, absolute = TRUE)
}
\arguments{
\item{id1}{input date vector (formatted character, IDate, POSIXlt, POSIXct, Date, etc.)}

\item{id2}{input date vector (formatted character, IDate, POSIXlt, POSIXct, Date, etc.)}

\item{absolute}{should the time difference be absolute, or oriented.
When `absolute=FALSE`, `d1` is expected to be before `d2`. Hence returned value will be negative if it's not the case.}
}
\value{
numeric value representing the number of months between two dates. If `absolute` is set to `TRUE`,
this difference will always be positive.
}
\description{
Month difference between two date vectors of same length.
}
\examples{
mondf("2019-01-01", as.Date(Sys.time()))
mondf(as.Date(Sys.time()), "2019-01-01", absolute=FALSE)
mondf(c("2019-01-01", "2018-01-01"), as.Date(Sys.time()))
mondf(c("2019-01-01", "2018-01-01"), rep(as.Date(Sys.time()), 2))
}
\keyword{date}
\keyword{difference}
\keyword{month}
