\docType{package}
\name{mopt}
\alias{mopt}
\alias{mopt-package}
\alias{rutils}
\alias{rutils-package}
\title{Utility functions for R}
\description{
  Utility functions for R
}
\author{
  Thibaut Lamadon \email{thibaut.lamadon@gmail.com}
}
\references{
  \url{https://github.com/tlamadon/rutils}
}
\seealso{
  rutils
}

