% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{most_frequent}
\alias{most_frequent}
\title{Returns the most frequent element of given vector \code{v}}
\usage{
most_frequent(v, na.rm = F)
}
\arguments{
\item{v}{Vector}

\item{na.rm}{Logical Should missing values be removed?}
}
\value{
The most frequent element of given vector \code{v}
}
\description{
Returns the most frequent element of given vector \code{v}
}
