% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_merge.R
\name{multi_merge}
\alias{multi_merge}
\title{Combine different matrices by row names and column names}
\usage{
multi_merge(..., na_to_zero = FALSE)
}
\arguments{
\item{...}{one or several \code{matrix} objects.}

\item{na_to_zero}{a \code{logical} value. If \code{TRUE} gaps (i.e. unknown edges) are
coded as \code{0}. Otherwise they are coded as \code{NA} (default).}
}
\value{
A \code{matrix} object.
}
\description{
This function combines different matrices by row names and column names by
performing a 2-dimension full join. Gaps are filled with \code{NA} (default) or
\code{0} (argument \code{na_to_zero}).
}
\examples{
mat1 <- matrix(rep(1, 9), nrow = 3)
colnames(mat1) <- c("A", "B", "C")
rownames(mat1) <- c("A", "B", "C")

mat2 <- matrix(rep(1, 9), nrow = 3)
colnames(mat2) <- c("D", "E", "F")
rownames(mat2) <- c("D", "E", "F")

mat3 <- matrix(rep(1, 9), nrow = 3)
colnames(mat3) <- c("F", "G", "H")
rownames(mat3) <- c("G", "A", "H")

multi_merge(mat1, mat2, mat3)

multi_merge(mat1, mat2, mat3, na_to_zero = TRUE)
}
