% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicator.R
\name{multiplicator}
\alias{multiplicator}
\title{Multiplicator helper}
\usage{
multiplicator(x)
}
\arguments{
\item{x}{the multiplicator}
}
\value{
A function
}
\description{
A helper function to create a function which
multiplies by a specific number
}
\examples{
double <- multiplicator(2)
triple <- multiplicator(3)
quadruple <- multiplicator(4)

x <- 1:3
double(x)
triple(x)
quadruple(x)
}
