% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_as.R
\name{na_as}
\alias{na_as}
\alias{na_as.logical}
\alias{na_as.character}
\alias{na_as.integer}
\alias{na_as.numeric}
\alias{na_as.Duration}
\alias{na_as.Period}
\alias{na_as.difftime}
\alias{na_as.hms}
\alias{na_as.Date}
\alias{na_as.POSIXct}
\alias{na_as.POSIXlt}
\alias{na_as.Interval}
\title{Return a \code{NA} value of the same type of an R object}
\usage{
na_as(x)

\method{na_as}{logical}(x)

\method{na_as}{character}(x)

\method{na_as}{integer}(x)

\method{na_as}{numeric}(x)

\method{na_as}{Duration}(x)

\method{na_as}{Period}(x)

\method{na_as}{difftime}(x)

\method{na_as}{hms}(x)

\method{na_as}{Date}(x)

\method{na_as}{hms}(x)

\method{na_as}{POSIXct}(x)

\method{na_as}{POSIXlt}(x)

\method{na_as}{Interval}(x)
}
\arguments{
\item{x}{An \link[checkmate:checkAtomic]{atomic} vector.}
}
\value{
A \code{NA} value with the same class and attributes of \code{x}.
}
\description{
\code{na_as()} returns an \code{NA} value with the same class and attributes as a
given R object.

This function is useful for dynamically assigning \code{NA} values that
preserve the type and attributes of the original object.
}
\examples{
na_as(TRUE)
#> [1] NA # Expected

class(na_as(TRUE))
#> [1] "logical" # Expected

na_as(as.Date("2020-01-01"))
#> [1] NA # Expected

class(na_as(as.Date("2020-01-01")))
#> [1] "Date"  # Expected
}
\concept{parsing/conversion functions}
