% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{nalocf}
\alias{nalocf}
\title{Replace \code{NA} values with the most recent non-\code{NA} values prior to
them.}
\usage{
nalocf(inputv, fromLast = FALSE, narm = FALSE, maxgap = NROW(inputv))
}
\arguments{
\item{\code{inputv}}{A \emph{numeric} or \emph{Boolean} vector or matrix, or
\emph{xts} time series.}

\item{\code{fromLast}}{A  \emph{Boolean} argument: should non-\code{NA}
values be carried backward rather than forward? (the default is \code{FALSE})}

\item{\code{narm}}{A \emph{Boolean} argument: should any remaining (leading
or trailing) \code{NA} values be removed? (the default is \code{FALSE})}

\item{\code{maxgap}}{The maximum number of neighboring \code{NA} values that
can be replaced (the default is \code{NROW(inputv)}).}
}
\value{
A vector, matrix, or \emph{xts} time series with the same dimensions
  and data type as the argument \code{inputv}.
}
\description{
Replace \code{NA} values with the most recent non-\code{NA} values prior to
them.
}
\details{
The function \code{nalocf()} replaces \code{NA} values with the
  most recent non-\code{NA} values prior to them.

  If the \code{fromLast} argument is \code{FALSE} (the default), then the
  previous or past non-\code{NA} values are carried forward to replace the
  \code{NA} values.
  If the \code{fromLast} argument is \code{TRUE}, then the following or
  future non-\code{NA} values are carried backward to replace the \code{NA}
  values.

  The function \code{nalocf()} performs the same operations as the function
  \code{zoo::na.locf()} from package
  \href{https://cran.r-project.org/web/packages/zoo/index.html}{zoo}.

  The function \code{nalocf()} calls the function \code{xts:::na.locf.xts()}
  from package
  \href{https://cran.r-project.org/web/packages/xts/index.html}{xts}.
}
\examples{
# Create vector containing NA values
inputv <- sample(22)
inputv[sample(NROW(inputv), 4)] <- NA
# Replace NA values with the most recent non-NA values
rutils::nalocf(inputv)
# Create matrix containing NA values
inputv <- sample(44)
inputv[sample(NROW(inputv), 8)] <- NA
inputv <- matrix(inputv, nc=2)
# Replace NA values with the most recent non-NA values
rutils::nalocf(inputv)
# Create xts series containing NA values
inputv <- xts::xts(inputv, order.by=seq.Date(from=Sys.Date(),
  by=1, length.out=NROW(inputv)))
# Replace NA values with the most recent non-NA values
rutils::nalocf(inputv, fromLast=TRUE)
}
