% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inc.utils.r
\name{nonlinear.map}
\alias{nonlinear.map}
\title{nonlinear mapping from [a,b] to [0,c]}
\usage{
nonlinear.map(z, low, high, type)
}
\arguments{
\item{x}{ordered array. a grid.}

\item{new.up}{new upper bound}
}
\description{
nonlinear mapping from [a,b] to [0,c]
}

