% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Title}
\usage{
normalize(x)
}
\arguments{
\item{x}{spatRaster}
}
\value{
spat raster with range 0-1
}
\description{
The normalize() function is based on the Min-Max Scaling method. The normalized values fall in
TODO: Rewrite to work on vector
}
