% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_extdata.R
\name{normalize_extdata}
\alias{normalize_extdata}
\title{Normalize file names from the \code{extdata} folder}
\usage{
normalize_extdata(package = NULL)
}
\arguments{
\item{package}{(optional) a string indicating the target package. If not
assigned, the function will try to use the name of the active project
directory (requires the \code{rstudioapi} package).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{normalize_extdata()} normalize file names founded in the \verb{./inst/extdata/}
folder of a package. \strong{This function must be used only with
\href{https://posit.co/products/open-source/rstudio/}{RStudio} IDE}.

The function lowercase the names, substitute \code{" - "} characters for \code{"_"},
and substitute \code{" "} characters for \code{"-"}.
}
\examples{
\dontrun{
normalize_extdata()}
}
\seealso{
Other R system functions: 
\code{\link{raw_data_1}()},
\code{\link{raw_data_2}()},
\code{\link{update_pkg_versions}()}
}
\concept{R system functions}
