% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_names.R
\name{normalize_names}
\alias{normalize_names}
\title{Normalize file names}
\usage{
normalize_names(
  path = clipr::read_clip(),
  exceptions = c("^README", "^OFL.txt$", "^DESCRIPTION", "^Google README.txt$"),
  just_dirs = FALSE
)
}
\arguments{
\item{path}{(Optional) A \code{\link[base:character]{character}} string indicating
the files directory path. Defaults to clipboard content.}

\item{exceptions}{(Optional) A \code{\link[base:character]{character}} vector of
regex patterns for files to exclude from renaming.}

\item{just_dirs}{(Optional) A \code{\link[base:logical]{logical}} flag indicating
if only directory names should be normalized (Default: \code{FALSE}).}
}
\value{
An invisible \code{NULL}. This function is used for its side effect.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{normalize_names()} normalize file names to make them machine-readable.
It converts names to lowercase and replaces spaces and special characters
with hyphens.

This function follows the naming conventions from the
\href{https://style.tidyverse.org/files.html#names}{Tidyverse Style Guide}.
}
\examples{
dir <- tempfile("")
dir.create(dir) |> invisible()

file <- tempfile("Test - Test-", tmpdir = dir)
file.create(file) |> invisible()

normalize_names(dir)

list.files(dir)
#> [1] "test-test-***" # Expected
}
\concept{file functions.}
