% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{notify}
\alias{NOTIFY}
\alias{notify}
\title{Stub per funzione di notify}
\usage{
notify(titolo, msg, id)

NOTIFY(titolo, msg, id = NULL)
}
\arguments{
\item{titolo}{titolo messaggio}

\item{msg}{testo del messaggio}

\item{id}{eventuale id (per usi futuri)}
}
\description{
Per ora non fan nulla ma la modifichero' a breve.
}

