% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{notify_system}
\alias{notify_system}
\title{Use system's notify-send to send a message}
\usage{
notify_system(..., title, message = NULL, icon = NULL, urgency = NULL)
}
\arguments{
\item{...}{Ignored arguments (to make it compatible with a pipe)}

\item{title}{Title of the message: character shorter than 40 characters.}

\item{message}{Message of the notification.}

\item{icon}{Icon, to pick from the system.}

\item{urgency}{Urgency of the system}
}
\value{
Always TRUE (even if no notification can be created).
}
\description{
Creates a pop up message on the system if it is run from an interactive
session.
}
\examples{
notify_system(title = "Works", message = "My first message")
}
