% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{order_df}
\alias{order_df}
\title{Order a dataframe by a given column}
\usage{
order_df(df, x, decreasing = FALSE)
}
\arguments{
\item{df}{the dataframe to order}

\item{x}{the column to order by}

\item{decreasing}{logical. Is the order increasing or decreasing?}
}
\value{
An ordered dataframe
}
\description{
A simple wrapper round the base R \code{order} function allowing chaining with \code{\%>\%}
}
\examples{
order_df(mtcars, ~ cyl)

# Same operation can be done with the pipe operator

mtcars \%>\% order_df(~ cyl)
}
