% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeHome.R
\name{outputFile}
\alias{outputFile}
\title{provide wrapper to path output files}
\usage{
outputFile(fileName)
}
\arguments{
\item{fileName}{of expected file in path}
}
\description{
provide wrapper to path output files
}
