% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_as_list.R
\name{path_as_list}
\alias{path_as_list}
\title{Create a list of paths}
\usage{
path_as_list(x)
}
\arguments{
\item{x}{vector of paths to put in a list}
}
\description{
This function creates a named list with full paths from a vector of paths.
}
\details{
By default cleans the starting digits if any. E.g. will remove \code{01} from \code{foo/01_data}
}
\examples{
x <- c("foo/foo", "foo/bar", "foo/baz")
path_as_list(x)

}
