% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_basename.R
\name{path_basename}
\alias{path_basename}
\title{Take the file name}
\usage{
path_basename(x, ext = FALSE)
}
\arguments{
\item{x}{path string}

\item{ext}{logical. Whether to keep the extension. Default is FALE}
}
\description{
Return the filename without extension.
}
\details{
Wrapper around \code{fs::path_ext_remove(fs::path_file(x))}
}
