% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_date.R
\name{path_date}
\alias{path_date}
\title{Extract date from a (path) string}
\usage{
path_date(x, format = "\%Y\%m\%dT\%H\%M\%S", convert = TRUE)
}
\arguments{
\item{x}{string containing a date portion.
Currently only searches for this pattern \code{20220103T140533}}

\item{format}{output date format.}

\item{convert}{whether to convert to the specified format. If \code{FALSE} will return the literal string}
}
\description{
Originally created to extract the date from the Sentinel-2 filename.
Work reliably only for Sentinel-2 filenames
}
\details{
NB!: it will extract only the first match
TODO: make a generic function to search for a date-like substring
}
