% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_decompose.R
\name{path_decompose}
\alias{path_decompose}
\alias{path_compose}
\title{Decompose path}
\usage{
path_decompose(x)

path_compose(x)
}
\arguments{
\item{x}{path or a data.frame of decomposed paths (for \code{pathcompose()}}
}
\value{
returns a tibble with columns  \code{dir}, \code{name}, and \code{ext}
}
\description{
Decomposes a path string into its main elements: \code{directory}, \code{filename}, and \code{extension}
}
\details{
\code{path_compose()} NB! Experimental. It iss the opposite function, e.g. reverts back the decomposed.
}
\examples{
x <- c("foo/bar.ext", "long/path/foo/baz.ext")
path_decompose(x)
}
