% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_exists.R
\name{path_exists}
\alias{path_exists}
\title{Does a folder exist}
\usage{
path_exists(path, silent = TRUE, create = FALSE, verbose = FALSE)
}
\arguments{
\item{path}{The path to be checked}

\item{silent}{logical. whether to return answer or remain silent.}

\item{create}{logical. Should the folder be created?}

\item{verbose}{If TRUE and \code{silent = TRUE} returns the path. Useful for filename creation}
}
\value{
If \code{check = TRUE} returns the result from \code{dir.exists}
               If \code{create = TRUE} creates folder if missing
               If \code{verbose = TRUE} returns the path, otherwise silent .
}
\description{
Slightly more powerful check. It checks if a folder exists, creates it if \code{create = TRUE} (the default)
}
