% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_extract.R
\name{path_extract}
\alias{path_extract}
\title{Extract position of filename string}
\usage{
path_extract(string, pos = 1, sep = "_")
}
\arguments{
\item{string}{string (typically a filename) from which a prefix is to be extracted}

\item{pos}{the position of the prefix in the string (default 1 - first prefix)}

\item{sep}{the separator which divides the prefix from the rest of the string}
}
\value{
String
}
\description{
Used to extract a a certain position based on splitting regexp from a string (typically a filename). If the string is the \code{form/of/a/long/path}, the path will be drooped
}
\examples{
x <- c("foo/bar.ext", "long/path/foo/baz.ext")
path_extract(x, pos = 1, sep = "a")
path_extract(x, pos = 2, sep = "a")
}
