% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_find.R
\name{path_find}
\alias{path_find}
\title{File search}
\usage{
path_find(...)
}
\arguments{
\item{path}{path to search in}

\item{regexp}{regular expression}
}
\description{
Recursive search for files in given path, which will not fail if access is denied.
A wrapper around \code{fs::dir_ls(..., recurse = T, fail = F)}
}
\examples{
path_find(".") # prints everything in your current directory!
}
