% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_insert.R
\name{path_insert}
\alias{path_insert}
\title{Insert string}
\usage{
path_insert(x, string, pos, sep = "_")
}
\arguments{
\item{x}{string to be edited}

\item{string}{string to be inserted}

\item{pos}{position to be inserted to}

\item{sep}{separator passed to \code{str_split} to identify positions}
}
\description{
Insert a string  at specific separation of the filename portion of a path.
}
\details{
\code{path_insert("path/file_name.ext", "foo", 1)} -> \code{"path/file_foo_name.ext"}
}
\note{
if  the string is inserted AFTER the position.
       \code{pos = 0} the string is pasted at first place, if \code{pos >= length(str_split(x))} the string
       is appended to the end of the string.
}
