% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_tifs.R
\name{path_tifs}
\alias{path_tifs}
\title{Recursive search for files with extension \code{.tif$}}
\usage{
path_tifs(database = ".", pattern, negate = FALSE)
}
\arguments{
\item{database}{location to search in, it will search all subfolders (recursive search)}

\item{pattern}{optionally, a further pattern to subset by. passed to \code{stringr::str_subset}}

\item{negate}{weather to invert the subsetting and keep those record not matching the pattern}
}
\description{
Recursive search for files with extension \code{.tif$}
}
