% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_tile.R
\name{path_tile}
\alias{path_tile}
\title{Find Tile name}
\usage{
path_tile(path)
}
\arguments{
\item{path}{string. Path to standard Sentinel-2 file.}
}
\description{
Extracts the tile identifier from Sentinel-2 file name.
}
