% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{percent_empty}
\alias{percent_empty}
\title{Percent of values in a vector that are NA}
\usage{
percent_empty(x)
}
\arguments{
\item{x}{Vetor of values}
}
\value{
Percent of values in x that are NA
}
\description{
Takes a vector and calculates the percent of values that are NA
}
\examples{
percent_empty(c(1,2,NA,4))

# Use with purrr to test columns in matrix, dataframe etc.

nadf <- data.frame(x = rep(NA, 3), y = c(1, 2, NA))
purrr::map_dbl(nadf, ~ percent_empty(.))
}
