% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pie2}
\alias{pie2}
\title{Extended pie chart function}
\usage{
pie2(x, title = "")
}
\arguments{
\item{x}{a vector of numeric values}

\item{title}{the title of the chart}
}
\value{
a plot object that represents a pie chart
}
\description{
This function is a shortcut to make an ordered pie chart with clear labels.
}
\examples{
pie2(saf(round(rnorm(150))))
}
\keyword{chart}
\keyword{pie}
\keyword{plot}
