% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{plot_acf}
\alias{plot_acf}
\title{Calculate the autocorrelation function (ACF) for a time series of returns,
and plot it.}
\usage{
plot_acf(
  xtes,
  lagg = 10,
  plotobj = TRUE,
  xlab = "Lag",
  ylab = "",
  main = "",
  ...
)
}
\arguments{
\item{\code{xtes}}{A vector, matrix, or time series of returns.}

\item{\code{lagg}}{The maximum lag at which to calculate the ACF (default is
\code{10}).}

\item{\code{plotobj}}{A \emph{Boolean} argument: should a plot be made?
(default is \code{TRUE})}

\item{\code{xlab}}{A string with the x-axis label.}

\item{\code{ylab}}{A string with the y-axis label.}

\item{\code{main}}{A string with the plot title.}

\item{\code{...}}{Additional arguments to the function \code{stats::acf()}.}
}
\value{
Returns the ACF data invisibly and creates a plot.
}
\description{
Calculate the autocorrelation function (ACF) for a time series of returns,
and plot it.
}
\details{
The function \code{plot_acf()} calculates the autocorrelation
  function (ACF) for a time series of returns, and plots it.  The function
  \code{plot_acf()} is just a wrapper for the function \code{stats::acf()}.
  The function \code{stats::acf()} calculates the autocorrelation function,
  including the lag zero autocorrelation, which is by definition equal to
  \code{1}.

  The function \code{plot_acf()} calls the function \code{stats::acf()},
  removes the spurious lag zero autocorrelation, creates a plot, and returns
  the ACF data invisibly.
}
\examples{
# Plot the ACF of random returns
rutils::plot_acf(rnorm(1e4), lag=10, main="ACF of Random Returns")
# Plot the ACF of VTI returns
rutils::plot_acf(na.omit(rutils::etfenv$returns$VTI), lag=10, main="ACF of VTI Returns")
}
