% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_functions.R
\name{plot_auc}
\alias{plot_auc}
\title{Plot AUC Curve}
\usage{
plot_auc(results)
}
\arguments{
\item{results}{data frame with predicted probabilities under column named 'prob' and results
under column named 'actual'}
}
\value{
plots the ROC curve and returns the AUC
}
\description{
\code{plot_auc} plots the ROC curve and displays the AUC for a list of predictions and actuals
}
