% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_functions.R
\name{plot_auc_color}
\alias{plot_auc_color}
\title{Plot AUC Curve with Color Cutoff}
\usage{
plot_auc_color(results, ggplot = TRUE)
}
\arguments{
\item{results}{data frame with predicted probabilities under column named 'prob' and results
under column named 'actual'}

\item{ggplot}{TRUE or FALSE depending on if you want a ggplot or normal r plot}
}
\value{
plots the ROC curve
}
\description{
\code{plot_auc_color} plots the ROC curve with a color coded cutoff and displays the AUC for a
list of predictions and actuals
}
