% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_hm}
\alias{plot_hm}
\title{Plots a heatmap ordered by cell and gene clusters of the logcounts.}
\usage{
plot_hm(
  sce,
  cell_clusters,
  gene_clusters,
  show_column_title = character(0),
  show_row_title = character(0),
  row_title_rot = 90,
  show_row_names = FALSE
)
}
\description{
Plots a heatmap ordered by cell and gene clusters of the logcounts.
}
