% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_functions.R
\name{plot_pr}
\alias{plot_pr}
\title{Plot PR Curve}
\usage{
plot_pr(results)
}
\arguments{
\item{results}{data frame with predicted probabilities under column named 'prob' and numeric
results under column named 'actual_num'}
}
\value{
plots the PR curve and returns the AUC
}
\description{
\code{plot_pr} plots the PR curve and displays the AUC for a list of predictions and actuals
}
