% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_functions.R
\name{plot_pr_color}
\alias{plot_pr_color}
\title{Plot PR Curve with Color Cutoff}
\usage{
plot_pr_color(results, ggplot = TRUE)
}
\arguments{
\item{results}{data frame with predicted probabilities under column named 'prob' and numeric
results under column named 'actual_num'}

\item{ggplot}{TRUE or FALSE depending on if you want a ggplot or normal r plot}
}
\value{
plots the PR curve
}
\description{
\code{plot_pr_color} plots the PR curve with a color coded cutoff and displays the AUC for a
list of predictions and actuals
}
