% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFCC.R
\name{plotFCC}
\alias{plotFCC}
\title{NIR-Red-Green plot of a multi-layered SpatRaster}
\usage{
plotFCC(x, r = 4, g = 1, b = 2, ...)
}
\arguments{
\item{x}{SpatRaster}

\item{...}{}
}
\description{
NIR-Red-Green plot of a multi-layered SpatRaster
}
