% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plotly_stacked_area}
\alias{plotly_stacked_area}
\title{Simple stacked area chart using plotly}
\usage{
plotly_stacked_area(
  dt,
  x,
  y,
  group,
  xname = NULL,
  yname = NULL,
  gname = NULL,
  title = NULL
)
}
\arguments{
\item{dt}{data.table to use data from}

\item{x}{x variable (typically, time)}

\item{y}{y variable (typically, what you want to observe)}

\item{group}{group variable (typically, what you want to compare)}

\item{xname}{label for x axis}

\item{yname}{label for y axis}

\item{gname}{label for group}

\item{title}{plot title}
}
\value{
a plot object that represents a stacked area chart
}
\description{
This function is a shortcut to plot stacked area chart using plotly.
}
\keyword{area}
\keyword{plotly}
\keyword{stack}
