% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeHome.R
\name{popFile}
\alias{popFile}
\title{provide wrapper to path to reference population data files}
\usage{
popFile(fileName)
}
\arguments{
\item{fileName}{of expected file in path}
}
\description{
provide wrapper to path to reference population data files
}
