% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{preprocess_data_set}
\alias{preprocess_data_set}
\title{Preprocess any SingleCellExperiment}
\usage{
preprocess_data_set(
  sce,
  org,
  mt_filter = TRUE,
  pct = 0.01,
  modPoisson = FALSE,
  rmbatch = FALSE,
  batchcol = NULL,
  truth = NULL,
  return_plots = FALSE
)
}
\description{
Preprocess any SingleCellExperiment
}
