% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA_utils.R
\name{prin_to_std}
\alias{prin_to_std}
\title{prin_coords: principal coordinates (row-vectors)
D: singular vectors.}
\usage{
prin_to_std(prin_coords, D)
}
\description{
prin_coords: principal coordinates (row-vectors)
D: singular vectors.
}
