% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintF.R
\name{PrintF}
\alias{PrintF}
\title{PrintF}
\usage{
PrintF(...)
}
\arguments{
\item{\code{...}}{Passed to \code{sprintf}}
}
\value{
Character vector
}
\description{
Implementation of \code{bash}-like \code{printf}. Wrapper to \code{sprintf}
that prints to \code{stderr}.
}

