% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop}
\alias{prop}
\title{Compute the proportion of values in a vector}
\usage{
prop(x, value, na_rm = TRUE)
}
\arguments{
\item{x}{An \code{\link[base:vector]{atomic}} vector.}

\item{value}{An \code{\link[base:vector]{atomic}} vector containing the unique values
in \code{x} for which to compute proportions.}

\item{na_rm}{A \code{\link[base:logical]{logical}} flag indicating whether to
remove \code{NA} values from \code{x} before computing proportions (default: \code{TRUE}).}
}
\value{
A \code{\link[base:numeric]{numeric}} vector with the proportions of each
value in \code{value} relative to the total number of values in \code{x}.
}
\description{
\code{prop()} computes the proportion of specified values in a vector.
}
\examples{
prop(c("a", "b", "a", "c", "b", "a"), c("a", "b"))
#> [1] 0.500000 0.3333333

prop(c(1, 2, 1, 2, 1), c(1, 2))
#> [1] 0.6 0.4

prop(c(TRUE, TRUE, TRUE, FALSE), c(TRUE, FALSE))
#> [1] 0.75 0.25
}
\seealso{
Other statistical functions: 
\code{\link{cut_interval_mean}()},
\code{\link{remove_outliers}()},
\code{\link{std_error}()},
\code{\link{test_outlier}()}
}
\concept{statistical functions}
