% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull}
\alias{pull}
\alias{first}
\alias{last}
\title{Pull element}
\usage{
pull(x, v)

first(x)

last(x)
}
\arguments{
\item{x}{list like object}

\item{v}{the elements to pull}
}
\description{
Pull an element from a list like object. Same as \code{x[[v]]}.
For use in native pipe synthax
}
\details{
\code{pull(x, v)} is same as \code{x[[v]]}

\code{first(x)} is same as \code{x[[1]]}

\code{last(x)} is same as \code{x[[length(x)]]}
}
\examples{
x <- list(1,2,3)
pull(x, 2)
first(x)
last(x)

}
