% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_code.R
\name{qr_code}
\alias{qr_code}
\title{Generate a QR code}
\usage{
qr_code(link, file, logo = NULL, frame = NULL, offset = "+50+50")
}
\arguments{
\item{link}{A \code{\link[base:character]{character}} string containing the link to
encode.}

\item{file}{A \code{\link[base:character]{character}} string indicating the path to
the output file.}

\item{logo}{(optional) A \code{\link[base:character]{character}} string indicating
the path to a logo image file.}

\item{frame}{(optional) A \code{\link[base:character]{character}} string indicating
the path to a frame file.}

\item{offset}{(optional) A \code{\link[base:character]{character}} string indicating
the offset for positioning the QR code within the frame. See
\code{\link[magick:composite]{image_composite()}} to learn more.}
}
\value{
An invisible \code{NULL}. This function is called for its side effects.
}
\description{
\code{qr_code()} creates a QR code from a given link and saves it as a
\href{https://en.wikipedia.org/wiki/PNG}{PNG} file.

The QR code has a fixed size: 500 px without a frame, and 450 px when a
frame is included.
}
\examples{
library(ggplot2)
library(magick)

link <- "https://github.com/danielvartan"
file <- tempfile(fileext = ".png")

# A Simple QR Code -----

qr_code(link = link, file = file)

file |>
  image_read() |>
  image_ggplot()

# QR Code + Logo -----

qr_code(
  link = link,
  file = file,
  logo = raw_data_1("github-icon.svg", "rutils")
)

file |>
  image_read() |>
  image_ggplot()

# QR Code + Logo + Frame -----

qr_code(
  link = "https://github.com/danielvartan",
  file = file,
  logo = raw_data_1("github-icon.svg", "rutils"),
  frame = raw_data_1("qr-code-frame.svg", "rutils")
)

file |>
  image_read() |>
  image_ggplot()
}
\concept{image functions}
