% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_quarter.r
\name{quarter}
\alias{quarter}
\title{Return the current quarter}
\usage{
quarter(date = Sys.Date())
}
\arguments{
\item{date}{the date of the quarter}
}
\value{
the numeric repr of quarter
}
\description{
Return the current quarter
}
