% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querying_redshift.R
\name{queryingRedshift}
\alias{queryingRedshift}
\title{A function to querying Redshift}
\usage{
queryingRedshift(qry, dbname, host, port, user, password)
}
\arguments{
\item{qry}{vector with words that we want to search in "searching_in" parameter}

\item{...}{credentials}
}
\description{
A function to querying Redshift
}
\examples{
matchingStringVectors()


}
\keyword{query}
\keyword{redshift}
