% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querying_presto.R
\name{queryingSource}
\alias{queryingSource}
\title{A function to querying Presto}
\usage{
queryingSource(qry)
}
\arguments{
\item{qry}{query to launch}
}
\description{
A function to querying Presto
}
\examples{
queryingSource()


}
\keyword{presto}
\keyword{query}
