% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_crs_check_crop.R
\name{r_crs_check_crop}
\alias{r_crs_check_crop}
\title{Crop with crs check}
\usage{
r_crs_check_crop(x, y, ...)
}
\arguments{
\item{x}{spatial raster}

\item{y}{object of class sf}

\item{...}{further parameters to `\code{\link[terra]{crop}} and \code{\link[terra]{writeRaster}}}
}
\description{
CRS safe crop. Checks if the crs of y is same as this of x and transforms it if not.
}
