% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_fc.R
\name{r_fc}
\alias{r_fc}
\title{Rescale to 0 - 1}
\usage{
r_fc(x, upper = 0.3, filename = "", clamping = TRUE, wopt = list())
}
\arguments{
\item{x}{object of class \code{SpatRaster}}

\item{upper}{upper threshold to clamp the raster}

\item{wopt}{additional list of named arguments to \code{terra::app}}
}
\value{
rescaled raster
}
\description{
rescale a raster from 0 to the specified upper threshold
}
\examples{
a<- terra::rast(system.file("ex/logo.tif", package="terra"))
res<- r_fc(a, upper = 100)
plot(res)
# with no clamping:
res2<- r_fc(a, upper = 100, clamping = FALSE)
plot(res2)
}
