% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_harden.R
\name{r_harden}
\alias{r_harden}
\title{Harden a soft raster layer}
\usage{
r_harden(r, threshold, above = FALSE, ...)
}
\arguments{
\item{r}{spatRaster object}

\item{threshold}{numeric threshold value}

\item{above}{invert the mask, e.g. keep pixels with values above the threshold as 1}

\item{...}{further arguments to \code{terra::classify}, useful to decide how to treat the \code{threshold}
e.g \code{include.lowest, right}`}
}
\description{
Produce a Mask layer from a raster with continuous values by converting
all values bellow \code{threshold} into 1 and everything else into NA.
}
\details{
TODO: user-defined value instead of 1
TODO: user-defined value instead of NA
TODO: keep = TRUE to keep original values
TODO: explain how to include or exclude the threshold (include.lowest, right options in reclassify)
TODO: see if its worth merging with r_reclassify_threshold
}
\examples{
a<- terra::rast(system.file("ex/logo.tif", package="terra"))
res<- r_harden(a, 100)
res
# plot(res)
}
