% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_hasdata_cell.R
\name{r_hasdata_cell}
\alias{r_hasdata_cell}
\title{Raster cell has data across all cells in a stack}
\usage{
r_hasdata_cell(r, na.rm = TRUE)
}
\arguments{
\item{r}{raster stack}

\item{na.rm}{if TRUE tests if ALL cells are NA, if FALSE test if ANY cell is NA}
}
\value{
a vector with the indices of the cells with non-NA values
}
\description{
TODO: rewrite is.na(r)
}
\examples{
a<- terra::rast(system.file("ex/logo.tif", package="terra"))
res<- r_hasdata_cell(a)
res
}
