% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_hasdata_layer.R
\name{r_hasdata_layer}
\alias{r_hasdata_layer}
\title{Raster layer has data}
\usage{
r_hasdata_layer(r)
}
\arguments{
\item{r}{is a spatRaster}
}
\value{
logical indicating indices of the layers with data
}
\description{
Checks if layer is empty
}
\examples{
a<- terra::rast(system.file("ex/logo.tif", package="terra"))
a_r<- raster::stack(system.file("ex/logo.tif", package="terra"))
res<- r_hasdata_layer(a)
res

}
