% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_init.R
\name{r_init_const}
\alias{r_init_const}
\title{Initialize a constant raster}
\usage{
r_init_const(x, const, ...)
}
\arguments{
\item{x}{spatRaster object from which to take the metadata}

\item{const}{numeric value to be used as constant for the entire raster}

\item{...}{further parameters to \code{writeRaster}}
}
\value{
spatRaster
}
\description{
Casts a raster from a single value
Not needed anymore - sill reimplemented in terra
}
\examples{

r <- terra::rast(ncols=36, nrows=18)

r_init_const(r, 1)

# Equivalent to:
const <- 1
set_const <- function(x){rep(const, x)}
init(r, fun=set_const)
}
