% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_mask_fill.R
\name{r_mask_fill}
\alias{r_mask_fill}
\title{Fill holes}
\usage{
r_mask_fill(r, hole.size, directions = 8, ...)
}
\arguments{
\item{r}{spatRaster image with holes (NA)}

\item{hole.size}{maximum hole size (in pixels) to close}

\item{directions}{should be 4 or 8 (default), for rook or queen neigbhours respectively.}

\item{...}{further arguments to `raster::writeRaster`}
}
\value{
mask raster with holes filled
}
\description{
Fill holes in mask image
}
\examples{
library(terra)

r <- terra::rast(ncol = 10, nrow = 10)
values(r) <- runif(100)
r[r<0.7] <- NA

clumped <- patches(r, directions = 8)
plot(clumped)
masked <- r_mask_patches(clumped, 2)
plot(masked)

masked_fill<- r_mask_fill(r, Inf)
plot(masked_fill)


}
\seealso{
r_mask_patches
}
