% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_mask_invert.R
\name{r_mask_invert}
\alias{r_mask_invert}
\title{Invert mask}
\usage{
r_mask_invert(mask, ...)
}
\arguments{
\item{mask}{input mask}

\item{...}{additional arguments for writing files as in writeRaster}
}
\value{
raster with values 1 and NA
}
\description{
Produce the opposite (negative) mask
}
\examples{
library(terra)
r <- rast(ncol = 10, nrow = 10)
values(r) <- runif(100)
r[r<0.7] <- NA
plot(r)

inverted <- r_mask_invert(r)
plot(inverted)
}
