% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_mask_patches.R
\name{r_mask_patches}
\alias{r_mask_patches}
\title{Mask out small patches}
\usage{
r_mask_patches(r, patch.size = 30, ...)
}
\arguments{
\item{r}{clumped raster to be analyzed, as produced by \code{terra::patches}}

\item{patch.size}{minimum patch size to be kept (pixels). All patches smaller than (<) this value are masked out}

\item{...}{further parameters passed to \code{writeRaster}}
}
\value{
Raster
}
\description{
This function  masks patches of a clumped raster smaller than user defined size (in pixels). The clumped raster
has to have unique value for each clump (as produced by \code{terra::patches}).
}
\examples{
library(terra)

r <- terra::rast(ncol = 10, nrow = 10)
values(r) <- runif(100)
r[r<0.7] <- NA

clumped <- patches(r, directions = 8)
plot(clumped)
masked <- r_mask_patches(clumped, 2)
plot(masked)
}
\seealso{
\link[gdfR]{r_mask_fill}
}
