% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_reclassfiy_threshold.R
\name{r_reclassify_threshold}
\alias{r_reclassify_threshold}
\title{Reclassify raster based on threshold}
\usage{
r_reclassify_threshold(r, threshold, becomes, ...)
}
\arguments{
\item{r}{raster to reclassify}

\item{threshold}{sequence of threshold values}

\item{becomes}{the output classes}

\item{...}{parameters to \code{terra::classify}}
}
\description{
Wrapper around \code{terra::classify()} which creates
the reclassification matrix from \code{threshold}, \code{becomes}
}
\examples{
a<- terra::rast(system.file("ex/logo.tif", package="terra"))
# make everything above 50 to be 50 and everything below NA
res<- r_reclassify_threshold(a, 50, c(NA, 50))
res
plot(res)
}
