% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_residuals.R
\name{r_residuals}
\alias{r_residuals}
\title{Calculate residuals image}
\usage{
r_residuals(x, y, ...)
}
\arguments{
\item{x}{the predictor raster}

\item{y}{the predictant raster. If missing \code{x} must be a raster with two layers
(or list length 2 of one-layer rasters), where first layer \code{x[[1]]} is the predictor (x)
and the second layer \code{x[[2]]} is the predictant (y)}

\item{...}{arguments passed to \code{writeRaster}}
}
\value{
\code{raster}
}
\description{
Calculates an image of linear fit residuals \code{residuals.lm(lm((y~x))}
}
\examples{
a<- terra::rast(system.file("ex/logo.tif", package="terra"))
res<- r_residuals(a[[1]], a[[2]])
res
plot(res)

}
