% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_stretch.R
\name{r_stretch}
\alias{r_stretch}
\title{Stretch raster}
\usage{
r_stretch(x, bit = 16, ...)
}
\arguments{
\item{x}{raster}

\item{bit}{bit rate}

\item{...}{further arguments to `writeRaster`}
}
\description{
Histogram equalization stretch of a raster image. A linear stretch is provided by
`terra::stretch`. Note that \code{terra::stretch} now also provides histogram equalization
via the argument \code{histeq = TRUE}. Function adapted for backwards compatibility
with old code.
}
\examples{

a<- terra::rast(system.file("ex/logo.tif", package="terra"))
res<- r_stretch(a)
res
plot(a*1)
plot(res)
}
