% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_vectorize.R
\name{r_vectorize}
\alias{r_vectorize}
\title{Vectorize}
\usage{
r_vectorize(x)
}
\arguments{
\item{x}{raster layer}
}
\description{
Convert a raster to a vector. Works in memory. Use with caution with big raster.
Note vectorize from the terra should also be enough
}
\examples{
a<- terra::rast(system.file("ex/logo.tif", package="terra"))
res<- r_vectorize(a)
res
plot(res)
}
