% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_vi.R
\name{r_vi}
\alias{r_vi}
\title{Vegetation (Spectral) indices}
\usage{
r_vi(
  x,
  indices = NULL,
  aoi,
  cores = 1,
  writeRaster = FALSE,
  outfolder,
  filename,
  overwrite = FALSE,
  clamp_evi = TRUE,
  wopt = list()
)
}
\arguments{
\item{x}{\code{terra::rast} object with band names set to those
returned by rutils::bandnames(). If missing returns all implemented indices}

\item{indices}{index to calculate. If missing calculates all indices.}

\item{aoi}{sf object. If provided, raster is cropped to this aoi}

\item{cores}{number of cores to use for parallel processing}

\item{writeRaster}{whether to write output to disc. If \code{filename} is
not specified, it will write to the same location
as the source file.  If the source file contains the
 string \code{"10-Band"} it will be replaced with
index name. Otherwise will append the index at the end of
the source file name (e.g. \code{"filename_NDVI.tif"}.}

\item{outfolder}{the folder where to write the file. the filename is taken from x. The same rules
for appending index as above. It will be ignored if \code{filename} is given.}

\item{filename}{filename to write to a specific location. The same rules
for appending index as above. If given overwrites \code{outfolder}}

\item{overwrite}{whether to overwrite if \code{writeRaster = TRUE}}

\item{clamp_evi}{whether or not to clamp EVI to range(-1,1)}

\item{wopt}{list of further options for \code{terra::writeRaster()}}
}
\value{
named \code{list} with \code{rast} objects for each index
}
\description{
Calculate a suite of multispectral indices such as NDVI, SAVI etc.
}
\details{
This function knows which bands it needs to calculate the selected indices
and searches for them in \code{names(x)}. Accepted band names are
\code{"blue", "green", "red", "re1", "re2", "re3", "nir1", "nir2", "swir1", "swir2"}.

If not all bands are available the function will calculate only those indices which are possible
and throw a warning.

Use \code{rutils::bandnames()} with \code{setNames()} to set the names properly.
}
\examples{
# run empty to see which indices are currently implemented
r_vi()

# bulk calculate all possible indices
s <- rast(system.file("ex/logo.tif", package="terra"))
r_vi(s)

# NDVI is not possible
r_vi(s, "NDVI")
TODO:: delete r_vi in changeR and make changeR load this one
}
