% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{.randomString}
\alias{.randomString}
\title{genera stringhe random}
\usage{
.randomString(length, prefix)
}
\arguments{
\item{length}{lunghezza della stringa (default = 8)}

\item{prefix}{evantuale prefisso (default = "")}
}
\value{
stringa con caratteri random di lunghezza = `length` ed un prefisso = `prefix`
}
\description{
genera stringhe random
}

