% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{raw_data_1}
\alias{raw_data_1}
\title{Get paths to \code{insert_package_name} raw data}
\usage{
raw_data_1(file = NULL, package = "insert_package_name")
}
\arguments{
\item{file}{(optional) a \code{character} object indicating the raw data file
name(s). If \code{NULL}, all raw data file names will be returned (default:
\code{NULL}).}

\item{package}{(optional) a string indicating the target package. If not
assigned, the function will try to use the name of the active project
directory (requires the \code{rstudioapi} package).}
}
\value{
If \code{file == NULL}, a \code{\link[=character]{character}} object with all file
names available. Else, a string with the file name path.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{raw_data_1()} returns the raw data paths of the \code{insert_package_name}
package.
}
\examples{
\dontrun{
## To list all raw data file names available

raw_data_1()

## To get the file path from a specific raw data

raw_data_1(raw_data()[1])}
}
\seealso{
Other R system functions: 
\code{\link{raw_data_2}()},
\code{\link{update_pkg_versions}()}
}
\concept{R system functions}
