% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{raw_data_2}
\alias{raw_data_2}
\title{Get paths to \verb{[INSERT PACKAGE NAME]} raw data}
\usage{
raw_data_2(type = NULL, file = NULL, package = NULL)
}
\arguments{
\item{type}{(Optional) A \code{\link[base:character]{character}} string indicating
the file type of the raw data (Default: \code{NULL}).}

\item{file}{(Optional) A \code{\link[base:character]{character}} vector indicating
the file name(s) of the raw data (Default: \code{NULL}).}

\item{package}{(Optional) A \code{\link[base:character]{character}} string indicating
the package with the database data. If \code{NULL}, the function will try to
use the basename of the working directory (Default: \code{NULL}).}
}
\value{
\itemize{
\item If \code{type = NULL}, a \code{character} vector with all file type names
available.
\item If \code{type != NULL} and \code{file = NULL}, a \code{character} vector with all file
names available from type.
\item If \code{type != NULL} and \code{file != NULL}, a \code{character} vector with the file
name(s) path.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{raw_data_2()} returns the raw data paths of the \verb{[INSERT PACKAGE NAME]}
package.
}
\examples{
if (interactive()) {
  raw_data_2()
}
}
\seealso{
Other R package functions: 
\code{\link{raw_data_1}()},
\code{\link{update_pkg_versions}()},
\code{\link{update_pkg_year}()}
}
\concept{R package functions}
\keyword{internal}
