% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_behavioural_results}
\alias{read_behavioural_results}
\title{Import all behavioural data}
\usage{
read_behavioural_results(behavioural_results_dir)
}
\arguments{
\item{behavioural_results_dir}{The directory that contains the _results.json behavioral data files.}
}
\value{
A date frame that concatenates the data frames from all subjects
}
\description{
Read in each behavioural results json file that is in the results directory.
Concatenate the resulting data frames. Create a new "subject" data frame with
values s1, s2, ... sn where s1 is the first participant by date, s2 is the
second, and so on.
}
\examples{
\dontrun{
data_df <- read_behavioural_results("foo_results_dir")
}
}
