% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_sheets.R
\name{read_excel_sheets}
\alias{read_excel_sheets}
\title{Read Excel to list}
\usage{
read_excel_sheets(path, sheets = NULL)
}
\arguments{
\item{path}{character. path to \code{.xlsx} file}

\item{sheets}{either numeric vector representing sheet number, or character vector.
If \code{NULL} it will read all sheets (default)
giving the names of the sheets}
}
\description{
Reads entire excel or selected sheets into a list.
A for-loop Wrapper around \code{readxl::read_excel()} to recursively read all or selected sheets.
always returns a list. so if you want to load only one sheet you might be better off using \code{read_excel()}
}
\details{
Names of sheets are used to name the list elements.
}
