% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{Read Files in Various Formats}
\usage{
read_files(.path)
}
\arguments{
\item{.path}{A character string representing the path to the file to be read.}
}
\value{
A tibble containing the data from the file.
}
\description{
This function reads a file from the specified path and converts it into a tibble.
Supported file formats include RDS, FST, XLSX, XLS, CSV, DTA, Parquet, and Feather.
}
\examples{
\dontrun{
data <- read_files("data/sample.csv")
data <- read_files("data/sample.rds")
}
}
