% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{read_last_file}
\alias{read_last_file}
\title{Read the last file in the directory by the file's extension}
\usage{
read_last_file(dir, file_regex = NULL)
}
\arguments{
\item{dir}{Directory}

\item{file_regex}{File regex to search}
}
\value{
data.frame
}
\description{
Read the last file in the directory by the file's extension
}
