% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_results_json}
\alias{read_results_json}
\title{Read the behavioural results json file}
\usage{
read_results_json(results_json_file)
}
\arguments{
\item{results_json_file}{The path to the results json file.}
}
\value{
A tibble data frame with one row per experimental trial and columns
for participant ID, block number, stimulus type, stimulus details, reaction
time, accuracy, etc.
}
\description{
The results from the Psychopy Python experiment are exported as a json file.
This command converts the main information contained in that json file into a
standard R data frame with, for example, one row per trial.
}
\examples{
\dontrun{
data_df <- read_results_json("foo_results.json")
}
}
