\name{recover_once}
\alias{recover_once}
\title{allows to recover the error once! just call recover_once() and the
error will recover the context. The next time, the error is back to normal
and the context is not recovered}
\usage{
  recover_once()
}
\description{
  allows to recover the error once! just call
  recover_once() and the error will recover the context.
  The next time, the error is back to normal and the
  context is not recovered
}

