% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_go_redundancy.R
\name{reduce_go_redundancy}
\alias{reduce_go_redundancy}
\title{Reduce redundancy of human GO terms}
\usage{
reduce_go_redundancy(pathway_df, threshold, scores = NULL)
}
\arguments{
\item{pathway_df}{a \code{data.frame} or \link[tibble:tbl_df-class]{tibble} object,
with the following columns:
\itemize{
\item \code{go_type}: the sub-ontology the
GO term relates to. Should be one of \code{c("BP", "CC", "MF")}.
\item \code{go_id}: the gene ontology identifier (e.g. GO:0016209)
}}

\item{threshold}{\code{numeric()} vector. Similarity threshold (0-1) for
\code{rrvgo::\link[rrvgo:reduceSimMatrix]{reduceSimMatrix()}}. Some
guidance: Large (allowed similarity=0.9), Medium (0.7), Small (0.5), Tiny
(0.4) Defaults to Medium (0.7).}

\item{scores}{\strong{named} vector, with scores (weights) assigned to each
term. Higher is better. Can be NULL (default, means no scores. In this case,
a default score based on set size is assigned, thus favoring larger sets).
Note: if you have p-values as scores, consider log-transforming them
(\code{-log10(p)}).}
}
\value{
a \link[tibble:tbl_df-class]{tibble} object of pathway results, a
"reduced" parent term to which pathways have been assigned. New columns:
\itemize{
\item \code{parent_id}: the GO ID of the parent term
\item \code{parent_term}: a description of the GO ID
\item \code{parent_sim_score}: the similarity score between the child GO term and
its parent term
}
}
\description{
This function will reduce GO redundancy first by creating a
semantic similarity matrix (using
\code{GOSemSim::\link[GOSemSim:mgoSim]{mgoSim}}), which is then passed
through \code{rrvgo::\link[rrvgo:reduceSimMatrix]{reduceSimMatrix()}},
which will reduce a set of GO terms based on their semantic similarity and
scores (in this case, a default score based on set size is assigned.)
}
\details{
Semantic similarity is calculated using the "Wang" method, a
graph-based strategy to compute semantic similarity using the topology of
the GO graph structure. \code{GOSemSim::\link[GOSemSim:mgoSim]{mgoSim}} does
permit use of other measures (primarily information-content measures), but
"Wang" is used as the default in GOSemSim (and was, thus, used as the
default here).

\code{rrvgo::\link[rrvgo:reduceSimMatrix]{reduceSimMatrix()}} creates a
distance matrix, defined as (1-simMatrix). The terms are then hierarchically
clustered using complete linkage (an agglomerative, or "bottom-up"
clustering approach), and the tree is cut at the desired threshold. The term
with the highest "score" is used to represent each group.
}
\examples{
file_path <- 
  system.file(
    "testdata", 
    "go_test_data.txt", 
    package = "rutils", 
    mustWork = TRUE
  )
  
pathway_df <- 
  readr::read_delim(file_path,
                    delim = "\t")

reduce_go_redundancy(
  pathway_df = pathway_df,
  threshold = 0.9,
  scores = NULL)
  
}
\references{
\itemize{
\item Yu et al. (2010) GOSemSim: an R package for measuring semantic
similarity among GO terms and gene products \emph{Bioinformatics} (Oxford,
England), 26:7 976--978, April 2010.
\url{http://bioinformatics.oxfordjournals.org/cgi/content/abstract/26/7/976}
PMID: 20179076
\item Sayols S (2020). rrvgo: a Bioconductor package to
reduce and visualize Gene Ontology terms. \url{https://ssayols.github.io/rrvgo}
}
}
