% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeHome.R
\name{refFile}
\alias{refFile}
\title{provide wrapper to path to reference data files}
\usage{
refFile(fileName)
}
\arguments{
\item{fileName}{of expected file in path}
}
\description{
provide wrapper to path to reference data files
}
