\name{renameany}
\alias{renameany}
\title{rename anything in the table (columns or cells)
you can specify a subset of columns to look at
if the selector is ambiguous.
note that you can have circular renaming, because the data is copied first}
\usage{
  renameany(d, rules, cols = c())
}
\description{
  rename anything in the table (columns or cells) you can
  specify a subset of columns to look at if the selector is
  ambiguous. note that you can have circular renaming,
  because the data is copied first
}

