% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{repeating_column}
\alias{repeating_column}
\title{Check if all values in a column are the same}
\usage{
repeating_column(x, roundx = NA)
}
\arguments{
\item{x}{Vector of values}

\item{roundx}{Only use if data is numeric, rounds the calculated variance if a value is given}
}
\value{
Logical
}
\description{
Checks if all values in a column are the same, works for any type of data
}
\details{
Be careful using the \code{roundx} argument. e.g. repeating_column(c(1, 1, 1.0001), roundx = 8) will
return TRUE.
}
\examples{
repeating_column(rep(1, 5))

repeating_column(c("dog", "dog", "dog", "cat"))

# Use with purrr to test multiple columns

testdf <- data.frame(x = rep(1, 5), y = c(1, 2, 1, 1, 1))
purrr::map_lgl(testdf, ~ repeating_column(.))
}
