% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_10band.R
\name{replace_10band}
\alias{replace_10band}
\title{Replace 10-Band string in S2-filename}
\usage{
replace_10band(path, replacement)
}
\arguments{
\item{replacement}{}
}
\description{
Shorthand for `str_replace(s2_path, "10-Band", "replacement")`.
It is useful for constructing filenames for `writeRaster`
}
