% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{rolling_file}
\alias{rolling_file}
\title{Rolls the file based on parameters}
\usage{
rolling_file(filename, max_size, max_files, lock_file)
}
\description{
Rolling of a file is based on the following conditions
 - file is bigger than a given size
 - number of rolling files are bigger than
}
