% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{round_clever}
\alias{round_clever}
\title{Clever rounding function}
\usage{
round_clever(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
a numeric vector rounded accordingly
}
\description{
This function tries to guess what is the best way to round a value to the closest meaningful value.
}
\examples{
round_clever(c(123456789, 12345.6789, 1.2346789))
}
\keyword{round}
