% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_parallel_tasks}
\alias{run_parallel_tasks}
\title{Run tasks in parallel and collect results}
\usage{
run_parallel_tasks(
  task_description,
  inputs,
  func_to_run,
  poll_interval = 1,
  ...
)
}
\arguments{
\item{task_description}{A description of the task.}

\item{inputs}{A list of inputs for the parallel tasks.}

\item{func_to_run}{The function to execute for each input.}

\item{poll_interval}{Interval in seconds between polling for task completion.}

\item{...}{Additional arguments passed to the function.}
}
\value{
A list of results keyed by the input items.
}
\description{
Run tasks in parallel and collect results
}
