% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runWithCpuParallelBackend.R
\name{runWithCpuParallelBackend}
\alias{runWithCpuParallelBackend}
\title{CPU Parallel Processing}
\usage{
runWithCpuParallelBackend(taskToBeExecuted, nCores = NULL, ...)
}
\arguments{
\item{taskToBeExecuted}{a function representing a task to be executed
with parallel processing.}

\item{nCores}{the numbers of CPU cores to be used in the parallel
processing. By default, it uses the numbers of cores returned by the
\code{\link[parallel]{detectCores}} function. It must be an integer.}

\item{...}{additional paramaters to be delegated to \code{taskToBeExecuted}
function.}
}
\description{
Helper function which automatically configures a CPU parallel
backend either on Linux or Windows (It was not tested in MAC OS).
It is specially useful when there is a requirement to write R
code that should be portable between windows and linux.
}
\details{
This function saves in global environment a boolean variable
called \code{activeParallelBackend}, indicating whether or not
the parallel backend is available for use.
}
\examples{
library(foreach)

# a function simulating an useful parallel task
summation <- function(){
   # saves the result in the parent enviroment
   result <<- sum(foreach(i = 1:10000, .combine = "c") \%dopar\% {sqrt(i)})
}

# this code is portable between windows and linux, if the right packages
# are installed in both of them.
runWithCpuParallelBackend(summation)
print(result) ## 666716.5
}

