% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runWithLinuxCpuParallelBackend.R
\name{runWithLinuxCpuParallelBackend}
\alias{runWithLinuxCpuParallelBackend}
\title{CPU Parallel Processing in Linux}
\usage{
runWithLinuxCpuParallelBackend(taskToBeExecuted, nCores = NULL, ...)
}
\arguments{
\item{taskToBeExecuted}{a function representing a task to be executed
with parallel processing.}

\item{nCores}{the numbers of CPU cores to be used in the parallel
processing. By default, it uses the numbers of cores returned by the
\code{\link[parallel]{detectCores}} function. It must be an integer.}

\item{...}{additional paramaters to be delegated to \code{taskToBeExecuted}
function.}
}
\description{
Utility function which executes a task with automatic configured linux parallel
backend. Under the hoods, it uses the \code{\link[doMC]{registerDoMC}}
function from the \code{doMC} package.
}
\details{
This function also saves in the global environment a boolean variable called
\code{activeParallelBackend} indicating if the parallel backend is
available to use.
}
\examples{
library(foreach)

# a function simulating an useful parallel task
summation <- function(){
   # saves the result in the parent enviroment
   result <<- sum(foreach(i = 1:10000, .combine = "c") \%dopar\% {sqrt(i)})
}

runWithLinuxCpuParallelBackend(summation)
print(result) ## 666716.5
}

