% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils-package.R
\docType{package}
\name{rutils-package}
\alias{rutils}
\alias{rutils-package}
\title{rutils: Utilties for Analysing Data from Project Purple}
\description{
Utilities for reading, processing, exploring, and modelling EEG and behavioural data from Project Purple. Project Purple is an EEG based experimental investigation of the human Approximate Number System (ANS).
}
\author{
\strong{Maintainer}: Mark Andrews \email{mark.andrews@ntu.ac.uk} (\href{https://orcid.org/0000-0001-7499-9521}{ORCID})

}
\keyword{internal}
