% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rutils-package.r
\docType{package}
\name{rutils}
\alias{rutils}
\alias{rutils-package}
\title{Utility functions for R}
\description{
Utility functions for R
}
\author{
Thibaut Lamadon \email{thibaut.lamadon@gmail.com}
}
\references{
\url{https://github.com/tlamadon/rutils}
}
\seealso{
rutils
}

