% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_ingest_cloudmask.R
\name{s2_ingest_cloudmask}
\alias{s2_ingest_cloudmask}
\title{Ingest Sentinel-2 Cloud Mask from .SAFE data}
\usage{
s2_ingest_cloudmask(path, outfolder, return = TRUE, ...)
}
\arguments{
\item{path}{path to an unzipped .SAFE folder}

\item{outfolder}{the folder where to write the file. the filename is taken from the .SAFE file}

\item{return}{whether to return a \code{terra::rast} object back into the session - default.
Otherwise returns the path to the newly created \code{tif}}

\item{...}{further parameters to \code{writerRaster}}
}
\description{
This function loads and saves the cloud mask layer available in the .SAFE folder containing
Sentinel-2 scene. The bands area resampled to 10m spatial resolution.
}
\details{
Currently works only for Level-2A products.
}
\examples{
# .SAFE folder
path <- path_daten9(content = T, regexp = "creodias/.*SAFE$" ,recurse = T, fail = F)[[1]]

# set outfolder and create if need
outfolder <- "data/cloudmask"
rutils::path_exists(outfolder, create = TRUE)


r <- s2_ingest_cloudmask(path = path, outfolder = outfolder)
}
