% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{saf}
\alias{saf}
\title{Summary Function}
\usage{
saf(vector)
}
\arguments{
\item{vector}{vector you want to summarize}
}
\description{
This function is a shortcut for `summary(as.factor(x))`
}
\examples{
saf(sample(c(TRUE,FALSE), 100, rep=TRUE))
}
\keyword{summary}
