% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{safr}
\alias{safr}
\title{Relative Summary Function}
\usage{
safr(vector, pretty = c(0, 1, 2)[1])
}
\arguments{
\item{vector}{vector you want to summarize}

\item{pretty}{how pretty should the result be? `0` will be a raw 0-1 value, `2` will be a 0-100 character string with percentage symbol.}
}
\description{
This function is a shortcut for `summary(as.factor(x))`
}
\examples{
safr(sample(c(TRUE,FALSE), 100, rep=TRUE), p=1)
}
\keyword{summary}
