% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sample_df}
\alias{sample_df}
\title{Get random samples from a dataframe, without replacement}
\usage{
sample_df(df, n)
}
\arguments{
\item{df}{Dataframe from which to sample}

\item{n}{Integer value of number of samples to take}
}
\value{
Dataframe with sampled rows
}
\description{
Takes \code{n} random sample from the dataframe \code{df} without replacement
}
\examples{
set.seed(1)

test <- data.frame(x = seq(1, 10, 1), y = rnorm(10))
test_sample <- sample_df(test, 5)
}
