% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_png.R
\name{save_png}
\alias{save_png}
\title{Save a plot as a .png}
\usage{
save_png(plot_code, path, width = 8, height = 6, units = "in", res = 600, ...)
}
\arguments{
\item{plot_code}{code required to generate the plot, sandwiched in curly
braces.}

\item{path}{\code{character(1)} specifying path for where to save your plot,
including file name.}

\item{width}{the width of the device.}

\item{height}{the height of the device.}

\item{units}{The units in which \code{height} and \code{width} are
    given.  Can be \code{px} (pixels, the default), \code{in} (inches),
    \code{cm} or \code{mm}.}

\item{res}{The nominal resolution in ppi which will be recorded in the
    bitmap file, if a positive integer.  Also used for \code{units}
    other than the default, and to convert points to pixels.}

\item{...}{additional arguments are passed to \code{grDevices::png()}.}
}
\value{
Invisibly, the path to saved plot.
}
\description{
A wrapper for calling \code{grDevices::png()} then \code{grDevices::dev.off()}. Allows
users to save a plot to a specified path. The preferred method for
\code{ggplot2::ggplot} objects remains \code{ggplot::ggsave()}. However if you wish to
use this function to save \code{ggplot2::ggplot} objects, print the plot as the
last step of \code{plot_code}.
}
\examples{

save_png(
  {
    plot(x = 1:5)
  },
  path = file.path(tempdir(), "test_plot.png"),
)
}
