% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_htmlwidget.R
\name{save_widget}
\alias{save_widget}
\title{Save htmlwidgets to an html file}
\usage{
save_widget(
  widget,
  path,
  title = class(widget)[[1]],
  type = "direct",
  ...,
  libdir = NULL,
  libpath = "relative"
)
}
\arguments{
\item{widget}{htmlwidget}

\item{path}{Where to save HTML file}

\item{title}{Title of HTML page}

\item{type}{One of \code{"direct"}, \code{"lib"}, or \code{"intermediate"}}

\item{libdir}{Directory for HTML dependencies (if \code{type} is \code{"lib"})}

\item{libpath}{Specify \code{"relative"} or \code{"absolute"} HTML dependency paths
(if \code{type} is \code{"lib"})}
}
\description{
Convenience wrapper around \link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget}
}
\details{
There are three ways to save an htmlwidget. If \code{type} is \code{"direct"},
a self-contained HTML file is saved as usual. \code{save_widget} helps in this case
due to an apparent bug in \code{saveWidget} if attempting to save to a directory
other than the current working directory.

If \code{type} is \code{"lib"}, a non-self-contained HTML file is created. In this
case, \code{"libdir"} may be provided. Also, \code{"libpath"} may be provided as either
\code{"relative"} (default) or \code{"absolute"}, which specifies whether paths
in the html file to the library files are relative or absolute. In contrast to
previous function \code{save_leaflet}, \code{"libdir"} does not necessarily need
to be a child of the directory containing \code{path}.

If \code{type} is \code{"intermediate"}, a self-contained HTML file is generated, but
through an intermediate step of first creating a non-self-contained HTML file, then
producing the self-contained file from this by writing a new HTML after copying the
HTML dependencies into the HTML file. This can be useful when saving large htmlwidgets,
as the attempt to save as a self-contained file using \code{saveWidget} will
fail on a pandoc memory error. This approach enables the saving of larger
widgets as self-contained HTML files.
}
