% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveChart.R
\name{saveChart}
\alias{saveChart}
\title{save a chart in a standard way}
\usage{
saveChart(fileName = "Chart", chart, theDate = NULL, onlyLatest = TRUE)
}
\arguments{
\item{fileName}{Leading part of file name}

\item{chart}{a ggplot2 chart}

\item{theDate}{data of chart - defaults to today()}

\item{onlyLatest}{ignores the data parameter and overwrites chart}
}
\description{
save a chart in a standard way
}
