% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoadFunctions.R
\name{saveModel}
\alias{saveModel}
\title{Save Model}
\usage{
saveModel(..., name = stop("'name' must be specified"), overwrite = FALSE,
  save.fun = save)
}
\arguments{
\item{...}{R object to be saved}

\item{name}{filename of the object}

\item{overwrite}{if file exists, overwrites the model if true otherwise stops}

\item{save.fun}{save function, defaults to save}
}
\description{
\code{saveModel} saves an Robject to a modelpath specified in user's Rprofile with filename name
}
