% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormap.R
\name{select_colors}
\alias{select_colors}
\title{Select Color Palette}
\usage{
select_colors(palette)
}
\arguments{
\item{palette}{A character string indicating the color palette to use.
Possible values are "col29", "col28", "col_sim", "col_sex", and "col_33".}
}
\value{
A vector of colors corresponding to the selected palette.
}
\description{
This function selects a color palette based on the specified palette name.
}
