% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_en_us_locale.R
\name{set_en_us_locale}
\alias{set_en_us_locale}
\title{Set the system locale to English (United States)}
\usage{
set_en_us_locale()
}
\value{
An invisible \code{NULL}. This function is called for its side effects.
}
\description{
\code{set_en_us_locale()} sets the system locale to English (United States)
trying to overcome OS-dependent differences in locale handling.
}
\examples{
get_locale()

set_en_us_locale()

get_locale()
}
\seealso{
Other system functions.: 
\code{\link{get_locale}()}
}
\concept{system functions.}
