% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.r
\name{setOption}
\alias{setOption}
\title{Handling of options in R sucks chode big time.}
\usage{
setOption(key, value)
}
\arguments{
\item{key}{key for the option}

\item{value}{value for the option}
}
\description{
I'm gonna let parametrize the key for options
}

