% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_setup.R
\name{setup_Rprofile}
\alias{setup_Rprofile}
\alias{setup_gitconfig}
\alias{setup_gitignore_global}
\title{Set up a default ~/.Rprofile}
\usage{
setup_Rprofile(
  path = "~/.Rprofile",
  template = "template_Rprofile",
  append = TRUE
)

setup_gitconfig(
  path = "~/.gitconfig",
  template = "template_gitconfig",
  append = FALSE
)

setup_gitignore_global(
  path = "~/.gitignore_global",
  template = "template_gitignore_global",
  append = FALSE
)
}
\arguments{
\item{path}{\code{character(1)} path to the configuration file.}

\item{template}{\code{character(1)} name of the template configuration files
stored in \code{inst/extdata}}

\item{append}{\code{logical(1)} whether to append to or overwrite the \code{path} with
the \code{template}.}
}
\description{
Set of functions to setup default R and git configuration files
}
\details{
These are used to setup default Rprofile/gitconfig/gitignore_global
configurations. They may be useful if you consistently need to set up new
instances e.g. via \code{docker_run_rserver}.
}
\section{Functions}{
\itemize{
\item \code{setup_gitconfig}: Set up a default ~/gitconfig

\item \code{setup_gitignore_global}: Set up a default ~/.gitconfig
}}

\examples{

\dontrun{
setup_Rprofile()
setup_gitconfig()
setup_gitignore_global()
}
}
