% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{should_roll_file}
\alias{should_roll_file}
\title{predicate that tells if a file needs to be "rolled"}
\usage{
should_roll_file(filename, max_size)
}
\arguments{
\item{filename}{path to be checked}

\item{max_size}{size beyond that filename has to be rolled}
}
\description{
predicate that tells if a file needs to be "rolled"
}
