\name{showMemoryUse}
\alias{showMemoryUse}
\title{gets all obejcts with their sizes
shamelessly copied from http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
\usage{
  showMemoryUse(sort = "size", decreasing = FALSE, limit)
}
\description{
  gets all obejcts with their sizes shamelessly copied from
  http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
}

