% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shush.R
\name{shush}
\alias{shush}
\title{Suppress messages and warnings}
\usage{
shush(x, quiet = TRUE)
}
\arguments{
\item{x}{Any expression, usually a function call.}

\item{quiet}{(optional) A \link[base:logical]{logical} flag value indicating
whether to suppress messages and warnings. This is can be used for
condition messages and warnings inside functions (Default: \code{TRUE}).}
}
\value{
The same object as \code{x} with messages and warnings suppressed.
}
\description{
\code{shush()} is a wrapper around \code{suppressMessages()} and \code{suppressWarnings()}
that allows you to suppress messages and warnings in a single function call.
It was designed to be used with pipes.
}
\examples{
message("test") |> shush()
message("test") |> shush(quiet = FALSE)
#> test # Expected

warning("test") |> shush()
warning("test") |> shush(quiet = FALSE)
#> Warning: test # Expected
}
\concept{dialog functions.}
