% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signi.R
\name{signi}
\alias{signi}
\title{Produce produces significance symbols for the values of p (ns., *, **, ***)}
\usage{
signi(p)
}
\arguments{
\item{p}{a p value}
}
\value{
a character "ns", "\emph{", "\strong{", "}}"
}
\description{
Produce produces significance symbols for the values of p (ns., *, **, ***)
}
\examples{
signi(p=0.004)

}
