% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{slice}
\alias{slice}
\title{Slides the \code{x} object in \code{n} parts}
\usage{
slice(x, n)
}
\arguments{
\item{x}{the object to be sliced}

\item{n}{the number of parts}
}
\value{
the list containing the \code{n} parts
}
\description{
Slides the \code{x} object in \code{n} parts
}
\author{
Giuseppe Acito
}

