% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{slug}
\alias{slug}
\title{Slug}
\usage{
slug(x, sep = "-")
}
\arguments{
\item{x}{input character string}

\item{sep}{separator}
}
\value{
the same character vector in its normal form
}
\description{
Reduces the input character string to a normal form.
This can be useful to match user-input names if you don't feel like doing any stemming/lemming.
}
\examples{
slug("La magie d'Aladin")
}
\keyword{form}
\keyword{normal}
\keyword{slug}
