% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf_smooth_data.R
\name{smooth_byRow}
\alias{smooth_byRow}
\title{Smooth Data by Row}
\usage{
smooth_byRow(
  data,
  method = c("moving_average", "loess"),
  window_size = 3,
  span = 0.2
)
}
\arguments{
\item{data}{A matrix or data frame where smoothing is to be applied.}

\item{method}{Method of smoothing: either "moving_average" or "loess".}

\item{window_size}{Size of the moving average window (only applicable if method is "moving_average").}

\item{span}{Span parameter for loess (only applicable if method is "loess").}
}
\value{
A matrix of smoothed data.
}
\description{
This function smooths the data by row using either moving average or loess smoothing.
}
