% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/source.all.files.R
\name{source.all.files}
\alias{source.all.files}
\title{Source Multiple Files}
\usage{
source.all.files(path = NULL)
}
\arguments{
\item{path}{a character vector of path names;
the default corresponds to the working directory,
 \code{getwd()}}
}
\description{
Source multiples files whose extensions are .R and .rda in the
paths passed as parameter.
}
\examples{
source.all.files(file.path("dir1", "dir2"))
source.all.files()
}

