% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sourceAllFiles.R
\name{sourceAllFiles}
\alias{sourceAllFiles}
\title{Source Multiple Files}
\usage{
sourceAllFiles(path = NULL)
}
\arguments{
\item{path}{a character vector of path names;
the default corresponds to the working directory,
 \code{getwd()}}
}
\description{
Source multiples files whose extensions are .R and .rda in the
paths passed as parameter.
}
\examples{
sourceAllFiles(file.path("dir1", "dir2"))
sourceAllFiles()
}

