% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA_utils.R
\name{sphere_cutoff}
\alias{sphere_cutoff}
\title{Removes everything within a sphere
of the defined quantile of the vector norm.}
\usage{
sphere_cutoff(x, qcutoff = 0.8)
}
\arguments{
\item{x}{matrix of row vectors}

\item{qcutoff}{quantile.}
}
\value{
Matrix of the vectors longer than the defined cutoff.
}
\description{
Removes everything within a sphere
of the defined quantile of the vector norm.
}
