% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inc.matlab.r
\name{spread}
\alias{spread}
\title{spread an array along a given dimension (similar to repmat in Matlab or spread in Fortran)
what dim specifies in the order given
spread(A , c(2) , 10) will insert a dimension  in the second index of size 10}
\usage{
spread(A, loc, dims)
}
\description{
spread an array along a given dimension (similar to repmat in Matlab or spread in Fortran)
what dim specifies in the order given
spread(A , c(2) , 10) will insert a dimension  in the second index of size 10
}

