% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{spread_multi}
\alias{spread_multi}
\title{Spread multiple columns from long to wide format}
\usage{
spread_multi(data, key_col, ..., fill = NA, convert = TRUE, drop = TRUE,
  sep = "_")
}
\description{
Spread multiple columns from long to wide format
}
