% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bbox_to_sfc.R
\name{st_bbox_as_sfc}
\alias{st_bbox_as_sfc}
\title{Spatial object bbox as sfc}
\usage{
st_bbox_as_sfc(x)
}
\arguments{
\item{x}{bbox or object from which bbox can be calculated. See \code{st_bbox()}}
}
\value{
sfc
}
\description{
Extracts extent from a spatial object and convert to sfc
}
