% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_bbox_transform.R
\name{st_bbox_transform}
\alias{st_bbox_transform}
\title{Transform bounding box}
\usage{
st_bbox_transform(x, crs)
}
\arguments{
\item{x}{bbox or an object from which a bbox can be extracted}

\item{crs}{the target crs}
}
\value{
bbox object
}
\description{
Reproject the bbox (extend) of a spatial object
}
\seealso{
st_bbox_to_sfc
}
