% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_country.R
\name{st_country}
\alias{st_country}
\title{Country geometry}
\usage{
st_country(country)
}
\arguments{
\item{country}{\code{character} the name of the country}
}
\value{
an \code{sf} object
}
\description{
Queries the \code{rutils::world} dataset and returns a country of choice.
}
\details{
TODO: make it work with many countries

English name of countries required. e.g. one of \code{print(rutils::world$name)}
}
