% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_crs_equal.R
\name{st_crs_equal}
\alias{st_crs_equal}
\title{Are CRS equal}
\usage{
st_crs_equal(x, y)
}
\arguments{
\item{x}{spatial object on which `st_crs()` can be called}

\item{y}{spatial object on which `st_crs()` can be called}
}
\value{
Logical
}
\description{
Checks whether two spatial objects have the same `crs`
}
