% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_from_wkt.R
\name{st_from_wkt}
\alias{st_from_wkt}
\title{Simple Feature from WKT}
\usage{
st_from_wkt(x, crs)
}
\arguments{
\item{x}{\code{character}. Geometry as WKT}
}
\value{
\code{sfc}
}
\description{
Create a simple feature from a well known text
}
\examples{

string <- "MULTIPOLYGON (((-19.8394 18.1518, -19.1212 20.7523,  -24.6239 30.3655, -25.1478 30.4619,  -30.9179 20.4919, -30.4307 20.4011, -28.0359 19.9526,  -20.9466 18.4181, -20.4804 18.3004, -20.0118 18.1969, -19.8394 18.1518)))"

st_from_wkt(string, crs = 4326)
}
