% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_gdb_pull.R
\name{st_gdb_pull}
\alias{st_gdb_pull}
\title{Pull a layer from GDB}
\usage{
st_gdb_pull(gdb, layer, aoi)
}
\arguments{
\item{gdb}{Geodatabase (path)}

\item{layer}{Layer name (character)}

\item{aoi}{Either a path to or an sf object for the ptional extent to crop to. (reprojected to \code{st_crs(gdb)})}
}
\description{
Pull a layer from GDB optionally cropping it to an extend. In fact,
if no cropping extend is provided this function is the same as
\code{sf::st_read(gdb, layer)}
}
