% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grow_shape.R
\name{st_grow_shape}
\alias{st_grow_shape}
\title{Expand/Grow a polygon shape}
\usage{
st_grow_shape(x, buffer)
}
\arguments{
\item{x}{object of class \code{sf} polygon}

\item{buffer}{the distance to grow}
}
\value{
x object with the last polygon being the expanded geometry
}
\description{
This function is used to expand the shapefile beyond its boundaries.
This is useful when a layer needs to be slightly bigger than the AOI to which
it has been cut.
}
\details{
Be careful with the how much you grow the area. The bigger the buffer the bigger the error.
works good with 10-50 m and can go up to 200m if polygons are big enough.
}
\examples{
a<- terra::rast(system.file("ex/logo.tif", package="terra"))
x<- st_as_sf(terra::ext(30,50,30,50) \%>\% terra::vect())
res <- st_grow_shape(x, 10)
plot(a[[1]])
plot(res, add = TRUE)
# the inner polygon is the original polygon,
# the outer polygon was generated
}
