% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_remove_duplicates.R
\name{st_remove_duplicates}
\alias{st_remove_duplicates}
\title{Remove duplicates}
\usage{
st_remove_duplicates(sf)
}
\arguments{
\item{sf}{a spatial object of class `sf`}
}
\description{
Removes duplicate geometries
}
