% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_rename_geometry.R
\name{st_rename_geometry}
\alias{st_rename_geometry}
\title{Rename Geometry Column}
\usage{
st_rename_geometry(g, name)
}
\arguments{
\item{g}{the spatial object `sf`}

\item{name}{the new name to be given}
}
\description{
give a new name to a `geometry` column in `sf, sfc`
}
