% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.r
\name{Stack}
\alias{Stack}
\title{Stack data structure}
\description{
This is a stack data structure, operations you can do with it are:
}
\details{
\itemize{
\item push
\item pop
\item peek
}

Constructor for Stack

Push an item in the stack

Peek an item from the stack

Return the item on top of the stack. Raises an error if the
stack is empty

Pops an item from the stack

returns and removes the item on top of the stack. If the stack
is empty, it raises an error.

\code{TRUE} is stack is empty

Returns the length of the stack

Internal method for object rappresentatio on REPL
}
\examples{
\dontrun{
 s <- Stack()
 s$push(1)
 s$peek()
 s$pop()
}

## ------------------------------------------------
## Method `Stack$new`
## ------------------------------------------------

Stack$new("first")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Stack-new}{\code{Stack$new()}}
\item \href{#method-Stack-push}{\code{Stack$push()}}
\item \href{#method-Stack-peek}{\code{Stack$peek()}}
\item \href{#method-Stack-pop}{\code{Stack$pop()}}
\item \href{#method-Stack-empty}{\code{Stack$empty()}}
\item \href{#method-Stack-length}{\code{Stack$length()}}
\item \href{#method-Stack-show}{\code{Stack$show()}}
\item \href{#method-Stack-clone}{\code{Stack$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stack-new"></a>}}
\if{latex}{\out{\hypertarget{method-Stack-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stack$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{items to be pushed in the stack}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Stack$new("first")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stack-push"></a>}}
\if{latex}{\out{\hypertarget{method-Stack-push}{}}}
\subsection{Method \code{push()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stack$push(item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item}}{object to be pushed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stack-peek"></a>}}
\if{latex}{\out{\hypertarget{method-Stack-peek}{}}}
\subsection{Method \code{peek()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stack$peek()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
element on top of the stack
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stack-pop"></a>}}
\if{latex}{\out{\hypertarget{method-Stack-pop}{}}}
\subsection{Method \code{pop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stack$pop()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
element on top of the stack
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stack-empty"></a>}}
\if{latex}{\out{\hypertarget{method-Stack-empty}{}}}
\subsection{Method \code{empty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stack$empty()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if empty, \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stack-length"></a>}}
\if{latex}{\out{\hypertarget{method-Stack-length}{}}}
\subsection{Method \code{length()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stack$length()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the number of elements in the stack
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stack-show"></a>}}
\if{latex}{\out{\hypertarget{method-Stack-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stack$show(sep = "\\n")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sep}}{separator used on the output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Stack-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Stack-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stack$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
