% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_error.R
\name{std_error}
\alias{std_error}
\title{Compute the standard error of the sample mean}
\usage{
std_error(x)
}
\arguments{
\item{x}{A \code{\link[base:numeric]{numeric}} vector with the sample data.}
}
\value{
A number representing the standard error.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{std_error()} computes the
\href{https://en.wikipedia.org/wiki/Standard_error}{standard error}
of the sample mean.
}
\details{
The \href{https://en.wikipedia.org/wiki/Standard_error}{standard error}
can be estimated by the standard deviation of the sample
(\eqn{\sigma_{\bar{x}}}) divided by the square root of the sample size
(\eqn{\sqrt{n}}).

\deqn{\sigma_{\bar{x}} = \frac{\sigma}{\sqrt{n}} \approx
\frac{s}{\sqrt{n}}}
}
\examples{
1:100 |> std_error() |> round(5)
#> [1] 2.90115 # Expected
}
\seealso{
Other stat functions: 
\code{\link{cut_interval_mean}()}
}
\concept{stat functions}
