% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stopif}
\alias{stopif}
\title{Stop if a condition is TRUE, with a custom error message}
\usage{
stopif(condition, message = "Condition failed")
}
\arguments{
\item{condition}{A logical condition to test.}

\item{message}{A string message to show if the condition is TRUE.}
}
\value{
None. Stops execution with an error if the condition is TRUE.
}
\description{
Stop if a condition is TRUE, with a custom error message
}
