% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_if.R
\name{swap_if}
\alias{swap_if}
\title{Swap two values if a condition is met}
\usage{
swap_if(x, y, condition = TRUE)
}
\arguments{
\item{x}{Any type of R object.}

\item{y}{An R object of the same class and length as \code{x}.}

\item{condition}{A \code{\link[base:logical]{logical}} flag indicating if the
values should be swapped (Default: \code{TRUE}).}
}
\value{
A \code{\link[base:list]{list}} with the swapped values.
}
\description{
\code{swap_if()} swaps two values if a condition is met.
}
\examples{
swap_if(1, 2)
#> $x # Expected
#> [1] 2
#>
#> $y
#> [1] 1

swap_if(1, 2, condition = FALSE)
#> $x # Expected
#> [1] 1
#>
#> $y
#> [1] 2

swap_if(letters, LETTERS, c(rep(TRUE, 13), rep(FALSE, 13)))
#> $x
#>  [1] "A" "B" "C" "D" "E" "F" "G" "H" "I" "J" "K" "L" "M" "n"
#> [15] "o" "p" "q" "r" "s" "t" "u" "v" "w" "x" "y" "z"
#>
#> $y
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j" "k" "l" "m" "N"
#> [15] "O" "P" "Q" "R" "S" "T" "U" "V" "W" "X" "Y" "Z"
}
\seealso{
Other conditional functions: 
\code{\link{vct_replace_all}()}
}
\concept{conditional functions}
