% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_vector.R
\name{swap_vector}
\alias{swap_vector}
\title{Swap values in a vector}
\usage{
swap_vector(x, swap = na_as(x), except = NULL, preserve_nas = FALSE)
}
\arguments{
\item{x}{An \code{\link[checkmate:checkAtomic]{atomic}} vector.}

\item{swap}{(Optional) A single value to swap the values in \code{x}
(Default: \code{na_as(x)}).}

\item{except}{(Optional) A vector of values to exclude from swapping.}

\item{preserve_nas}{(Optional) A \code{\link[base:logical]{logical}} flag to
indicating if the function must preserve \code{NA} values in \code{x}
(Default: \code{FALSE}).}
}
\value{
The same type of object in \code{x} with swapped values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{swap_vector()} swaps values in a vector with a specified value.
}
\examples{
swap_vector(1:10)
#> [1] NA NA NA NA NA NA NA NA NA NA

swap_vector(1:10, swap = 0L, except = 5:7)
#> [1] 0 0 0 0 5 6 7 0 0 0

swap_vector(c("a", "b", NA, "c"), swap = "d", preserve_nas = TRUE)
#> [1] "d" "d" NA  "d"
}
\seealso{
Other conditional functions: 
\code{\link{swap_if}()}
}
\concept{conditional functions}
