% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_rmd_to_blog_dirs.R
\name{sync_rmd_to_blog_dirs}
\alias{sync_rmd_to_blog_dirs}
\title{syncs/copies Rmd files to blog directories
For example: I keep my blog files as hugo+bookdown files inside some directory.
But I keep my Rmd files in some external directories.
By using sync_rmd_to_blog_dirs(".") I can copy all Rmd files in current directory to 
those blog directories with external dependency files included.
Note that, external dependency resources, should be explicitly specified in yaml header
such as:
resource_files:
- ./globals.R}
\usage{
sync_rmd_to_blog_dirs(dir = ".")
}
\arguments{
\item{dir}{source directory for Rmd files}
}
\description{
syncs/copies Rmd files to blog directories
For example: I keep my blog files as hugo+bookdown files inside some directory.
But I keep my Rmd files in some external directories.
By using sync_rmd_to_blog_dirs(".") I can copy all Rmd files in current directory to 
those blog directories with external dependency files included.
Note that, external dependency resources, should be explicitly specified in yaml header
such as:
resource_files:
- ./globals.R
}
\examples{
sync_rmd_to_blog_dirs(".")
}
