% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tabler}
\alias{tabler}
\title{Confusion table}
\usage{
tabler(r, c, scale.by = 1, pretty = 1)
}
\arguments{
\item{r}{numeric or character vector, factor or ordered. Will be displayed as row in result.}

\item{c}{numeric or character vector, factor or ordered, the same length as r. Will be displayed as column in result.}

\item{scale.by}{1 for row, 2 for column}

\item{pretty}{0 for 0-1 numeric, 1 for 0-100 numeric, anything else for simple table.}
}
\value{
a convenient confusion matrix
}
\description{
Pretty and relative confusion table
}
\examples{
tabler(iris$Species, iris$Sepal.Length>5.8, scale=1, p=1)
}
\keyword{confusion}
\keyword{matrix}
\keyword{table}
