% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{tdistr}
\alias{tdistr}
\title{Calculate the density of the non-standard Student's t-distribution.}
\usage{
tdistr(x, dfree = 3, loc = 0, scalev = 1)
}
\arguments{
\item{\code{x}}{A \emph{numeric} value for which to calculate the density of
the t-distribution.}

\item{\code{dfree}}{An \emph{integer} value equal to the degrees of freedom
(the default is \code{3}).}

\item{\code{loc}}{A \emph{numeric} value equal to the location (center) of
the distribution (the default is \code{0}).}

\item{\code{scalev}}{A \emph{numeric} value equal to the scale (width) of the
distribution (the default is \code{1}).}
}
\value{
A \emph{numeric} value equal to the density of the non-standard
  Student's t-distribution.
}
\description{
Calculate the density of the non-standard Student's t-distribution.
}
\details{
The function \code{tdistr()} calculates the density of the
  non-standard Student's t-distribution by calling the function
  \code{gamma()}.
  The density function of the non-standard Student's t-distribution is given
  by:
  \deqn{
    f(t) = \frac{\Gamma((\nu+1)/2)}{\sqrt{\pi \nu} \, \sigma \, \Gamma(\nu/2)} \, (1 + (\frac{t - \mu}{\sigma})^2/\nu)^{-(\nu+1)/2}
  }
Where \eqn{\Gamma()} is the gamma function, and \eqn{\nu} are the degrees of
freedom.

The non-standard Student's density function has a mean equal to the location
parameter \eqn{\mu}, and a standard deviation proportional to the scale
parameter \eqn{\sigma}.
}
\examples{
\dontrun{
# Student t-distribution at x=1, with df=4, location=-1 and scale=2
tdistr(1, df=4, loc=-1, scalev=2)
# Student t-distribution at x=1, with location=0 and scale=1 - same as dt()
all.equal(tdistr(1, df=3), dt(1, df=3))
}

}
