% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\docType{data}
\name{.tempdirs}
\alias{.tempdirs}
\title{Utils}
\format{\preformatted{ NULL
}}
\usage{
.tempdirs
}
\description{
holds all the tempdirs to be removed when this packages
is unloaded
}
\keyword{datasets}

