% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_effect.R
\name{test_effect}
\alias{test_effect}
\title{Test Delta Effect of Generic Function}
\usage{
test_effect(data, col, measure, delta, obs, FUN, ...)
}
\arguments{
\item{data}{data frame to test on}

\item{col}{name of the input column which is to be changed}

\item{measure}{name of the measure column in the resulting data frame}

\item{delta}{delta applied to observations in col}

\item{obs}{indexes for observations in data to test on}

\item{FUN}{function to apply to data frame in the form of FUN(data, ...)}

\item{...}{additional params for FUN}
}
\value{
Data frame with observation, original value, delta, original measure, new measure, effect
}
\description{
\code{test_effect} returns a data frame recording the results of changing an
input variable on a target variable after applying a function.
}
