% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_it.R
\name{time_it}
\alias{time_it}
\title{Time execution measurement}
\usage{
time_it(operation, type = c("elapsed", "user.self", "sys.self"))
}
\arguments{
\item{operation}{a target function to have its execution time
measured.}

\item{type}{the OS time type to be returned by time_it.
Defaults to 'elapsed'.}
}
\description{
Function to measure how much time (in seconds) some operation
has taken to execute.
}
\examples{

f <- function() { sapply(1:1000000, sum) }
timeElapsed <- time_it(f)

}

