% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timeOperation.R
\name{timeOperation}
\alias{timeOperation}
\title{Time execution measurement}
\usage{
timeOperation(operation, type = c("elapsed", "user.self", "sys.self"))
}
\arguments{
\item{operation}{a target function to have its execution time
measured.}

\item{type}{the OS time type to be returned by timeOperation.
Defaults to "elapsed".}
}
\description{
Function to measure how much time (in seconds) some operation
has taken to execute.
}
\examples{
f <- function() { sapply(1:1000000, sum) }
timeElapsed <- timeOperation(f)
}

