% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{titlecase_one}
\alias{titlecase_one}
\title{Title Case (on one word)}
\usage{
titlecase_one(s)
}
\arguments{
\item{s}{input character vector string}
}
\value{
the same character vector with only first letter in upper case
}
\description{
Puts the first letter of the input character string in upper case, and the rest in lower case.
}
\examples{
titlecase_one("adrian")
}
\keyword{case}
\keyword{name}
\keyword{title}
