% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tls_regression}
\alias{tls_regression}
\title{Total Least Squares Regression for a set of points.}
\usage{
tls_regression(points)
}
\arguments{
\item{points}{A matrix of points. Rows are observations.}
}
\description{
Total Least Squares Regression for a set of points.
}
