% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_binomial_name.R
\name{to_binomial_name}
\alias{to_binomial_name}
\title{Correct species binomial name case}
\usage{
to_binomial_name(x)
}
\arguments{
\item{x}{a \code{character} vector with misspelled binomial names.}
}
\value{
A \code{character} vector of same length as \code{x}.
}
\description{
Corrects species binomial name case.
}
\examples{
x <- c("sAlMO saLAR (Linnee, 1765)", "tyranosurus_REX rex.1")
to_binomial_name(x)
}
