% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_title_case_pt.R
\name{to_title_case_pt}
\alias{to_title_case_pt}
\title{Convert to title case considering Portuguese rules}
\usage{
to_title_case_pt(
  x,
  articles = TRUE,
  conjunctions = TRUE,
  oblique_pronouns = TRUE,
  prepositions = TRUE,
  custom_rules = c(`(.)\\\\bD(el)\\\\b` = "\\\\1d\\\\2")
)
}
\arguments{
\item{x}{A \code{\link[base:character]{character}} vector.}

\item{articles}{(Optional) A \code{\link[base:logical]{logical}} flag indicating if
articles should be converted (default: \code{TRUE}).}

\item{conjunctions}{(Optional) A \code{\link[base:logical]{logical}} flag indicating
if conjunctions should be converted (default: \code{TRUE}).}

\item{oblique_pronouns}{(Optional) A \code{\link[base:logical]{logical}} flag
indicating if oblique pronouns should be converted (default: \code{TRUE}).}

\item{prepositions}{(Optional) A \code{\link[base:logical]{logical}} flag
indicating if prepositions should be converted (default: \code{TRUE}).}

\item{custom_rules}{(Optional) A \code{\link[base:character]{character}} vector
with custom rules to be applied. The syntax is
\code{c("regex" = "replacement")}. The default is
\code{c("(.)\\\\bD(el)\\\\b" = "\\\\1d\\\\2")}, which converts \emph{Del} to \emph{del}.}
}
\value{
A \code{\link[base:character]{character}} vector.
}
\description{
\code{to_title_case_pt()} converts a character vector to title case, but
keeping some classes of words in lower case.

In written Portuguese (PT), when converting to title case, it is not usual
to keep in title case some words, like prepositions, conjunctions,
articles and some kinds of pronouns. This function locates those
cases and converts them to lower case.

(Adapted from the original function \code{to_title_case()} by \strong{Jose de Jesus
Filho})
}
\examples{
to_title_case_pt("Desterro de Entre Rios")
#> [1] "Desterro de entre Rios" # Expected

to_title_case_pt("Pablo Del Rei")
#> [1] "Pablo del Rei" # Expected

to_title_case_pt("Sant'ana do Livramento")
#> [1] "Sant'Ana do Livramento" # Expected

to_title_case_pt("Alta Floresta d'Oeste")
#> [1] "Alta Floresta D'Oeste" # Expected
}
\seealso{
Other string functions.: 
\code{\link{extract_initials}()}
}
\author{
Jose de Jesus Filho
}
\concept{string functions.}
