% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ini_parse.r
\name{toml_parse}
\alias{toml_parse}
\title{Parses a TOML file}
\usage{
toml_parse(filename)
}
\arguments{
\item{filename}{filename to parse}
}
\value{
a list containing the TOML files' informations
}
\description{
Parses a TOML file
}
