% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{tracker}
\alias{tracker}
\title{Named and numbered list of values}
\usage{
tracker()
}
\value{
(function) A function that receives two arguments, `value` and
   `val_name`. Calling the function with arguments populates a named list,
   where the name of the object is `val_name` and the stored value is `val`.
   This function can be called without arguments to return the named list.
}
\description{
Function to keep a named list of recorded values; useful when performing
}

