% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{trim}
\alias{trim}
\title{Character trim function}
\usage{
trim(s)
}
\arguments{
\item{s}{input character vector string}
}
\value{
the same character vector without spaces at the begining or at the end
}
\description{
This functions removes trailing space characters at the begining and at the end of the input character vector.
}
\examples{
trim(" lorem ipsum	")
}
\keyword{trim}
