% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_matrix.R
\name{trim_matrix}
\alias{trim_matrix}
\title{Trim matrix on each side}
\usage{
trim_matrix(m, row, col)
}
\arguments{
\item{row}{numeric. number of rows to trim on each side}

\item{col}{numeric. number of cols to trim on each side}
}
\description{
Removes n number of rows and columns on each side of a matrix
}
\details{
a 5x5 matrix with nx == nx = 1 will become a 3x3 matrix
(e.g. will remove the first and last col and rows)
}
\examples{

m <-
matrix(rep(1:9, each = 9), ncol = 9) +
matrix(rep(seq(10,90, by = 10), each = 9), ncol = 9, byrow = T)

m

trim_matrix(m, 1, 2)
}
