% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{unescape_html}
\alias{unescape_html}
\title{Unescape HTML strings}
\usage{
unescape_html(strings)
}
\arguments{
\item{strings}{character vector}
}
\value{
character vector
}
\description{
Unescape HTML strings
}
