% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{update_pb}
\alias{update_pb}
\title{Update Progress Bar}
\usage{
update_pb(pb, index)
}
\arguments{
\item{pb}{progress bar previously created with `create_pb` (or NULL in a non-TTY context, in which case this is a no-op)}

\item{index}{progress level (bounded between 0 and pb`$tot_iter`, that was set by parameter `nb_iter` in `create_pb`)}
}
\description{
Create an amazingly stylish progress bar.
}
\examples{
NB_ITER = 20
pb <- create_pb(NB_ITER)
\dontrun{for (i in 1:NB_ITER) {
    update_pb(pb,i)
    Sys.sleep(0.5)
}}

NB_ITER = 1000
pb <- create_pb(NB_ITER, bar_style="simple", time_style="cd")
\dontrun{for (i in 1:NB_ITER) {
    update_pb(pb,i)
    Sys.sleep(0.5)
}}

NB_ITER = 10000
# NOT RUN
pb <- create_pb(NB_ITER, bar_style="pc", time_style="end")
\dontrun{for (i in 1:NB_ITER) {
    update_pb(pb,i)
    Sys.sleep(0.5)
}}
}
\seealso{
create_pb
}
\keyword{bar}
\keyword{progress}
