% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pkg_year.R
\name{update_pkg_year}
\alias{update_pkg_year}
\title{Update the year in system files of a package}
\usage{
update_pkg_year(
  file = c(here::here("LICENSE"), here::here("LICENSE.md"), here::here("inst",
    "CITATION"))
)
}
\arguments{
\item{file}{(Optional) A \code{\link[base:character]{character}} vector indicating
the path of the files that must be updated. Default to \code{LICENSE},
\code{LICENSE.md}, and \code{inst/CITATION}.}
}
\value{
An invisible \code{NULL}. This function is used for its side effect.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{update_pkg_year()} updates the year in system files of a package.
}
\seealso{
Other R package functions: 
\code{\link{raw_data_1}()},
\code{\link{raw_data_2}()},
\code{\link{update_pkg_versions}()}
}
\concept{R package functions}
