% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{upf}
\alias{upf}
\title{Unprocessed Files}
\usage{
upf(.dir_out, .dir_in = NULL, .names = NULL)
}
\arguments{
\item{.dir_out}{A directory with Files}

\item{.dir_in}{A directory with Files}

\item{.names}{A vector of file names}
}
\value{
...
}
\description{
Filter only those file names that are not already processed
}
