% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{urns.character}
\alias{urns.character}
\title{Unique Remove-Nas Sort -> URNS}
\usage{
\method{urns}{character}(x)
}
\arguments{
\item{x}{vector to clean}
}
\description{
Cleans a vector (removes NAs & duplicates, sorts values ascending).
}
\details{
default S3 method for `urns` for class `character`
}
\examples{
\dontrun{urns(sample(letters, 100, replace=TRUE))}
}
\seealso{
urns.numeric urns.character sort unique na.rm
}
\keyword{NA}
\keyword{sort}
\keyword{unique}
\keyword{urns}
