% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{use_functions}
\alias{use_functions}
\title{Create Script Function File}
\usage{
use_functions(.dir_here, .dir_script = NULL, .name_script, .browse = FALSE)
}
\arguments{
\item{.dir_here}{Main Directory.}

\item{.dir_script}{Script Directory (optional).}

\item{.name_script}{Script Name.}

\item{.browse}{Logical, if TRUE, opens the file in a web browser. Default is FALSE.}
}
\value{
A path to the created script function file.
}
\description{
This function creates a script function file within the specified directory and sources it.
Optionally, the function can open the file in a web browser.
}
