% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{use_funding}
\alias{use_funding}
\title{Funding}
\usage{
use_funding(path = ".")
}
\arguments{
\item{path}{Path to project/package folder.}
}
\value{
Called by its side effects of creating a file.
}
\description{
Set github file for funding.
}
\examples{
use_funding()
}
