% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodeHome.R
\name{utilFile}
\alias{utilFile}
\title{provide wrapper to path to reference util files}
\usage{
utilFile(fileName)
}
\arguments{
\item{fileName}{of expected file in path}
}
\description{
provide wrapper to path to reference util files
}
