% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{.basename}
\alias{.basename}
\title{Just like basename without the ext}
\usage{
.basename(path)
}
\arguments{
\item{path}{path}
}
\value{
just the name of the path
}
\description{
Just like basename without the ext
}
\author{
Giuseppe Acito
}

