% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utm_zone.R
\name{utm_zone}
\alias{utm_zone}
\alias{utm_zone,numeric,numeric-method}
\alias{utm_zone,numeric,numeric,logical-method}
\alias{utm_zone,Spatial,missing-method}
\alias{utm_zone,sf,missing-method}
\alias{utm_zone,Spatial,missing,logical-method}
\title{UTM Zone}
\usage{
utm_zone(x, y, proj4string = FALSE)

\S4method{utm_zone}{numeric,numeric}(x, y, proj4string = FALSE)

\S4method{utm_zone}{Spatial,missing}(x, y, proj4string = FALSE)

\S4method{utm_zone}{sf,missing}(x, proj4string)
}
\arguments{
\item{x}{longitude or Spatial object of class sf or sp}

\item{y}{latitude or missing}

\item{proj4string}{boolean}
}
\description{
Find UTM Zone from spatial object or x y coordinates
TODO: implement a raster method
}
\examples{
utm_zone(6, 24)

# if a polygon is give, takes its centroid
library(sf)
bayern <- germany.sf[germany.sf$NAME_1 == "Bayern", ]
utm_zone(bayern)

}
