% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_description.r
\name{validate_description}
\alias{validate_description}
\title{Validates a DESCRIPTION because nobody thought how to do it.}
\usage{
validate_description(path = "DESCRIPTION")
}
\arguments{
\item{path}{path to the DESCRIPTION file}
}
\description{
dumbasses.
}
