\name{var.sd}
\alias{var.sd}
\title{Variance with standard error}
\usage{
  var.sd(x, y = c(), att = list())
}
\arguments{
  \item{x}{vector fo which the variance needs to be
  computed}

  \item{y}{is optional, in case you want to compute
  covariance}

  \item{att}{is the list of attributes to append to the
  results}
}
\description{
  compute the variance and returns the standard error on
  the estimate. It the return value is a data.frame
}

