% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{vect.feed.backward}
\alias{vect.feed.backward}
\title{Feeds each missing value with the value of its next element. Order input vector appropriately brfore calling this function.}
\usage{
vect.feed.backward(v, index = sequence(length(v)) \%>\% rev)
}
\description{
Feeds each missing value with the value of its next element. Order input vector appropriately brfore calling this function.
}
