% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{vect.feed.forward}
\alias{vect.feed.forward}
\title{Feeds each missing value with the value of its previous element. Order input vector appropriately brfore calling this function.}
\usage{
vect.feed.forward(v, index = sequence(length(v)))
}
\description{
Feeds each missing value with the value of its previous element. Order input vector appropriately brfore calling this function.
}
