% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/vector.to.string.R
\name{vector.to.string}
\alias{vector.to.string}
\title{Vector Formatting}
\usage{
vector.to.string(vec, collapse = ", ")
}
\arguments{
\item{vec}{a vector containing the elements to be joined.}

\item{collapse}{character value to use as a separator between
the elements of \code{vec}.}
}
\description{
This function joins every element of a vector into a single
character element. Useful for formatting output.
}
\examples{
vector.to.string(1:10)
## Produces "1, 2, 3, 4, 5, 6, 7, 8, 9, 10"
}

