% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetor_to_c.R
\name{vector_to_c}
\alias{vector_to_c}
\title{Put a vector into \code{c()}}
\usage{
vector_to_c(x, quote = TRUE, clipboard = TRUE)
}
\arguments{
\item{x}{An \code{\link[checkmate:checkAtomic]{atomic}} vector.}

\item{quote}{A \code{\link[base:logical]{logical}} flag indicating if the values
should be wrapped in double quotes (Default: \code{TRUE}).}

\item{clipboard}{A \code{\link[base:logical]{logical}} flag indicating if the output
should be copied to the clipboard (Default: \code{TRUE}).}
}
\value{
A \code{\link[base:character]{character}} string with all values from \code{x}
inside the \code{\link[base:c]{c()}} function.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{vector_to_c()} put a vector into the \code{\link[base:c]{c()}} function.

This function is useful when you want to assign values in a vector
while programming in R.
}
\examples{
vector_to_c(letters[1:5], clipboard = FALSE)
#> c("a", "b", "c", "d", "e") # Expected

vector_to_c(1:5, quote = FALSE, clipboard = FALSE)
#> c(1, 2, 3, 4, 5) # Expected
}
\concept{parsing/conversion functions.}
