% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WebDetailsCache}
\alias{WebDetailsCache}
\title{WebDetailsCache class}
\description{
WebDetailsCache class

WebDetailsCache class
}
\details{
A class for caching web details including keys, metadata, and entry details.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cache_name}}{Unique name for the cache.}

\item{\code{cache_dir}}{Directory for cache files.}

\item{\code{log_dir}}{Directory for log files.}

\item{\code{all_entry_keys}}{List of cached entry keys.}

\item{\code{all_entry_details}}{List of cached entry details.}

\item{\code{metadata}}{Metadata associated with the cache.}

\item{\code{list_all_entries_url}}{URL for listing all entries.}

\item{\code{key_field_extraction_func}}{Function to extract the key from entries.}

\item{\code{single_entry_details_url}}{URL template for fetching single entry details.}

\item{\code{all_entry_keys_file}}{File path for storing entry keys.}

\item{\code{metadata_file}}{File path for storing metadata.}

\item{\code{all_entry_details_file}}{File path for storing entry details.}

\item{\code{flush_threshold}}{Number of entry detils fetched before flushing to disk. (improves perf since writing to disk can take a long time)}

\item{\code{list_all_entries_page_size}}{Page size for listing entries.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WebDetailsCache-new}{\code{WebDetailsCache$new()}}
\item \href{#method-WebDetailsCache-load_cache}{\code{WebDetailsCache$load_cache()}}
\item \href{#method-WebDetailsCache-build_cache}{\code{WebDetailsCache$build_cache()}}
\item \href{#method-WebDetailsCache-are_all_entry_keys_cached}{\code{WebDetailsCache$are_all_entry_keys_cached()}}
\item \href{#method-WebDetailsCache-are_all_entry_details_cached}{\code{WebDetailsCache$are_all_entry_details_cached()}}
\item \href{#method-WebDetailsCache-fetch_all_entry_keys}{\code{WebDetailsCache$fetch_all_entry_keys()}}
\item \href{#method-WebDetailsCache-fetch_all_entry_details}{\code{WebDetailsCache$fetch_all_entry_details()}}
\item \href{#method-WebDetailsCache-flush_all_entry_keys}{\code{WebDetailsCache$flush_all_entry_keys()}}
\item \href{#method-WebDetailsCache-flush_all_entry_details}{\code{WebDetailsCache$flush_all_entry_details()}}
\item \href{#method-WebDetailsCache-clone}{\code{WebDetailsCache$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebDetailsCache-new"></a>}}
\if{latex}{\out{\hypertarget{method-WebDetailsCache-new}{}}}
\subsection{Method \code{new()}}{
initialize a new WebDetailsCache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebDetailsCache$new(
  cache_name,
  cache_dir,
  log_dir,
  list_all_entries_url,
  key_field_extraction_func,
  single_entry_details_url,
  flush_threshold = 100,
  list_all_entries_page_size = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cache_name}}{Unique name for the cache.}

\item{\code{cache_dir}}{Directory to store cache files.}

\item{\code{log_dir}}{Directory for log files.}

\item{\code{list_all_entries_url}}{URL for listing all entries.}

\item{\code{key_field_extraction_func}}{Function to extract the key from entries.}

\item{\code{single_entry_details_url}}{URL template for fetching details for a single entry.}

\item{\code{flush_threshold}}{flush_threshold Number of entry detils fetched before flushing to disk (default 100).}

\item{\code{list_all_entries_page_size}}{Number of entries per page (default 1000).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebDetailsCache-load_cache"></a>}}
\if{latex}{\out{\hypertarget{method-WebDetailsCache-load_cache}{}}}
\subsection{Method \code{load_cache()}}{
Load the cache data from disk

Loads cached keys, metadata, and entry details from their respective files if they exist.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebDetailsCache$load_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebDetailsCache-build_cache"></a>}}
\if{latex}{\out{\hypertarget{method-WebDetailsCache-build_cache}{}}}
\subsection{Method \code{build_cache()}}{
Build the cache.

Loads existing cache data and fetches missing entry keys and details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebDetailsCache$build_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebDetailsCache-are_all_entry_keys_cached"></a>}}
\if{latex}{\out{\hypertarget{method-WebDetailsCache-are_all_entry_keys_cached}{}}}
\subsection{Method \code{are_all_entry_keys_cached()}}{
Check if all entry keys are cached.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebDetailsCache$are_all_entry_keys_cached()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if the number of cached keys equals the total number of entries; FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebDetailsCache-are_all_entry_details_cached"></a>}}
\if{latex}{\out{\hypertarget{method-WebDetailsCache-are_all_entry_details_cached}{}}}
\subsection{Method \code{are_all_entry_details_cached()}}{
Check if all entry details are cached.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebDetailsCache$are_all_entry_details_cached()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if every key has corresponding details cached; FALSE otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebDetailsCache-fetch_all_entry_keys"></a>}}
\if{latex}{\out{\hypertarget{method-WebDetailsCache-fetch_all_entry_keys}{}}}
\subsection{Method \code{fetch_all_entry_keys()}}{
Fetch all entry keys from the web.

Retrieves entry keys in a paginated manner and updates the cache.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebDetailsCache$fetch_all_entry_keys()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebDetailsCache-fetch_all_entry_details"></a>}}
\if{latex}{\out{\hypertarget{method-WebDetailsCache-fetch_all_entry_details}{}}}
\subsection{Method \code{fetch_all_entry_details()}}{
Fetch all entry details from the web.

Iterates over all entry keys and fetches detailed information for each.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebDetailsCache$fetch_all_entry_details()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebDetailsCache-flush_all_entry_keys"></a>}}
\if{latex}{\out{\hypertarget{method-WebDetailsCache-flush_all_entry_keys}{}}}
\subsection{Method \code{flush_all_entry_keys()}}{
Flush cached entry keys to disk.

Atomically saves the cached keys and metadata to disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebDetailsCache$flush_all_entry_keys()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebDetailsCache-flush_all_entry_details"></a>}}
\if{latex}{\out{\hypertarget{method-WebDetailsCache-flush_all_entry_details}{}}}
\subsection{Method \code{flush_all_entry_details()}}{
Flush cached entry details to disk.

Atomically saves the cached entry details to disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebDetailsCache$flush_all_entry_details()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WebDetailsCache-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WebDetailsCache-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebDetailsCache$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
