% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{workDir}
\alias{workDir}
\title{Creates a working directory
The user is in charge of deleting it.}
\usage{
workDir(prefix = NULL)
}
\value{
the path of the newly created workdir
}
\description{
Creates a working directory
The user is in charge of deleting it.
}
\author{
Giuseppe Acito
}

