% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{write_files}
\alias{write_files}
\title{Write Files in Various Formats}
\usage{
write_files(.obj, .path)
}
\arguments{
\item{.obj}{An R Object}

\item{.path}{A character string representing the path where the file should be saved.}
}
\description{
This function writes a tibble to the specified path in various formats.
Supported file formats include RDS, FST, XLSX, CSV, DTA, Parquet, and Feather.
}
\examples{
\dontrun{
write_files(data, "data/sample.csv")
write_files(data, "data/sample.rds")
}
}
