% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{ws}
\alias{ws}
\title{Wide screen}
\usage{
ws(ret = TRUE)
}
\arguments{
\item{ret}{return new terminal width}
}
\value{
terminal width
}
\description{
Fit R session output width to terminal window width.
}
\examples{
\dontrun{ws()}
}
\keyword{screen}
\keyword{wide}
\keyword{ws}
