% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo_webhook_push.R
\name{zenodo_webhook_push}
\alias{zenodo_webhook_push}
\title{Push the a GitHub release to Zenodo via webhook}
\usage{
zenodo_webhook_push(
  user,
  repository,
  webhook_token = askpass::askpass("Enter Zenodo webhook token: "),
  release_tag = "latest",
  sandbox = FALSE
)
}
\arguments{
\item{user}{A \code{\link[base:character]{character}} string with the GitHub
username or organization name of the repository.}

\item{repository}{A \code{\link[base:character]{character}} string with the GitHub
repository name.}

\item{webhook_token}{A \code{\link[base:character]{character}} string with the
Zenodo webhook token to authenticate the deposit. This token can be found
on the repository settings under \emph{Webhooks} after \href{https://help.zenodo.org/docs/github/enable-repository/}{linking the GitHub repository to Zenodo}.
If not provided, the function will prompt you to enter it
interactively using the \code{\link[askpass:askpass]{askpass}} function.}

\item{release_tag}{(optional) A \code{\link[base:character]{character}} string with
the GitHub release tag to push to Zenodo. If set to \code{"latest"} (default),
the latest release will be pushed (default: \code{"latest"}).}

\item{sandbox}{(optional) A \code{\link[base:logical]{logical}} flag indicating
whether to the webhook token is for the \href{https://sandbox.zenodo.org}{Zenodo Sandbox} Instance or the main Zenodo platform
(default: \code{FALSE}).}
}
\value{
An \link[base:invisible]{invisible} \code{NULL}. This function is called for
its side effects.
}
\description{
\code{zenodo_webhook_push()} provides an interface to the \href{https://developers.zenodo.org/}{Zenodo REST API} to deposit the latest release of your
GitHub repository using the provided webhook token for authentication. This
is especially useful in cases when you need to manually trigger the Zenodo
webhook.

Please note that the function assumes that the GitHub repository is already
\href{https://help.zenodo.org/docs/github/enable-repository/}{linked to Zenodo}
and that a release exists.

I recommend testing this action in the \href{https://sandbox.zenodo.org}{Zenodo Sandbox} instance before using it with the main
Zenodo platform to avoid unintended data submissions.
}
\details{
Based on Pieter Huybrechts' script, found
\href{https://github.com/zenodo/zenodo/issues/1463#issuecomment-2812815208}{here}.
}
\examples{
\dontrun{
  zenodo_webhook_push(
    user = "danielvartan",
    repository = "actverse",
    webhook_token = "your_zenodo_webhook_token_here",
    release_tag = "latest",
    sandbox = FALSE
  )
}
}
\concept{API functions}
