% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{add_month}
\alias{add_month}
\title{Add or subtract months to a given date character}
\usage{
add_month(input_date, num_months = 1)
}
\arguments{
\item{num_months}{(numeric or integer): Number of months to be added to the given date. It can be a negative number.}
}
\value{
(character): date pointing to the input date with added months
}
\description{

}
\examples{
add_month('2012-06-01', 3)
[1] "2012-09-01"
add_month('2010-03-12', -6)
[1] "2009-09-12"
}
