% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{bucket2}
\alias{bucket2}
\title{Threshold-based bucketing function}
\usage{
bucket2(v, splits = quantile(v, probs = seq(0, 1, by = 1/10)[2:(10 + 1)]))
}
\arguments{
\item{v}{input numeric vector}

\item{splits}{should values of v be rounded}
}
\value{
a numeric vector where each value represents a class/bucket
}
\description{
Associates each value of a numeric input vector to a threshold-based bucket.
}
\examples{
bucket2(rnorm(100, 100, 25), c(50, 75, 100, 125, 150))
}
\keyword{bucket}
\keyword{ditribution}
