% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{chart_dygraph2y}
\alias{chart_dygraph2y}
\title{Plot an interactive \emph{dygraphs} line plot for two \emph{xts} time series,
with two \emph{"y"} axes.}
\usage{
chart_dygraph2y(x_ts, ...)
}
\arguments{
\item{x_ts}{An \emph{xts} time series with two columns.}

\item{...}{additional arguments to function \code{dygraphs::dygraph()}.}
}
\value{
A \code{dygraphs} plot object.
}
\description{
Plot an interactive \emph{dygraphs} line plot for two \emph{xts} time series,
with two \emph{"y"} axes.
}
\details{
The function \code{chart_dygraph2y()} creates an interactive dygraphs
  line plot with two \emph{"y"} axes.
  The function \code{chart_dygraph2y()} uses plotting functions from the
  package
  \href{https://cran.r-project.org/web/packages/dygraphs/index.html}{dygraphs}.
}
\examples{
# plot an interactive dygraphs line plot with two "y" axes
price_s <- cbind(Ad(rutils::etf_env$VTI), Ad(rutils::etf_env$IEF))
colnames(price_s) <- get_name(colnames(price_s), field=2)
rutils::chart_dygraph2y(price_s)
}
