% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{chart_xts}
\alias{chart_xts}
\title{Plot an \code{xts} time series with custom y-axis range and with vertical
background shading.}
\usage{
chart_xts(x_ts, ylim = NULL, in_dex = NULL, ...)
}
\arguments{
\item{x_ts}{\code{xts} time series.}

\item{ylim}{\code{numeric} vector with two elements containing the y-axis
range.}

\item{in_dex}{\code{Boolean} vector or \code{xts} time series for specifying
the shading areas, with TRUE indicating "lightgreen" shading, and FALSE
indicating "lightgrey" shading.}

\item{...}{additional arguments to function \code{chart_Series()}.}
}
\value{
chart object \code{chob} returned invisibly.
}
\description{
A wrapper for function \code{chart_Series()} from package
\href{https://cran.r-project.org/web/packages/quantmod/index.html}{quantmod}.
}
\details{
Extracts the chart object and modifies its ylim parameter using
  accessor and setter functions.
  Also adds background shading using function \code{add_TA()}. The
  \code{in_dex} argument should have the same length as the \code{x_ts} time
  series.
  Finally the function \code{chart_xts()} plots the chart object and returns
  it invisibly.
}
\examples{
\dontrun{
chart_xts(env_etf$VTI["2015-11"],
  name="VTI in Nov 2015", ylim=c(102, 108),
  in_dex=zoo::index(env_etf$VTI["2015-11"]) > as.Date("2015-11-18"))
}
}

