% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{chart_xts2y}
\alias{chart_xts2y}
\title{Plot two \emph{xts} time series with two y-axes in an x11 window.}
\usage{
chart_xts2y(xtsv, color = "red", x11 = TRUE, ...)
}
\arguments{
\item{\code{xtsv}}{An \emph{xts} time series with two columns.}

\item{\code{color}}{A string specifying the color of the second line and
axis (the default is \code{"red"}).}

\item{\code{x11}}{A \emph{Boolean} argument: if \code{TRUE} then open x11
window for plotting, else plot in standard window (the default is \code{TRUE}).}

\item{\code{...}}{Additional arguments to function \code{plot.zoo()}.}
}
\value{
The \emph{xtsv} column names returned invisibly, and a plot
  in an x11 window produced as a side effect.
}
\description{
Plot two \emph{xts} time series with two y-axes in an x11 window.
}
\details{
The function \code{chart_xts2y()} creates a plot of two \emph{xts}
  time series with two y-axes.
  By default \code{chart_xts2y()} opens and plots in an x11 window.
  The function \code{chart_xts2y()} uses the standard plotting functions from
  base \emph{R}, and the function \code{plot.zoo()} from package
  \href{https://cran.r-project.org/web/packages/zoo/index.html}{zoo}.
}
\examples{
# Plot two time series
rutils::chart_xts2y(cbind(quantmod::Cl(rutils::etfenv$VTI),
                          quantmod::Cl(rutils::etfenv$IEF))["2015"])
}
