% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/cluster.r
\docType{class}
\name{Cluster}
\alias{Cluster}
\title{A Cluster proxy to internal cluster functions}
\description{
This proxy has been made to be less error prone and a little easier
to emply clusters in computations
}
\details{
no value added, just a small API
}
\section{Fields}{

\describe{
\item{\code{ncores}}{stores the number of cores used for this cluster}

\item{\code{cluster}}{the internal cluster}

\item{\code{working}}{a logical status to fix some erratic behaviour of the internal cluster}
}}
\section{Methods}{

\describe{
\item{\code{initialize(ncores = detectCores())}}{Initializes the Cluster by the number of cores argument (ncores)}

\item{\code{isWorking()}}{Returns `TRUE` is the cluster is in working state}

\item{\code{submit(X, fun, ...)}}{Submit a `fun` to the cluster to operate on the `X` data}
}}

