% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal.R
\name{create_pb}
\alias{create_pb}
\title{Create Progress Bar}
\usage{
create_pb(
  nb_iter,
  bar_style = c("simple", "pc"),
  time_style = c("cd", "end"),
  width = NULL
)
}
\arguments{
\item{nb_iter}{how many steps will there be}

\item{bar_style}{`simple` to have a simple progress bar, or `pc` if you want percentages}

\item{time_style}{`cd` to have a simple countdown, or `end` if you want an estimate of end date/time}

\item{width}{terminal width to draw into; detected from the terminal (falling back to 100) when NULL}
}
\description{
Create an amazingly stylish progress bar.
}
\examples{
NB_ITER = 20
pb <- create_pb(NB_ITER)
\dontrun{for (i in 1:NB_ITER) {
    update_pb(pb,i)
    Sys.sleep(0.5)
}}

NB_ITER = 1000
pb <- create_pb(NB_ITER, bar_style="simple", time_style="cd")
\dontrun{for (i in 1:NB_ITER) {
    update_pb(pb,i)
    Sys.sleep(0.5)
}}

NB_ITER = 10000
# NOT RUN
pb <- create_pb(NB_ITER, bar_style="pc", time_style="end")
\dontrun{for (i in 1:NB_ITER) {
    update_pb(pb,i)
    Sys.sleep(0.5)
}}
}
\seealso{
update_pb
}
\keyword{bar}
\keyword{progress}
