% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{cry}
\alias{cry}
\title{Naive encryption function}
\usage{
cry(x, key = 123)
}
\arguments{
\item{x}{character to encrypt}

\item{key}{random seed}
}
\value{
a character string, the same length as x, but non-human readable
}
\description{
Naive bijective encryption function
}
\examples{
a <- cry("Love is in the air", 1234)
print(a)
print(decry(a, 1235))
print(decry(a, 1234))
}
\seealso{
decry
}
\keyword{cipher}
\keyword{encryption}
