% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{decry}
\alias{decry}
\title{Naive decryption function}
\usage{
decry(x, key = 123)
}
\arguments{
\item{x}{character to decrypt}

\item{key}{random seed that was used to encrypt}
}
\value{
a character string, the same length as x, that should be human-readable if you did it right.
}
\description{
Naive bijective decryption function
}
\examples{
a <- cry("Love is in the air", 1234)
print(a)
print(decry(a, 1235))
print(decry(a, 1234))
}
\seealso{
cry
}
\keyword{cipher}
\keyword{encryption}
