% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{decry}
\alias{decry}
\title{Naive de-obfuscation function}
\usage{
decry(x, key = 123)
}
\arguments{
\item{x}{character to de-obfuscate}

\item{key}{random seed that was used with \code{cry}}
}
\value{
a character string, the same length as x, that should be human-readable if you used the right key.
}
\description{
Reverses \code{\link{cry}} for a given key. See \code{cry} for the security
caveats: this is obfuscation, not encryption.
}
\examples{
a <- cry("Love is in the air", 1234)
print(a)
print(decry(a, 1235))
print(decry(a, 1234))
}
\seealso{
cry
}
\keyword{cipher}
\keyword{obfuscation}
\keyword{substitution}
