% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{do_call}
\alias{do_call}
\title{Recursively apply a function to a list of objects, such as \emph{xts} time
series.}
\usage{
do_call(func_tion, li_st, ...)
}
\arguments{
\item{func_tion}{The name of function that returns a single object from a
list of objects.}

\item{li_st}{A list of objects, such as \code{vectors}, \code{matrices},
\code{data frames}, or \code{time series}.}

\item{...}{additional arguments to function \code{func_tion()}.}
}
\value{
A single \code{vector}, \code{matrix}, \code{data frame}, or
  \code{time series}.
}
\description{
Performs a similar operation as \code{do.call()}, but using recursion, which
is much faster and uses less memory. The function \code{do_call()} is a
generalization of function \code{do_call_rbind()}.
}
\details{
The function \code{do_call()} performs an lapply loop, each time
  binding neighboring elements and dividing the length of \code{li_st} by
  half. The result of performing \code{do_call(rbind, list_xts)} on a list of
  \emph{xts} time series is identical to performing \code{do.call(rbind,
  list_xts)}. But \code{do.call(rbind, list_xts)} is very slow, and often
  causes an \sQuote{out of memory} error.
}
\examples{
# create xts time series
x_ts <- xts(x=rnorm(1000), order.by=(Sys.time()-3600*(1:1000)))
# split time series into daily list
list_xts <- split(x_ts, "days")
# rbind the list back into a time series and compare with the original
identical(x_ts, rutils::do_call(rbind, list_xts))
}

