% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{do_call_assign}
\alias{do_call_assign}
\title{Apply a function to a list of objects, merge the outputs into a single
object, and assign the object to the output environment.}
\usage{
do_call_assign(
  func,
  symbolv = NULL,
  outv,
  inenv = .GlobalEnv,
  outenv = .GlobalEnv,
  ...
)
}
\arguments{
\item{\code{func}}{The name of a function that returns a single object
(\emph{vector}, \emph{xts} time series, etc.)}

\item{\code{symbolv}}{A vector of \emph{character} strings with the names of
input objects.}

\item{\code{outv}}{The string with name of output object.}

\item{\code{inenv}}{The environment containing the input \code{symbolv}.}

\item{\code{outenv}}{The environment for creating the \code{outv}.}

\item{\code{...}}{Additional arguments to function \code{func()}.}
}
\value{
A single object (\emph{matrix}, \emph{xts} time series, etc.)
}
\description{
Apply a function to a list of objects, merge the outputs into a single
object, and assign the object to the output environment.
}
\details{
The function \code{do_call_assign()} performs an lapply loop over
  \code{symbolv}, applies the function \code{func()}, merges the
  outputs into a single object, and creates the object in the environment
  \code{outenv}.  The output object is created as a side effect, while its
  name is returned invisibly.
}
\examples{
newenv <- new.env()
rutils::do_call_assign(
   func=get_col,
   symbolv=rutils::etfenv$symbolv,
   outv="prices",
   inenv=etfenv, outenv=newenv)
}
