% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_shell.R
\name{docker_run_rserver}
\alias{docker_run_rserver}
\title{Run an RStudio instance on the bioconductor docker image}
\usage{
docker_run_rserver(
  image = "bioconductor/bioconductor_docker:devel",
  port = 8888,
  password = "bioc",
  name = "dz_bioc",
  rm = FALSE,
  volumes = NULL
)
}
\arguments{
\item{image}{\code{character(1)} name of the image to use}

\item{port}{\code{integer(1)} port of the host which will be used to access
RStudio server via \code{localhost::port}}

\item{password}{\code{character(1)} password to used for RStudio server}

\item{name}{\code{character(1)} name for the container.}

\item{rm}{\code{logical(1)} whether to use the \code{-rm} flag when running the
container. Should the container be removed when stopped?}

\item{volumes}{\code{character(1)} paths for the hosts files that you want
accessible in the container. These will be mounted using the identical
paths on the host and with \code{read-only} permissions.}
}
\description{
A wrapper for the docker function used to run a RStudio container on a
specified docker image. This will be available at the \code{port} specified on the
host.
}
\examples{

\dontrun{
docker_run_rserver(volumes = "/home/dzhang/projects/dasper_analysis/raw_data;/home/dzhang/projects/dasper_analysis/scripts")
}
}
\references{
\code{docker_run_rserver} is based on
\verb{https://github.com/LieberInstitute/megadepth/blob/master/R/megadepth.R}.
}
