% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_shell.R
\name{docker_run_rserver}
\alias{docker_run_rserver}
\title{Run an RStudio instance on the bioconductor docker image}
\usage{
docker_run_rserver(
  image = "bioconductor/bioconductor_docker:devel",
  port = 8888,
  password = "bioc",
  name = "dz_bioc",
  rm = FALSE,
  volumes = NULL,
  volumes_ro = NULL,
  permissions = "match",
  USERID = 1002,
  GROUPID = 1024,
  verbose = TRUE
)
}
\arguments{
\item{image}{\code{character(1)} name of the image to use}

\item{port}{\code{integer(1)} port of the host which will be used to access
RStudio server via \code{localhost::port}.}

\item{password}{\code{character(1)} password to used for RStudio server.}

\item{name}{\code{character(1)} name for the container.}

\item{rm}{\code{logical(1)} whether to use the \code{-rm} flag when running the
container. Should the container be removed when stopped?}

\item{volumes}{\code{character(1)} paths for the hosts files that you want
accessible in the container. See argument \code{permissions} for read/write
access to these. Can either be a path on host or in the form
"host_path:container_path".}

\item{volumes_ro}{\code{character(1)} paths for the hosts files that you want
accessible in the container. These will be read-only on the container and
this option is recommended for any volumes you don't need to modify.}

\item{permissions}{\code{character(1)} if set to "match" (be careful!), then the
permissions of the mounted volumes on the container, will match that of the
host. I.e. the user executing the docker command will have permissions to
read/write/execute in the container as they did on the host.}

\item{USERID}{\code{integer(1)} the USERID of the user for which you would like
the permissions of the mounted volumes to match.}

\item{GROUPID}{\code{integer(1)} the GROUPID of the user for which you would like
the permissions of the mounted volumes to match.}

\item{verbose}{\code{logical(1)} whether to display the command to be run (for
debugging purposes).}
}
\description{
A wrapper for the docker function used to run a RStudio container on a
specified docker image. This will be available at the \code{port} specified on the
host. Currently, the \code{permissions} argument operates through setting a USERID
and GROUPID, which are by default set to my user's settings. In order to make
sure your volumes have the correct permissions, these settings should be
modified to match the user of interest.
}
\examples{

\dontrun{
docker_run_rserver()
}
}
\references{
\code{docker_run_rserver} is based on
\verb{https://github.com/LieberInstitute/megadepth/blob/master/R/megadepth.R}.
The solution for the \code{permissions} of the \code{volumes} was taken from
https://github.com/rocker-org/rocker/wiki/Sharing-files-with-host-machine.
}
