% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Download files from the internet to a local directory}
\usage{
download_file(url, dir = ".", broken_links = FALSE)
}
\arguments{
\item{url}{A \code{\link[base:character]{character}} vector of URLs pointing to
files.}

\item{dir}{(optional) A string specifying the directory where the files
should be downloaded (default: ".").}

\item{broken_links}{(optional) A \code{\link[base:logical]{logical}} flag indicating
whether to return a list of broken links (default: \code{FALSE}).}
}
\value{
If \code{broken_links} is \code{TRUE}, an invisible
\code{\link[base:character]{character}} vector of broken links. Otherwise, a
invisible \code{\link[base:character]{character}} vector of file paths where
the files were downloaded.
}
\description{
\code{download_file()} downloads files from the internet to a local directory. It
can handle multiple files at once and provides progress updates during the
download process.

The function also checks for broken links and can return a list of any
broken links encountered during the download.
}
\examples{
library(curl)

if (has_internet()) {
  urls <- paste0(
    "ftp://ftp.datasus.gov.br/dissemin/publicos/IBGE/POPSVS/",
     c("POPSBR00.zip", "POPSBR01.zip")
  )

  dir <- tempfile("dir")
  dir.create(dir)

  download_file(urls, dir)
}
}
\seealso{
Other file functions: 
\code{\link{get_file_size}()}
}
\concept{file functions}
