% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na.R
\name{drop_na}
\alias{drop_na}
\title{Remove \code{NA} values from an atomic object}
\usage{
drop_na(x)
}
\arguments{
\item{x}{An \code{\link[checkmate:checkAtomic]{atomic}} object of any type.}
}
\value{
An object of the same type as \code{x} with all \code{NA} values removed.
}
\description{
\code{drop_na()} remove all \code{NA} values from an atomic object.
}
\examples{
drop_na(c(1, NA, 2, 3, NA, 4))
#> [1] 1 2 3 4 # Expected

drop_na(c("a", NA, "b", "c"))
#> [1] "a" "b" "c" # Expected
}
\seealso{
Other vector functions.: 
\code{\link{count_na}()},
\code{\link{inbetween_integers}()}
}
\concept{vector functions.}
