% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{fetch_safe}
\alias{fetch_safe}
\title{Fetch safe}
\usage{
fetch_safe(url, max_attempts = 3, handle = NULL, logger = NULL)
}
\arguments{
\item{url}{url to fetch}

\item{max_attempts}{numeric that represents the maximum number of attempts}

\item{handle}{`curl::handle` to add to the connection}

\item{logger}{your custom logger if you want to keep track of potential warnings or errors}
}
\value{
a list that represents the result of the fetch (with headers and content still binarized),
		or NULL if `url` couldn't be fetched in the specified number of attempts.
}
\description{
Simple wrapper around `curl::curl_fetch_memory` with a few safeguards
}
\examples{
fetch_safe("http://www.qwant.com")
fetch_safe("http://www.qwant.comme")
}
\seealso{
curl::curl_fetch_memory
}
\keyword{curl}
\keyword{curl_fetch_memory}
\keyword{fetch}
