% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_web_content}
\alias{fetch_web_content}
\title{Fetch web content as text}
\usage{
fetch_web_content(url, query = list())
}
\arguments{
\item{url}{The URL to fetch content from.}

\item{query}{A list of query parameters.}
}
\value{
The response content as text, or NULL if the fetch failed.
}
\description{
Fetch web content as text
}
