% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flypwd.r
\name{flypwd}
\alias{flypwd}
\title{Returns the password stored by flypwd}
\usage{
flypwd(service, clean)
}
\arguments{
\item{service}{service name}

\item{clean}{flush credentials stored in session}
}
\value{
the current user's password
}
\description{
Returns the password stored by flypwd
}
\author{
Giuseppe Acito
}
