% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get_date_from_file}
\alias{get_date_from_file}
\title{Get date from file name}
\usage{
get_date_from_file(path)
}
\arguments{
\item{path}{Charactoer vector of file paths}
}
\value{
Date vector
}
\description{
Get date from file name
}
