% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_ext.R
\name{get_file_ext}
\alias{get_file_ext}
\title{Get file extensions}
\usage{
get_file_ext(file)
}
\arguments{
\item{file}{A \code{\link[base:character]{character}} vector with file paths.}
}
\value{
A \code{\link[base:character]{character}} vector containing the file
extension of the files.
}
\description{
\code{get_file_ext()} is similar to \code{\link[tools:fileutils]{file_ext()}}, the
difference is that it returns the file extension with the dot
separator (e.g., \code{.csv}).
}
\examples{
get_file_ext("example.txt")
#> [1] ".txt" # Expected

get_file_ext("/path/to/file.tar.gz")
#> [1] ".gz"  # Expected

get_file_ext("no-extension")
#> [1] NA # Expected

get_file_ext(c("data.csv", "data.rds"))
#> [1] ".csv" ".rds" # Expected
}
\seealso{
Other string functions: 
\code{\link{get_file_name_without_ext}()}
}
\concept{string functions}
