% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_size.R
\name{get_file_size}
\alias{get_file_size}
\title{Get the sizes of local files or files from URLs}
\usage{
get_file_size(file)
}
\arguments{
\item{file}{A \code{\link[base:character]{character}} vector of file paths.
The function also works with URLs.}
}
\value{
A \code{\link[fs:fs_bytes]{fs_bytes}} vector of file sizes.
}
\description{
\code{get_file_size()} returns the sizes of files in bytes. It works with local
files and URLs.
}
\examples{
library(fs)
library(readr)

files <- c("file1.txt", "file2.txt", "file3.txt")

dir <- tempfile("dir")
dir.create(dir)

for (i in files) {
  write_lines(rep(letters, sample(1000:10000, 1)), file.path(dir, i))
}

urls <- c(
  paste0(
    "https://geodata.ucdavis.edu/climate/worldclim/2_1/base/",
    "wc2.1_2.5m_tavg.zip"
  ),
  paste0(
    "https://geodata.ucdavis.edu/climate/worldclim/2_1/base/",
    "wc2.1_10m_prec.zip"
  )
)

c(urls, path(dir, files)) |> get_file_size()
}
\seealso{
Other file functions: 
\code{\link{download_file}()}
}
\concept{file functions}
