% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gutils-package.R
\docType{package}
\name{gutils-package}
\alias{gutils}
\alias{gutils-package}
\title{gutils: Utility Functions Created by the GIPSO Dev Team}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An R package to store and organize utility functions created by the GIPSO developer team.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gipso.github.io/gutils/}
  \item \url{https://github.com/gipso/gutils/}
  \item Report bugs at \url{https://github.com/gipso/gutils/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Mario Pedrazzoli \email{pedrazzo@usp.br} (\href{https://orcid.org/0000-0002-5257-591X}{ORCID}) [scientific advisor]
}

Other contributors:
\itemize{
  \item Interdisciplinary Sleep Research Group (GIPSO) [funder]
  \item University of Sao Paulo (USP) [funder]
}

}
\keyword{internal}
