% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is.windows}
\alias{is.windows}
\title{Misc Utils}
\usage{
is.windows()
}
\value{
\code{TRUE} if the current R sessions runs on a Windows system
}
\description{
returns \code{TRUE} if the current R sessions runs on a Windows system
}
\author{
Giuseppe Acito
}

