% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{lag_it}
\alias{lag_it}
\title{Apply a lag to a \code{numeric} vector or matrix.}
\usage{
lag_it(in_put, lag = 1)
}
\arguments{
\item{in_put}{a \code{numeric} vector or matrix.}

\item{lag}{integer equal to the number of periods of lag.}
}
\value{
vector or matrix with the same dimensions as the input object.
}
\description{
Apply a lag to a \code{numeric} vector or matrix.
}
\details{
Applies a lag to a vector or matrix, by shifting its values by a
  certain number of rows, equal to the integer \code{lag}, and pads the
  leading or trailing stub periods with \code{zeros}. Positive \code{lag}
  means that values in the current row are replaced with values from the row
  that are \code{lag} rows above. (vice versa negative \code{lag}).
}
\examples{
# lag vector by 2 periods
lag_it(1:10, lag=2)
# lag matrix by negative 2 periods
lag_it(matrix(1:10, ncol=2), lag=-2)
}

