% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioinfo.R
\name{llrs_cellranger_multi}
\alias{llrs_cellranger_multi}
\title{Create multi file}
\usage{
llrs_cellranger_multi(config, gex, library, samples, vdj, feature, out)
}
\arguments{
\item{out}{Path to a directory where to save the file.}

\item{path}{Files to be used./ Info to be used.}
}
\value{
The path to a file saved with the configuration as specified.
}
\description{
Prepare the cellranger multi file for some samples.
}
\examples{
llrs_cellranger_multi(out = "/S2_cellranger.csv")
}
\references{
\url{https://www.10xgenomics.com/support/software/cell-ranger/latest/advanced/cr-multi-config-csv-opts}
}
