% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnag.R
\name{llrs_cnag_cellranger}
\alias{llrs_cnag_cellranger}
\title{Prepare CNAG data for cellranger}
\usage{
llrs_cnag_cellranger(path)
}
\arguments{
\item{path}{Path to a file with information about the delivery file.}
}
\value{
A data.frame with the output of the names of the files ("d1" and "d2")
and the new names c("cr1", "cr2")they should have.
}
\description{
CNAG returns the data in a format that is not compatible with cellranger.
}
\examples{
# out <- llrs_cnag_cellranger("PROJECT_01.xls")
# out[, c("SAMPLE NAME", "d1", "cr1", "d1", "cr2")]
}
\references{
\url{https://www.10xgenomics.com/support/software/cell-ranger/latest/analysis/inputs/cr-specifying-fastqs#file-naming-convention}
}
