% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnag.R
\name{llrs_cnag_deliver}
\alias{llrs_cnag_deliver}
\title{Read the delivery file}
\usage{
llrs_cnag_deliver(path)
}
\arguments{
\item{path}{Path to the project file.}
}
\description{
Import the delivery file which has information about the machine, application
and samples provided by the researcher.
}
\examples{
# llrs_cnag_deliver("AUSER_01.xls")
}
\references{
\href{https://www.cnag.eu/}{CNAG}
}
