% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/governance.R
\name{llrs_governance}
\alias{llrs_governance}
\title{Add the governance}
\usage{
llrs_governance(path = ".")
}
\arguments{
\item{path}{Path to the directory where the file will be added.}
}
\value{
A logical value if the file has been added.
}
\description{
To make clear expectations and how people can join projects state how their
contributors can lead to higher rights (and responsibilities).
}
\examples{
\dontrun{
llrs_governance()
}
}
\references{
\url{https://antonin.delpeuch.eu/posts/off-the-shelf-governance-models-for-small-foss-projects/}
}
