% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_profile.R
\name{llrs_profile}
\alias{llrs_profile}
\title{Provide a R global user profile}
\usage{
llrs_profile()
}
\value{
\code{TRUE} or \code{FALSE} depending on if it has overwritten the files or not.
}
\description{
Write my preferences on a R user profile and R environment.
}
