% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onedrive.R
\name{llrs_upload}
\alias{llrs_upload}
\title{Upload a file to Microsoft cloud in One Drive}
\usage{
llrs_upload(path)
}
\arguments{
\item{path}{File you want to upload}
}
\value{
The id of the file uploaded
}
\description{
Upload a file to Microsoft cloud in One Drive
}
