% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{mondf}
\alias{mondf}
\title{Month difference (vectorized)}
\usage{
mondf(id1, id2, absolute = TRUE)
}
\arguments{
\item{id1}{input date vector (formatted character, IDate, POSIXlt, POSIXct, Date, etc.)}

\item{id2}{input date vector (formatted character, IDate, POSIXlt, POSIXct, Date, etc.)}

\item{absolute}{should the time difference be absolute, or oriented.
When `absolute=FALSE`, `id1` is expected to be before `id2`. Hence the returned value will be negative if it's not the case.}
}
\value{
numeric value representing the number of months between two dates. If `absolute` is set to `TRUE`,
this difference will always be positive.
}
\description{
Number of calendar-month boundaries crossed between two dates (the day of the
month is ignored, so e.g. Jan 31 -> Feb 1 counts as 1). Inputs are coerced
with \code{data.table::as.IDate}; only date formats it understands parse
correctly. The two arguments must have the same length, or one must be
length 1 (it is then recycled).
}
\examples{
mondf("2019-01-01", as.Date(Sys.time()))
mondf(as.Date(Sys.time()), "2019-01-01", absolute=FALSE)
mondf(c("2019-01-01", "2018-01-01"), as.Date(Sys.time()))
mondf(c("2019-01-01", "2018-01-01"), rep(as.Date(Sys.time()), 2))
}
\keyword{date}
\keyword{difference}
\keyword{month}
