% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicals.R
\name{na.false}
\alias{na.false}
\title{Logical evaluation that treats NA as FALSE}
\usage{
na.false(x)
}
\arguments{
\item{x}{logical expression to evaluate}
}
\value{
TRUE if x is TRUE, FALSE if it is FALSE or NA
}
\description{
Binarily evaluates a logical expression with NA giving FALSE
}
\examples{
na.false(FALSE) == na.false(NA)
}
\seealso{
na.true
}
\keyword{NA}
