% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicals.R
\name{na.true}
\alias{na.true}
\title{Logical evaluation that treats NA as TRUE}
\usage{
na.true(x)
}
\arguments{
\item{x}{logical expression to evaluate}
}
\value{
FALSE if x is FALSE, TRUE if it is TRUE or NA
}
\description{
Binarily evaluates a logical expression with NA giving TRUE
}
\examples{
na.true(TRUE) == na.true(NA)
}
\seealso{
na.false
}
\keyword{NA}
