% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_as.R
\name{na_as}
\alias{na_as}
\title{Return a \code{NA} value with the same characteristics of an R object}
\usage{
na_as(x)
}
\arguments{
\item{x}{An \link[checkmate:checkAtomicVector]{atomic vector}, provided that
\code{na_as()} have a method for it.}
}
\value{
A \code{NA} value with the same class and attributes of \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{na_as()} returns a \code{NA} value of the same characteristics of an R object.
It really helps when you need to assign \code{NA}s dynamically.
}
\examples{
na_as(TRUE)
#> [1] NA # Expected
class(na_as(TRUE))
#> [1] "logical" # Expected

na_as(as.Date("2020-01-01"))
#> [1] NA # Expected
class(na_as(as.Date("2020-01-01")))
#> [1] "Date"  # Expected
}
\concept{parsing/conversion functions}
