% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{na_locf}
\alias{na_locf}
\title{Replace \code{NA} values with the most recent non-\code{NA} values prior to
them.}
\usage{
na_locf(in_put, from_last = FALSE, na_rm = FALSE, max_gap = NROW(in_put))
}
\arguments{
\item{\code{in_put}}{A \emph{numeric} or \emph{Boolean} vector or matrix, or
\emph{xts} time series.}

\item{\code{from_last}}{A  \emph{Boolean} argument: should non-\code{NA}
values be carried backward rather than forward? (default is \code{FALSE})}

\item{\code{na_rm}}{A \emph{Boolean} argument: should any remaining (leading
or trailing) \code{NA} values be removed? (default is \code{FALSE})}

\item{\code{max_gap}}{The maximum number of neighboring \code{NA} values that
can be replaced (default is \code{NROW(in_put)}).}
}
\value{
A vector, matrix, or \emph{xts} time series with the same dimensions
  and data type as the argument \code{in_put}.
}
\description{
Replace \code{NA} values with the most recent non-\code{NA} values prior to
them.
}
\details{
The function \code{na_locf()} replaces \code{NA} values with the
  most recent non-\code{NA} values prior to them.

  If the \code{from_last} argument is \code{FALSE} (the default), then the
  previous or past non-\code{NA} values are carried forward to replace the
  \code{NA} values.
  If the \code{from_last} argument is \code{TRUE}, then the following or
  future non-\code{NA} values are carried backward to replace the \code{NA}
  values.

  The function \code{na_locf()} performs the same operation as function
  \code{xts:::na.locf.xts()} from package
  \href{https://cran.r-project.org/web/packages/zoo/index.html}{zoo}, but
  it also accepts vectors as input.

  The function \code{na_locf()} calls the compiled function \code{na_locf()}
  from package
  \href{https://cran.r-project.org/web/packages/xts/index.html}{xts}, which
  allows it to perform its calculations about three times faster than
  \code{xts:::na.locf.xts()}.
}
\examples{
# Create vector containing NA values
in_put <- sample(22)
in_put[sample(NROW(in_put), 4)] <- NA
# Replace NA values with the most recent non-NA values
rutils::na_locf(in_put)
# Create matrix containing NA values
in_put <- sample(44)
in_put[sample(NROW(in_put), 8)] <- NA
in_put <- matrix(in_put, nc=2)
# Replace NA values with the most recent non-NA values
rutils::na_locf(in_put)
# Create xts series containing NA values
in_put <- xts::xts(in_put, order.by=seq.Date(from=Sys.Date(),
  by=1, length.out=NROW(in_put)))
# Replace NA values with the most recent non-NA values
rutils::na_locf(in_put)
}
