% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_quarter.r
\name{prev_quarter}
\alias{prev_quarter}
\title{this function returns the previous quarter compared on the current date}
\usage{
prev_quarter()
}
\description{
this function returns the previous quarter compared on the current date
}
