% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data.R
\name{raw_data_2}
\alias{raw_data_2}
\title{Get paths to \code{insert_package_name} raw data}
\usage{
raw_data_2(type = NULL, file = NULL, package = "insert_package_name")
}
\arguments{
\item{type}{(optional) a string indicating the file type of the raw data
(default: \code{NULL}).}

\item{file}{(optional) a \code{character} object indicating the file name(s) of
the raw data (default: \code{NULL}).}

\item{package}{(optional) a string indicating the package with the database
data. If \code{NULL}, the function will try to use the name of the active
project directory (requires the \code{rstudioapi} package) (default:
\code{insert_package_name}).}
}
\value{
\itemize{
\item If \code{type = NULL}, a \code{character} object with all file type names
available.
\item If \code{type != NULL} and \code{file = NULL}, a \code{character} object with all file
names available from type.
\item If \code{type != NULL} and \code{file != NULL}, a \code{character} with the file name(s)
path.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{raw_data_2()} returns the raw data paths of the \code{insert_package_name}
package.
}
\examples{
\dontrun{
raw_data_2()
}
}
\seealso{
Other R system functions: 
\code{\link{normalize_extdata}()},
\code{\link{raw_data_1}()},
\code{\link{update_pkg_versions}()}
}
\concept{R system functions}
