% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_behavioural_results}
\alias{read_behavioural_results}
\title{Import all behavioural data}
\usage{
read_behavioural_results(results_json_files)
}
\arguments{
\item{results_json_files}{A list or vector of file names of the json results
files. Use, for example, \code{fs::dir_ls(behavioural_results_dir, glob = '*.json')} to get this list.}
}
\value{
A data frame that concatenates the data frames from all subjects
}
\description{
Read in each behavioural results json file that is in the results directory.
Concatenate the resulting data frames. Create a new "subject" variable with
values s1, s2, ... sn where s1 is the first participant by date, s2 is the
second, and so on.
}
\examples{
\dontrun{
data_df <- read_behavioural_results(json_results_file_list)
}
}
