% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{Read Files}
\usage{
read_files(.path)
}
\arguments{
\item{.path}{Full path to the file}
}
\value{
An R object
}
\description{
Read any of the following files: .rds, .fst, .xlsx, .xls, .csv, .dta
}
\examples{

\dontrun{
library(rUtils)

dir   <- system.file("extdata/files", package = "rUtils")
files <- lfc(dir)

read_files(files[1])
read_files(files[2])

rm(dir, files)
}
}
