% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cachedFile.R
\name{readCachedFile}
\alias{readCachedFile}
\title{readCachedFile}
\usage{
readCachedFile(
  path,
  uniqueName = NULL,
  period = lubridate::hours(1),
  extDrive = FALSE
)
}
\arguments{
\item{path}{to file to cache}

\item{uniqueName}{unique name for cached file in cases where path has common names}

\item{period}{after which the cached file will no longer be valid}

\item{extDrive}{for large files to be stored on external drives}
}
\description{
download and cache frequently used files
}
\details{
for caching files that are used often
}
