% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{readLines}
\alias{readLines}
\title{base::readLines error reporting sucks.}
\usage{
readLines(con = stdin(), n = -1L, ok = TRUE, warn = TRUE,
  encoding = "unknown")
}
\arguments{
\item{con}{@see base::readLines}

\item{n}{@see base::readLines}

\item{ok}{@see base::readLines}

\item{warn}{@see base::readLines}

\item{encoding}{@see base::readLines}
}
\description{
base::readLines error reporting sucks.
}
\seealso{
base::readLines
}
