% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{reload_package}
\alias{reload_package}
\title{Reload Package}
\usage{
reload_package(package_name)
}
\arguments{
\item{package_name}{name of the package you want to reload}
}
\value{
logical that represents whether the package was properly reloaded
}
\description{
Reloads package that was previously loaded.
This is useful for instance when some objects of one package were masked and you want to reload them in a different order, but without loosing your current session's variables.
This function requires to have devtools installed.
}
\examples{
reload_package("rutils")
}
\keyword{package}
\keyword{reload}
