% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Convert factor to numeric}
\usage{
remove_outliers(data, x, outlier_percent = NULL)
}
\arguments{
\item{data}{data.frame}

\item{x}{column name to filter}

\item{outlier_percent}{Percent to remove}
}
\value{
data.frame
}
\description{
Convert factor to numeric
}
