% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{roll_sum}
\alias{roll_sum}
\title{Calculate the rolling sum of a \emph{numeric} vector, matrix, or \emph{xts}
time series over a sliding window (lookback period).}
\usage{
roll_sum(x_ts, look_back)
}
\arguments{
\item{\code{x_ts}}{A vector, matrix, or \emph{xts} time series containing one
or more columns of data.}

\item{\code{look_back}}{The size of the lookback window, equal to the number
of data points for calculating the rolling sum.}
}
\value{
A vector, matrix, or \emph{xts} time series with the same dimensions
  as the input series.
}
\description{
Calculate the rolling sum of a \emph{numeric} vector, matrix, or \emph{xts}
time series over a sliding window (lookback period).
}
\details{
For example, if look_back=3, then the rolling sum at any point is
  equal to the sum of \code{x_ts} values for that point plus two preceding
  points. The initial values of roll_sum() are equal to cumsum() values, so
  that roll_sum() doesn't return any \code{NA} values.

  The function \code{roll_sum()} performs the same operation as function
  \code{runSum()} from package
  \href{https://cran.r-project.org/web/packages/TTR/index.html}{TTR}, but
  using vectorized functions, so it's a little faster.
}
\examples{
# Rolling sum of vector
vec_tor <- rnorm(1000)
rutils::roll_sum(vec_tor, look_back=3)
# Rolling sum of matrix
mat_rix <- matrix(rnorm(1000), nc=5)
rutils::roll_sum(mat_rix, look_back=3)
# Rolling sum of xts time series
x_ts <- xts(x=rnorm(1000), order.by=(Sys.time()-3600*(1:1000)))
rutils::roll_sum(x_ts, look_back=3)
}
