% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.r
\name{rolling_file}
\alias{rolling_file}
\title{Rolls the file based on parameters}
\usage{
rolling_file(filename, max_size, max_files, lock_file)
}
\arguments{
\item{filename}{path of the logging file}

\item{max_size}{size in bytes before rolling (defaults 10Mb)}

\item{max_files}{number of files to keep (defaults 5)}

\item{lock_file}{lock file to prevent access to lock during rolling}
}
\description{
Rolling of a file is based on the following conditions
\itemize{
\item file is bigger than a given size
\item number of rolling files are bigger than
}
}
