% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils-package.R
\docType{package}
\name{rutils-package}
\alias{rutils}
\alias{rutils-package}
\title{rutils: Utility Functions for R}
\description{
An R package made to store and organize utility functions created by the Daniel Vartanian.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielvartan.github.io/rutils/}
  \item \url{https://github.com/danielvartan/rutils/}
  \item Report bugs at \url{https://github.com/danielvartan/rutils/issues/}
}

}
\author{
\strong{Maintainer}: Daniel Vartanian \email{danvartan@gmail.com} (\href{https://orcid.org/0000-0001-7782-759X}{ORCID}) [conceptor, copyright holder]

}
\keyword{internal}
