% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils-package.R
\docType{package}
\name{rutils-package}
\alias{rutils}
\alias{rutils-package}
\title{rutils: A Collection of R Functions}
\description{
A collection of R functions commonly used in FRB-CESAB projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/rutils}
  \item Report bugs at \url{https://github.com/frbcesab/rutils/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
