% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils-package.R
\docType{package}
\name{rutils-package}
\alias{rutils}
\alias{rutils-package}
\title{rutils: Personal collection of functions to work with R}
\description{
rutils provides with functions to use in every day situations.
It is thought as a leightweight package with many suggested dependencies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/llrs/rutils}
  \item \url{https://llrs.github.io/rutils/}
  \item Report bugs at \url{https://github.com/llrs/rutils/issues}
}

}
\author{
\strong{Maintainer}: Lluís Revilla Sancho \email{lluis.revilla@gmail.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID})

}
\keyword{internal}
