% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils-package.R
\docType{package}
\name{rutils-package}
\alias{rutils}
\alias{rutils-package}
\title{rutils: A bunch of (supposedly) useful R functions}
\description{
A collection of small, everyday helpers for string, numeric, date and
logical manipulation, plotting shortcuts, a terminal progress bar and a few
miscellaneous utilities.
}
\author{
\strong{Maintainer}: Adrian Delmarre \email{adelmarre@vongo.eu}

Authors:
\itemize{
  \item Adrian Delmarre \email{adelmarre@vongo.eu}
}

}
\keyword{internal}
