% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\docType{package}
\name{rutils}
\alias{rutils}
\alias{rutils-package}
\title{rutils: A package providing utility functions for R}
\description{
The rutils package provides utility functions
}
\section{AUC/PR Functions}{

Functions to calculate the auc for roc curves and pr curves and to plot both curves either plain
or with color as cutoff.
}

