% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_plot_png}
\alias{save_plot_png}
\title{Save a ggplot2 object as a .png}
\usage{
save_plot_png(plot, title, subfolder = NULL, pwidth = 790, pheight = 553)
}
\arguments{
\item{plot}{the plot to save}

\item{title}{name to give file}

\item{subfolder}{a subdirectory of existing 'Plots' directory where current plot
should be saved}

\item{pwidth}{width of the plot in pixels}

\item{pheight}{heigh of the plot in pixels}
}
\description{
Save ggplot2 object as a .png. If a folder called Plots exists
in your current working directory then the plot will be saved to this location.
}
