\name{stationary.dist}
\alias{stationary.dist}
\title{Obtain stationary distribution of transition matrix by iteration}
\usage{
  stationary.dist(G, g, n)
}
\arguments{
  \item{G}{transition matrix}

  \item{g}{initial distribution}

  \item{n}{maximal iterations}
}
\value{
  the stationary distribution of G
}
\description{
  Obtain stationary distribution of transition matrix by
  iteration
}

