% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_summary.R
\name{stats_summary}
\alias{stats_summary}
\title{Compute summary statistics}
\usage{
stats_summary(
  data,
  col,
  name = NULL,
  na_rm = TRUE,
  remove_outliers = FALSE,
  iqr_mult = 1.5,
  hms_format = TRUE,
  threshold = hms::parse_hms("12:00:00"),
  as_list = FALSE
)
}
\arguments{
\item{data}{A \code{\link[base:data.frame]{data.frame}}.}

\item{col}{A \code{\link[base:character]{character}} \code{string} specifying the column
name in \code{data} for which to compute summary statistics.}

\item{name}{(optional) A \code{\link[base:character]{character}} \code{string} with a
name to include in the summary statistics (default: \code{NULL}).}

\item{na_rm}{A \code{\link[base:logical]{logical}} flag indicating whether to remove
\code{NA} values when computing statistics (default: \code{TRUE}).}

\item{remove_outliers}{A \code{\link[base:logical]{logical}} flag indicating whether
to remove outliers using the IQR method (default: \code{FALSE}).}

\item{iqr_mult}{A \code{\link[base:numeric]{numeric}} value specifying the
Interquartile Range (IQR) multiplier for outlier detection
(default: \code{1.5}).}

\item{hms_format}{A \code{\link[base:logical]{logical}} flag indicating whether to
format temporal statistics as \code{\link[hms:hms]{hms}} objects (default: \code{TRUE}).}

\item{threshold}{A \code{\link[hms:hms]{hms}} object specifying the threshold time
for linking times to a timeline when computing temporal statistics.
See \code{\link[lubritime:link_to_timeline]{link_to_timeline()}} to learn more
(default: \code{hms::parse_hms("12:00:00")}).}

\item{as_list}{A \code{\link[base:logical]{logical}} flag indicating whether to
return the summary statistics as a \code{\link[base:list]{list}} (default: \code{FALSE}).}
}
\value{
If \code{as_list} is \code{TRUE}, a \code{\link[base:list]{list}}  with the summary
statistics. Otherwise, a one-row \code{\link[tibble:tibble]{tibble}} with the
summary statistics.
}
\description{
\code{stats_summary()} computes summary statistics for a specified column in a
data frame.
}
\examples{
library(dplyr)
library(hms)

rnorm(1000) |>
  as_tibble() |>
  stats_summary("value") |>
  print(n = Inf)

letters |>
  as_tibble() |>
  stats_summary("value") |>
  print(n = Inf)

sample(0:86399, 1000) |>
  as_hms() |>
  as_tibble() |>
  stats_summary("value") |>
  print(n = Inf)
}
\seealso{
Other statistical functions.: 
\code{\link{cohens_d}()},
\code{\link{mode}()}
}
\concept{statistical functions.}
