% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_if.R
\name{swap_if}
\alias{swap_if}
\title{Swap two values if a condition is met}
\usage{
swap_if(x, y, condition = TRUE)
}
\arguments{
\item{x}{Any type of R object.}

\item{y}{Any type of R object.}

\item{condition}{A \code{\link[base:logical]{logical}} flag indicating if the
values should be swapped (Default: \code{TRUE}).}
}
\value{
A \code{\link[base:list]{list}} with the swapped values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{swap_if()} swaps two values if a condition is met.
}
\examples{
swap_if(1, 2)
#> $x # Expected
#> [1] 2
#>
#> $y
#> [1] 1

swap_if(1, 2, condition = FALSE)
#> $x # Expected
#> [1] 1
#>
#> $y
#> [1] 2
}
\seealso{
Other conditional functions: 
\code{\link{swap_vector}()}
}
\concept{conditional functions}
