\name{tempdir}
\alias{tempdir}
\title{this is how a tempdir function sould work, not as the base::workdir does}
\usage{
tempdir(prefix = NULL)
}
\arguments{
  \item{prefix}{something to be appended betweeh
  \code{workDir} and the random dir generated}
}
\value{
the path of the working directory
}
\description{
this is how a tempdir function sould work, not as the
base::workdir does
}
\author{
Giuseppe Acito
}

