% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\docType{data}
\name{.tempdirs}
\alias{.tempdirs}
\title{Utils}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
.tempdirs
}
\description{
holds all the tempdirs to be removed when this packages
is unloaded
}
\keyword{datasets}
