% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{to_period}
\alias{to_period}
\title{Aggregate an \emph{xts} time series to a lower periodicity.}
\usage{
to_period(timeser, period = "minutes", k = 1)
}
\arguments{
\item{\code{timeser}}{An \emph{xts} time series of prices.}

\item{\code{period}}{Aggregation interval ("seconds", "minutes", "hours",
"days", "weeks", "months", "quarters", and "years").}

\item{\code{k}}{The number of periods to aggregate over (for example if
\code{period="minutes"} and \code{k=2}, then aggregate over two minute
intervals.)}
}
\value{
A \emph{OHLC} time series of prices in \emph{xts} format, with a
  lower periodicity.
}
\description{
Given an \emph{xts} time series at high periodicity (say seconds),
calculate the \emph{OHLC} prices at a lower periodicity (say minutes).
}
\details{
The function \code{to_period()} is a wrapper for the function
  \code{xts::to.period()} from package
  \href{https://cran.r-project.org/web/packages/xts/index.html}{xts}.
}
\examples{
\dontrun{
# Aggregate the OHLC prices from minutes to days:
ohlc <- rutils::to_period(timeser=HighFreq::SPY["2009"], period="days")
}

}
