% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pkg_versions.R
\name{update_pkg_versions}
\alias{update_pkg_versions}
\title{Update package versions in the `DESCRIPTION`` file}
\usage{
update_pkg_versions(
  file = here::here("DESCRIPTION"),
  old_r_version = bump_back_r_version()
)
}
\arguments{
\item{file}{(Optional) A string indicating the path to the \code{DESCRIPTION}
file.}

\item{old_r_version}{(Optional) A string indicating the previous minor
version of the current R version.}
}
\value{
An invisible \code{NULL}. This function is used for its side effect.
}
\description{
\code{update_pkg_versions()} updates the version of packages listed in the
\code{DESCRIPTION} file of an R package with their installed versions.

If the package comes with an typical installation of R (e.g., base, utils),
the function will update the version to the previous minor version of the
current R version. This is made to avoid errors with CI/CD.
}
\seealso{
Other R package functions: 
\code{\link{raw_data_1}()},
\code{\link{raw_data_2}()}
}
\concept{R package functions}
