% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_pkg_versions.R
\name{update_pkg_versions}
\alias{update_pkg_versions}
\title{Update package versions in DESCRIPTION file}
\usage{
update_pkg_versions(
  file = here::here("DESCRIPTION"),
  old_r_version = bump_back_r_version()
)
}
\arguments{
\item{file}{(optional) A string indicating the path to the \code{DESCRIPTION} file
(default: \code{here::here("DESCRIPTION")}).}

\item{old_r_version}{(optional) A string indicating the previous minor
version of the current R version (default: \code{bump_back_r_version()}).}
}
\value{
An invisible \code{NULL}. This function is used for its side effect.
}
\description{
\code{update_pkg_versions()} updates the version of the packages in the
\code{DESCRIPTION} file of R packages with their current version.

If the package is an R base package, the function will update the version to
the previous minor version of the current R version. This is made to avoid
errors with CI/CD.
}
\seealso{
Other R system functions: 
\code{\link{normalize_extdata}()},
\code{\link{raw_data_1}()},
\code{\link{raw_data_2}()}
}
\concept{R system functions}
