% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{use_funding}
\alias{use_funding}
\title{Funding}
\usage{
use_funding(path = ".", ...)
}
\arguments{
\item{path}{Path to project/package folder.}

\item{...}{Other named strings that are \href{https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/displaying-a-sponsor-button-in-your-repository}{accepted by GitHub}}
}
\value{
Called by its side effects of creating a file.
}
\description{
Set github file for funding.
}
\examples{
use_funding(path = tempdir())
}
