% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/utils.r
\name{whoami}
\alias{whoami}
\title{if "username" is set as environment variable, it's preferred.}
\usage{
whoami()
}
\value{
the OS's username
}
\description{
if "username" is set as environment variable, it's preferred.
}
\author{
Giuseppe Acito
}

