% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-package-functions.R
\name{with_package_bindings}
\alias{with_package_bindings}
\title{Temporarily change bindings in this package}
\usage{
with_package_bindings(expr, ...)
}
\arguments{
\item{expr}{An expression that makes use of the function in \code{functions}}

\item{...}{Pairs of names and values of functions to replace}
}
\value{
The result of the evaluated expression
}
\description{
Thin wrapper to \code{rlang::local_bindings}
}
\examples{
with_package_bindings(do_things(1), thing_1 = loud(thing_1))
}
